### 2025-05-28 #8304 @editorialbot
[PRE REVIEW]: Gyselalib++: A Portable C++ Library for Semi-Lagrangian Kinetic and Gyrokinetic Simulations
**Submitting author:** @EmilyBourne (<a href="http://orcid.org/0000-0002-3469-2338">Emily BOURNE</a>)
**Repository:** https://github.com/gyselax/gyselalibxx
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0
