### 2025-05-28 #8304 @danielskatz
👋 @EmilyBourne - thanks for your submission. Everything looks fine, but I just want to make sure this is in the right JOSS track before we start the process. It may take a day or two to get this going.
