+++ docs/joss_paper/paper.bib
@@ -161,7 +161,7 @@ volume = {234},
 pages = {1-13},
 year = {2019},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2018.08.008},
+doi = {10.1016/j.cpc.2018.08.008},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465518303035},
 author = {P. Donnel and X. Garbet and Y. Sarazin and V. Grandgirard and Y. Asahi and N. Bouzat and E. Caschera and G. Dif-Pradalier and C. Ehrlacher and P. Ghendrih and C. Gillot and G. Latu and C. Passeron},
 keywords = {Collision, Multispecies, H-theorem, Neoclassic},
