+++ docs/standards/coding_covariant_and_contravariant_tensors.md
@@ -30,6 +30,7 @@ In this context, $`\boldsymbol{\eta}_1`$ and $`\boldsymbol{\eta}_2`$ are contrav
 These concepts have a direct equivalent in the code.
 
 Every basis element that may appear in a tensor expression must define the following three pieces of compile-time information:
+
 - Whether the element can be used as a **contravariant** basis element
 - Whether the element can be used as a **covariant** basis element
 - What the associated basis element is in the **dual space** (the contravariant element associated with the covariant element and vice-versa).
@@ -53,16 +54,16 @@ struct Eta1_cov {
 
 Here:
 
-* `Eta1` represents the **contravariant** basis element $`\boldsymbol{\eta}_1`$,
-* `Eta1_cov` represents the **covariant dual** basis element $\boldsymbol{\eta}^1$,
-* Each type declares its **dual**, linking the **contravariant** and **covariant** basis elements.
+- `Eta1` represents the **contravariant** basis element $`\boldsymbol{\eta}_1`$,
+- `Eta1_cov` represents the **covariant dual** basis element $\boldsymbol{\eta}^1$,
+- Each type declares its **dual**, linking the **contravariant** and **covariant** basis elements.
 
 ### Naming Convention
 
 By convention:
 
-* Contravariant dimensions are named `Eta1`, `Eta2`, `X`, `R`, etc.
-* Covariant dimensions use the `_cov` suffix, such as `Eta1_cov`, `Eta2_cov`, `X_cov`, `R_cov`, etc.
+- Contravariant dimensions are named `Eta1`, `Eta2`, `X`, `R`, etc.
+- Covariant dimensions use the `_cov` suffix, such as `Eta1_cov`, `Eta2_cov`, `X_cov`, `R_cov`, etc.
 
 ### Cartesian Coordinates
 
@@ -90,8 +91,8 @@ This is done using the `VectorIndexSet` construct.
 
 A `VectorIndexSet` represents an **ordered list of basis directions**. It is used as a template parameter to vector and tensor types and describes:
 
-* Which basis elements are involved (e.g., `Eta1`, `Eta2`),
-* In what order.
+- Which basis elements are involved (e.g., `Eta1`, `Eta2`),
+- In what order.
 
 ### Example: Defining Bases
 
@@ -102,8 +103,8 @@ using EtaBasis_cov = VectorIndexSet<Eta1_cov, Eta2_cov>; // Covariant basis
 
 This tells the library:
 
-* `EtaBasis` defines a 2D contravariant space with directions `Eta1` and `Eta2`,
-* `EtaBasis_cov` defines the associated covariant dual space.
+- `EtaBasis` defines a 2D contravariant space with directions `Eta1` and `Eta2`,
+- `EtaBasis_cov` defines the associated covariant dual space.
 
 You will use these sets as the type parameters for `Tensor` later.
 
@@ -129,6 +130,7 @@ $$
 
 This object is defined over the tensor product of the contravariant basis $`\{\boldsymbol{\eta}_i\}`$ and the covariant basis $boldsymbol{b}^j$.
 Supposing the tensor components are of type `double`, this can be declared in Gyselalib++ as:
+
 ```cpp
 Tensor<double, EtaBasis, BBasis_cov> M;
 ```
@@ -142,6 +144,7 @@ DTensor<EtaBasis, BBasis_cov> M;
 
 A vector is simply a rank-1 tensor. The same logic applies, and several equivalent declarations are supported for convenience.
 For a vector expressed on the contravariant basis $`\{\boldsymbol{\eta}_1, \boldsymbol{\eta}_2\}`$, any of the following are valid:
+
 ```cpp
 Tensor<double, EtaBasis> v;
 DTensor<EtaBasis> v;
[…]
