+++ docs/mkdocs.yml
@@ -21,7 +21,7 @@ nav:
   - Standards:
       - Coding Standard: docs/standards/CODING_STANDARD.md
       - Maths and Physics: docs/standards/mathematical_and_physical_conventions.md
-      - Coding Covariant and Contravariant Tensors: coding_covariant_and_contravariant_tensors.md
+      - Coding Covariant and Contravariant Tensors: docs/standards/coding_covariant_and_contravariant_tensors.md
   - Gyselalib++:
       - Source: src
       - Simulations: simulations
