### 2025-05-30 caa4262
Update coding conventions to match enforced rules
