### 2025-05-30 467bdbc
Add a section about standard GPU usage
