### 2025-05-30 c818a73
Temporarily comment slicing functionalities
