### 2025-05-30 a3659ef
Use a VectorField for the electric field in the QN solver
