### 2025-05-30 a245edd
Use a vector field to store the electric field as entry to Vlasov
