### 2025-05-30 b00714f
Clang formatting
