+++ src/geometryXYVxVy/poisson/nullqnsolver.hpp
@@ -19,8 +19,7 @@ public:
     /** @brief A QN Solver which does nothing
      *
      * @param[out] electrostatic_potential The electrostatic potential, the result of the poisson solver.
-     * @param[out] electric_field_x The x-component of the electric field, the gradient of the electrostatic potential.
-     * @param[out] electric_field_y The y-component of the electric field, the gradient of the electrostatic potential.
+     * @param[out] electric_field The electric field, the gradient of the electrostatic potential.
      * @param[in] allfdistribu The distribution function.
      */
     void operator()(
