### 2025-05-30 5f37c7f
Deprecate constructor
