### 2025-05-30 80acacb
Update coding conventions file (#321)
The coding conventions file did not fully reflect the conventions that
we apply in the code and request in reviews. This PR updates the list to
add missing conventions and remove conventions that we are not
enforcing.

The convention of using the Gyselalib++ aliases was not being followed
for `ddc::Coordinate`. Remaining uses of `ddc::Coordinate` are removed
and its use is prohibited via the static analysis CI test.

**Removed conventions**
- we name everything using expressive English names (e.g. `temperature`)
and don't use variable names from the equati
[…]
