+++ assets/.doxy/gyselalibxx/gyselalibxx/polar__bsplines_8hpp_source.md
@@ -171,15 +171,14 @@ public:
                     = Idx<typename DiscreteMapping::BSplineR,
                           typename DiscreteMapping::BSplineTheta>;
             if constexpr (C > -1) {
-                const ddc::Coordinate<DimX, DimY> pole
-                        = curvilinear_to_cartesian(ddc::Coordinate<DimR, DimTheta>(0.0, 0.0));
+                const Coord<DimX, DimY> pole
+                        = curvilinear_to_cartesian(Coord<DimR, DimTheta>(0.0, 0.0));
                 const double x0 = ddc::get<DimX>(pole);
                 const double y0 = ddc::get<DimY>(pole);
                 double tau = 0.0;
                 for (std::size_t i(0); i < ddc::discrete_space<BSplinesTheta>().size(); ++i) {
-                    const ddc::Coordinate<DimX, DimY> point
-                            = curvilinear_to_cartesian.control_point(
-                                    mapping_tensor_product_index_type(1, i));
+                    const Coord<DimX, DimY> point = curvilinear_to_cartesian.control_point(
+                            mapping_tensor_product_index_type(1, i));
 
                     const double c_x = ddc::get<DimX>(point);
                     const double c_y = ddc::get<DimY>(point);
@@ -192,11 +191,9 @@ public:
                     tau = tau > tau3 ? tau : tau3;
                 }
                 // Determine the corners for the barycentric coordinates
-                const ddc::Coordinate<DimX, DimY> corner1(x0 + tau, y0);
-                const ddc::Coordinate<DimX, DimY>
-                        corner2(x0 - 0.5 * tau, y0 + 0.5 * tau * sqrt(3.0));
-                const ddc::Coordinate<DimX, DimY>
-                        corner3(x0 - 0.5 * tau, y0 - 0.5 * tau * sqrt(3.0));
+                const Coord<DimX, DimY> corner1(x0 + tau, y0);
+                const Coord<DimX, DimY> corner2(x0 - 0.5 * tau, y0 + 0.5 * tau * sqrt(3.0));
+                const Coord<DimX, DimY> corner3(x0 - 0.5 * tau, y0 - 0.5 * tau * sqrt(3.0));
 
                 const CartesianToBarycentric<DimX, DimY, Corner1Tag, Corner2Tag, Corner3Tag>
                         barycentric_coordinate_converter(corner1, corner2, corner3);
@@ -241,9 +238,8 @@ public:
                 for (IdxR const ir : IdxRange<BSplinesR>(IdxR(0), IdxStepR(C + 1))) {
                     for (IdxTheta const itheta :
                          poloidal_spline_idx_range.take_first(n_theta_in_singular)) {
-                        const ddc::Coordinate<DimX, DimY> point
-                                = curvilinear_to_cartesian.control_point(
-                                        mapping_tensor_product_index_type(ir, itheta));
+                        const Coord<DimX, DimY> point = curvilinear_to_cartesian.control_point(
+                                mapping_tensor_product_index_type(ir, itheta));
                         host_t<DFieldMem<IdxRange<BernsteinBasis>>> bernstein_vals(
                                 bernstein_idx_range);
                         ddc::discrete_space<BernsteinBasis>()
@@ -298,25 +294,19 @@ public:
 
         Impl& operator=(Impl&& x) = default;
 
[…]
