+++ assets/.doxy/gyselalibxx/gyselalibxx/classPolarSplineEvaluator.md
@@ -72,15 +72,15 @@ _Define an evaluator on polar B-splines._ [More...](#detailed-description)
 |   | [**PolarSplineEvaluator**](#function-polarsplineevaluator-24) (OuterExtrapolationRule const & outer\_bc) <br>_Instantiate a_ [_**PolarSplineEvaluator**_](classPolarSplineEvaluator.md) _with boundary evaluation conditions._ |
 |   | [**PolarSplineEvaluator**](#function-polarsplineevaluator-34) ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md) const & x) = default<br>_Instantiate a_ [_**PolarSplineEvaluator**_](classPolarSplineEvaluator.md) _from another._ |
 |   | [**PolarSplineEvaluator**](#function-polarsplineevaluator-44) ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md) && x) = default<br>_Instantiate a_ [_**PolarSplineEvaluator**_](classPolarSplineEvaluator.md) _from another temporary._ |
-|  double | [**deriv\_1\_and\_2**](#function-deriv_1_and_2) (ddc::Coordinate&lt; [**DimR**](classPolarSplineEvaluator.md#typedef-dimr), [**DimTheta**](classPolarSplineEvaluator.md#typedef-dimtheta) &gt; coord\_eval, host\_t&lt; [**ConstPolarSpline**](structConstPolarSpline.md)&lt; PolarBSplinesType &gt; &gt; const spline\_coef) const<br>_Get the value of the cross derivative of the spline function._  |
-|  double | [**deriv\_dim\_1**](#function-deriv_dim_1-12) (ddc::Coordinate&lt; [**DimR**](classPolarSplineEvaluator.md#typedef-dimr), [**DimTheta**](classPolarSplineEvaluator.md#typedef-dimtheta) &gt; coord\_eval, host\_t&lt; [**ConstPolarSpline**](structConstPolarSpline.md)&lt; PolarBSplinesType &gt; &gt; const spline\_coef) const<br>_Get the value of the derivative of the spline function on the first dimension._  |
-|  void | [**deriv\_dim\_1**](#function-deriv_dim_1-22) (DField&lt; Domain, Kokkos::HostSpace &gt; const spline\_eval, ConstField&lt; ddc::Coordinate&lt; [**DimR**](classPolarSplineEvaluator.md#typedef-dimr), [**DimTheta**](classPolarSplineEvaluator.md#typedef-dimtheta) &gt;, Domain, Kokkos::HostSpace &gt; const coords\_eval, host\_t&lt; [**ConstPolarSpline**](structConstPolarSpline.md)&lt; PolarBSplinesType &gt; &gt; const spline\_coef) const<br>_Get the values of the derivative of the spline function on the first dimension._  |
-|  void | [**deriv\_dim\_1\_and\_2**](#function-deriv_dim_1_and_2) (DField&lt; Domain, Kokkos::HostSpace &gt; const spline\_eval, ConstField&lt; ddc::Coordinate&lt; [**DimR**](classPolarSplineEvaluator.md#typedef-dimr), [**DimTheta**](classPolarSplineEvaluator.md#typedef-dimtheta) &gt;, Domain, Kokkos::HostSpace &gt; const coords\_eval, host\_t&lt; [**ConstPolarSpline**](structConstPolarSpline.md)&lt; PolarBSplinesType &gt; &gt; const spline\_coef) const<br>_Get the values of the cross derivative of the spline function._  |
[…]
