### 2025-05-30 9cac5e1
Fix cmake in submodule (#324)
Fix paths in root `CMakeLists.txt` file to ensure it can be correctly
used in a submodule.
Remove reference to non-existant folder `./cmake`. Correct path to
version.txt.
