+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix paths in root `CMakeLists.txt` file to ensure it can be correctly used in a submodule.
 - Update remaining use of `ddc::Coordinate` to use Gyselalib++ conventions (`Coord`).
 - Update coding conventions to match what is applied.
 
