+++ .github/workflows/static_analyses.yml
@@ -6,6 +6,8 @@ on:
       - opened
       - reopened
       - synchronize
+    branches-ignore:
+      - main
   merge_group:
     types: [checks_requested]
 
