+++ bin/ci_tools/gyselalib_static_analysis.py
@@ -33,6 +33,7 @@ ddc_keyword_map = {'DiscreteElement': 'Idx',
                    'Chunk'          : 'FieldMem',
                    'ChunkSpan'      : 'Field',
                    'ChunkView'      : 'ConstField',
+                   'Coordinate'     : 'Coord',
                    'span_view'      : 'get_field',
                    'span_cview'     : 'get_const_field',
                    'domain'         : 'get_idx_range',
