### 2025-05-30 3c8b900
Update changelog for new version (#326)
Update the changelog for a new version following the release of v0.1.1
