+++ .github/workflows/cpu_tests.yml
@@ -7,6 +7,8 @@ on:
       - reopened
       - synchronize
       - ready_for_review
+    branches-ignore:
+      - main
   merge_group:
     types:
       - checks_requested
