### 2025-06-02 a36b606
Use a vector field for the electric field in the geometryXYVxVy (#323)
Use a vector field for the electric field in the geometryXYVxVy. This
removes the need for the constructor of `VectorField` which takes
multiple `Field`s. This constructor is blocking for #322.
