+++ src/geometryXYVxVy/time_integration/predcorr.cpp
@@ -31,20 +31,13 @@ DFieldSpVxVyXY PredCorr::operator()(
 
     // electrostatic potential and electric field (depending only on x)
     DFieldMemXY electrostatic_potential(get_idx_range<GridX, GridY>(allfdistribu_v2D_split));
-    DFieldMemXY electric_field_x(get_idx_range<GridX, GridY>(allfdistribu_v2D_split));
-    DFieldMemXY electric_field_y(get_idx_range<GridX, GridY>(allfdistribu_v2D_split));
+    DVectorFieldMemXY electric_field(get_idx_range<GridX, GridY>(allfdistribu_v2D_split));
 
     host_t<DFieldMemXY> electrostatic_potential_host(
             get_idx_range<GridX, GridY>(allfdistribu_v2D_split));
 
     // a 2D memory block of the same size as fdistribu
     DFieldMemSpVxVyXY allfdistribu_half_t(get_idx_range(allfdistribu_v2D_split));
-
-    m_poisson_solver(
-            get_field(electrostatic_potential),
-            get_field(electric_field_x),
-            get_field(electric_field_y),
-            get_const_field(allfdistribu_v2D_split));
 
     int iter = 0;
     for (; iter < steps; ++iter) {
@@ -54,8 +47,7 @@ DFieldSpVxVyXY PredCorr::operator()(
         // the associated electric field
         m_poisson_solver(
                 get_field(electrostatic_potential),
-                get_field(electric_field_x),
-                get_field(electric_field_y),
+                get_field(electric_field),
                 get_const_field(allfdistribu_v2D_split));
 
         transpose_layout(
@@ -77,33 +69,23 @@ DFieldSpVxVyXY PredCorr::operator()(
         ddc::parallel_deepcopy(allfdistribu_half_t, allfdistribu_v2D_split);
 
         // predictor
-        m_vlasov_solver(
-                get_field(allfdistribu_half_t),
-                get_const_field(electric_field_x),
-                get_const_field(electric_field_y),
-                dt / 2);
+        m_vlasov_solver(get_field(allfdistribu_half_t), get_const_field(electric_field), dt / 2);
 
         // computation of the electrostatic potential at time tn+1/2
         // and the associated electric field
         m_poisson_solver(
                 get_field(electrostatic_potential),
-                get_field(electric_field_x),
-                get_field(electric_field_y),
+                get_field(electric_field),
                 get_const_field(allfdistribu_half_t));
 
         // correction on a dt
-        m_vlasov_solver(
-                get_field(allfdistribu_v2D_split),
-                get_const_field(electric_field_x),
-                get_const_field(electric_field_y),
-                dt);
+        m_vlasov_solver(get_field(allfdistribu_v2D_split), get_const_field(electric_field), dt);
     }
 
     double const final_time = iter * dt;
     m_poisson_solver(
             get_field(electrostatic_potential),
-            get_field(electric_field_x),
-            get_field(electric_field_y),
+            get_field(electric_field),
             get_const_field(allfdistribu_v2D_split));
 
 
