+++ assets/.doxy/gyselalibxx/gyselalibxx/geometryXYVxVy_2poisson_2nullqnsolver_8cpp_source.md
@@ -12,11 +12,7 @@
 
 #include "nullqnsolver.hpp"
 
-void NullQNSolver::operator()(
-        DFieldXY const,
-        DFieldXY const,
-        DFieldXY const,
-        DConstFieldSpVxVyXY const) const
+void NullQNSolver::operator()(DFieldXY const, DVectorFieldXY const, DConstFieldSpVxVyXY const) const
 {
 }
 ```
