+++ assets/.doxy/gyselalibxx/gyselalibxx/geometryXYVxVy_2poisson_2qnsolver_8hpp_source.md
@@ -33,8 +33,7 @@ public:
 
     void operator()(
             DFieldXY electrostatic_potential,
-            DFieldXY electric_field_x,
-            DFieldXY electric_field_y,
+            DVectorFieldXY electric_field,
             DConstFieldSpVxVyXY allfdistribu) const override;
 };
 ```

[…]
