### 2025-06-04 7b9e111
Remove ::contains and move IdxRangeSlice to ddc_aliases
