### 2025-06-04 816678d
Use ddc::StridedDiscreteDomain (#316)
The PR replaces the implementation of `IdxRangeSlice` by a type alias to
`ddc::StridedDiscreteDomain`.
- `using IdxRangeSlice = ddc::StridedDiscreteDomain`, moved in
`ddc_aliases.hpp`
- functions not present in the API of `ddc::StridedDiscreteDomain` have
been modified:
- `IdxRangeSlice::contains` has been inlined in the only placed where it
was used,
- `IdxRangeSlice::get_index` has been moved to
`DerivFieldCommon::get_index` replacing the unused existing
`DerivFieldCommon::get_index`
- using an alias broke some template deduction in `DerivFieldMem`, they
[…]
