+++ assets/.doxy/gyselalibxx/gyselalibxx/derivative__field__common_8hpp_source.md
@@ -103,14 +103,13 @@ protected:
     to_subidx_range_collection<physical_deriv_grids> m_cross_derivative_idx_range;
 
 protected:
-    template <class DElem>
-    KOKKOS_FUNCTION std::pair<int, index_type> get_index(DElem elem) const
+    template <class QueryDim>
+    KOKKOS_FUNCTION std::size_t get_index(Idx<QueryDim> elem) const noexcept
     {
-        discrete_deriv_index_type default_derivatives = detail::no_derivative_element<deriv_tags>();
-        discrete_deriv_index_type deriv_index(detail::select_default(elem, default_derivatives));
-        physical_index_type physical_index(elem);
-        index_type index(physical_index, deriv_index);
-        return std::pair<int, index_type>(get_array_index(deriv_index), index);
+        assert(IdxRangeSlice<QueryDim>(m_cross_derivative_idx_range).contains(elem));
+        return IdxRangeSlice<QueryDim>(m_cross_derivative_idx_range)
+                .distance_from_front(elem)
+                .value();
     }
 
     template <class... Tag>
@@ -134,8 +133,7 @@ protected:
                 // If information is available about the physical index range
                 if (array_idx & (1 << ddc::type_seq_rank_v<ddc::Deriv<QueryDDim>, deriv_tags>)) {
                     // If the derivative is being requested
-                    return m_cross_derivative_idx_range.get_index(
-                            ddc::select<QueryDDim>(slice_idx));
+                    return get_index(ddc::select<QueryDDim>(slice_idx));
                 }
             }
             if constexpr (ddc::in_tags_v<QueryDDim, physical_grids>) {
@@ -172,10 +170,14 @@ protected:
                 IdxRange<QueryDDim> idx_range_requested(slice_idx_range);
                 if (array_idx & (1 << ddc::type_seq_rank_v<ddc::Deriv<QueryDDim>, deriv_tags>)) {
                     // If the derivative is being requested
-                    assert(m_cross_derivative_idx_range.contains(idx_range_requested));
+                    assert(IdxRangeSlice<QueryDDim>(m_cross_derivative_idx_range)
+                                   .contains(idx_range_requested.front())
+                           && (idx_range_requested.extents() == 1
+                               || IdxRangeSlice<QueryDDim>(m_cross_derivative_idx_range).strides()
+                                          == 1));
                     return std::pair<std::size_t, std::size_t>(
-                            m_cross_derivative_idx_range.get_index(idx_range_requested.front()),
-                            m_cross_derivative_idx_range.get_index(idx_range_requested.back()) + 1);
+                            get_index(idx_range_requested.front()),
+                            get_index(idx_range_requested.back()) + 1);
                 }
             }
             if constexpr (ddc::in_tags_v<QueryDDim, physical_grids>) {
