### 2025-06-04 1efc291
Silence some warnings from the GPU CI
