### 2025-06-04 1357423
Fix ignored m_magnetic_field
