### 2025-06-05 420c958
Spack user changes
