+++ toolchains/mi250.hipcc.adastra.spack/prepare.sh
@@ -6,19 +6,29 @@ if [[ "${BASH_SOURCE[0]}" != "${0}" ]]; then
     return 1
 fi
 
-TOOLCHAIN_ROOT_DIRECTORY="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd -P)"
+TOOLCHAIN_ROOT_DIRECTORY="$(dirname -- "$(readlink -f -- "${BASH_SOURCE[0]:-${0}}")")"
 
 set -eu
 
 cd -- "${TOOLCHAIN_ROOT_DIRECTORY}"
 
+SPACK_USER_VERSION="spack-user-4.0.0"
+
+export SPACK_USER_PREFIX="${SHAREDWORKDIR}/gyselalibxx-spack-install-MI250/Configuration.${SPACK_USER_VERSION}"
+export SPACK_USER_CACHE_PATH="${SPACK_USER_PREFIX}/cache"
+
 module purge
-
-export SPACK_USER_PREFIX="${SHAREDWORKDIR}/gyselalibxx-spack-install-MI250"
-
-# FIXME: This loads unneeded modules. It should not interfere with our build.
 module load develop
-module load spack-user-4.0.0
+module load "${SPACK_USER_VERSION}"
+module list
+
+spack config --scope user update --yes-to-all packages
+spack config --scope user update --yes-to-all config
+spack config --scope user add 'packages:all:permissions:read:world'
+spack config --scope user add 'packages:all:permissions:write:group'
+spack config --scope user add 'concretizer:unify:true'
+# spack config --scope user add 'modules:default:tcl:hash_length:4'
+spack debug report
 
 # Inject PDI recipes into our local repo.
 git clone https://github.com/pdidev/spack pdi.spack || true
@@ -53,18 +63,13 @@ pdiplugin-trace%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 pdiplugin-mpi%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 ginkgo%gcc@13.2.1.mi250+openmp+rocm~shared amdgpu_target=gfx90a arch=linux-rhel8-zen3
 eigen%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
-py-matplotlib arch=linux-rhel8-zen3
-py-xarray arch=linux-rhel8-zen3
-py-h5py arch=linux-rhel8-zen3
+py-matplotlib%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
+py-xarray%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
+py-h5py%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 "
 # openblas@0.3.26%gcc@12.1.generic~bignuma~consistent_fpcsr+dynamic_dispatch+fortran~ilp64+locking+pic+shared build_system=makefile symbol_suffix=none threads=none arch=linux-rhel8-zen3
 
 which spack
-
-spack config --scope user update --yes-to-all packages
-spack config --scope user add 'packages:all:permissions:read:world'
-spack config --scope user add 'packages:all:permissions:write:group'
-spack config --scope user add 'concretizer:unify:true'
 
 # If we start preparing a new environment, ensure we wont get name clashes by
 # uninstalling previous products.
