### 2025-06-05 e0686c3
Keep the definition of the function in the class description
