### 2025-06-05 a5479a4
Remove unused objects (#331)
- Remove unused member variable `Tensor::m_data_alloc` to fix
`maybe-uninitialized` warning.
- Remove `R_cov::PERIODIC` and `Theta_cov::PERIODIC` and
`GradientTestFunction::operator()` to fix other unused symbols.
