+++ assets/.doxy/gyselalibxx/gyselalibxx/class_members.md
@@ -900,7 +900,6 @@
 * **m\_pseudo\_physical\_to\_logical** ([**SplinePolarFootFinder**](classSplinePolarFootFinder.md))
 * **m\_pseudo\_physical\_to\_physical** ([**SplinePolarFootFinder**](classSplinePolarFootFinder.md))
 * **m\_rhs** ([**SplitRightHandSideSolver**](classSplitRightHandSideSolver.md))
-* **m\_data\_alloc** ([**Tensor**](classTensor.md))
 * **m\_data** ([**TensorCommon**](classTensorCommon.md))
 * **m\_mapping\_2d** ([**ToroidalToCylindrical**](classToroidalToCylindrical.md))
 * **m\_coord\_changer** ([**TriangularBernsteinPolynomialBasis::Impl**](classTriangularBernsteinPolynomialBasis_1_1Impl.md))
