+++ assets/.doxy/gyselalibxx/gyselalibxx/tensor_8hpp_source.md
@@ -74,9 +74,8 @@ public:
     using base_type::rank;
 
 private:
-    using base_type::s_n_elements;
-    std::array<ElementType, s_n_elements> m_data_alloc;
     using base_type::m_data;
+    using base_type::s_n_elements;
 
 public:
     KOKKOS_DEFAULTED_FUNCTION Tensor() = default;
