### 2025-06-05 c731e56
Add changelog entry
