+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix uninitialized warning in the `Tensor` class.
+- Fix unused `m_magnetic_field` variable in `MaxwellianEquilibrium` class.
 
 ### Changed
 
