### 2025-06-05 400849d
Rename some variables to avoid shadowing
