### 2025-06-06 1d0a5f4
Generalise compute_coeffs_on_mapping to work for any mapping
