### 2025-06-12 bf76bbd
Remove unused m_iter_start from RestartInitialisation
