### 2025-06-13 689392b
Remove unused m_iter_start from RestartInitialisation (#335)
Remove unused m_iter_start from RestartInitialisation
