### 2025-06-13 3f9a2b4
Using default pdi packages and adding the external repos with the pdi file name
