+++ src/geometryXYVxVy/time_integration/predcorr.cpp
@@ -65,9 +65,9 @@ DFieldSpVxVyXY PredCorr::operator()(
                 .with("time_saved", iter_time)
                 .with("fdistribu", allfdistribu_host)
                 .with("electrostatic_potential", electrostatic_potential_host);
+        Kokkos::Profiling::popRegion();
         // copy fdistribu
         ddc::parallel_deepcopy(allfdistribu_half_t, allfdistribu_v2D_split);
-        Kokkos::Profiling::popRegion();
 
         // predictor
         m_vlasov_solver(get_field(allfdistribu_half_t), get_const_field(electric_field), dt / 2);
