### 2025-06-13 4c55f91
Add Kokkos::fence
