### 2025-06-13 4b32010
Enable SYSTEM in cmake add_subdirectory
On compatible compilers, this silences warnings of files included from
these directories.
