+++ CMakeLists.txt
@@ -73,10 +73,10 @@ if("${GYSELALIBXX_INCLUDE_TESTING_DEPENDENCIES}")
   if("${GYSELALIBXX_GTest_DEPENDENCY_POLICY}" STREQUAL "AUTO")
     find_package(GTest "1.12" QUIET)
     if(NOT "${GTest_FOUND}")
-      add_subdirectory("vendor/googletest/" "googletest") # SYSTEM)
+      add_subdirectory("vendor/googletest/" "googletest" SYSTEM)
     endif()
   elseif("${GYSELALIBXX_GTest_DEPENDENCY_POLICY}" STREQUAL "EMBEDDED")
-    add_subdirectory("vendor/googletest/" "googletest") # SYSTEM)
+    add_subdirectory("vendor/googletest/" "googletest" SYSTEM)
   elseif("${GYSELALIBXX_GTest_DEPENDENCY_POLICY}" STREQUAL "INSTALLED")
     find_package(GTest "1.12" REQUIRED)
   endif()
@@ -89,7 +89,7 @@ if(NOT "${Kokkos_FOUND}")
     option(Kokkos_ENABLE_CUDA_CONSTEXPR "Whether to activate experimental relaxed constexpr functions" ON)
     option(Kokkos_ENABLE_CUDA_RELOCATABLE_DEVICE_CODE "Whether to enable relocatable device code (RDC) for CUDA" ON)
   endif()
-  add_subdirectory("vendor/kokkos/" "kokkos") # SYSTEM)
+  add_subdirectory("vendor/kokkos/" "kokkos" SYSTEM)
 endif()
 
 find_package(KokkosKernels 4.5.1 QUIET)
@@ -98,20 +98,20 @@ if(NOT "${KokkosKernels_FOUND}")
   set(KokkosKernels_ENABLE_ALL_COMPONENTS OFF CACHE BOOL "")
   set(KokkosKernels_ENABLE_COMPONENT_BATCHED ON CACHE BOOL "")
   set(KokkosKernels_ENABLE_COMPONENT_BLAS ON CACHE BOOL "")
-  add_subdirectory("vendor/kokkos-kernels/" "kokkos-kernels") # SYSTEM)
+  add_subdirectory("vendor/kokkos-kernels/" "kokkos-kernels" SYSTEM)
 endif()
 
 
 find_package(DDC 0.5 QUIET COMPONENTS fft pdi splines)
 if(NOT "${DDC_FOUND}")
   ## Use the discrete domain computation library (ddc) from `vendor/`
-  add_subdirectory("vendor/ddc/" "ddc") # SYSTEM)
+  add_subdirectory("vendor/ddc/" "ddc" SYSTEM)
 endif()
 
 set(koliop_BUILD_TESTING OFF CACHE BOOL "")
 set(koliop_ENABLE_Kokkos        "SUBDIRECTORY" CACHE STRING "")
 set(koliop_ENABLE_KokkosKernels "SUBDIRECTORY" CACHE STRING "")
-add_subdirectory("vendor/koliop/" "koliop") # SYSTEM)
+add_subdirectory("vendor/koliop/" "koliop" SYSTEM)
 
 ###############################################################################################
 #                              Build libraries and executables
