+++ toolchains/mi250.hipcc.adastra.spack/environment.sh
@@ -6,16 +6,17 @@ module load cpe/24.07
 # FIXME:
 # craype-accel-amd-gfx90a Error with the cray wrappers clang: error: unsupported option '-fopenmp-targets=' for language mode 'HIP'
 module load craype-x86-trento
-module load PrgEnv-cray-amd amd-mixed/6.1.2
+# NOTE: Force 6.3.3 due to startup failures (https://github.com/gyselax/gyselalibxx/pull/198#issuecomment-2943081411)
+module load PrgEnv-cray-amd amd-mixed/6.3.3
 
-module load rocm/6.1.2
+module load rocm/6.3.3
 module load cray-fftw
 module load cray-hdf5-parallel
 module load cray-python
 
-# FIXME: SPACK_USER_PREFIX path too long
-# export SPACK_USER_PREFIX="${ENVIRONMENT_ROOT_DIRECTORY}/gyselalibxx-spack-install-MI250"
-export SPACK_USER_PREFIX="${SHAREDWORKDIR}/gyselalibxx-spack-install-MI250"
+SPACK_USER_VERSION="spack-user-4.0.0"
+
+export SPACK_USER_PREFIX="${SHAREDWORKDIR}/gyselalibxx-spack-install-MI250/Configuration.${SPACK_USER_VERSION}"
 module use "${SPACK_USER_PREFIX}/modules/tcl/linux-rhel8-zen3"
 
 module load \
@@ -24,9 +25,14 @@ module load \
     gcc/13.2.1.mi250/zen3/paraconf \
     gcc/13.2.1.mi250/zen3/pdi \
     gcc/13.2.1.mi250/zen3/pdiplugin-decl-hdf5 \
+    gcc/13.2.1.mi250/zen3/pdiplugin-mpi \
     gcc/13.2.1.mi250/zen3/pdiplugin-set-value \
     gcc/13.2.1.mi250/zen3/pdiplugin-trace \
     gcc/13.2.1.mi250/zen3/ginkgo \
-    gcc/13.2.1.mi250/zen3/ninja
+    gcc/13.2.1.mi250/zen3/eigen \
+    gcc/13.2.1.mi250/zen3/ninja \
+    gcc/13.2.1.mi250/zen3/py-matplotlib \
+    gcc/13.2.1.mi250/zen3/py-xarray \
+    gcc/13.2.1.mi250/zen3/py-h5py
 
 module list
