### 2025-06-17 6f462ec
Remove eigen installation
