+++ toolchains/genoa.gcc.adastra.spack/environment.sh
@@ -19,7 +19,6 @@ module load \
     gcc/13.2.1.genoa/zen4/pdiplugin-set-value \
     gcc/13.2.1.genoa/zen4/pdiplugin-trace \
     gcc/13.2.1.genoa/zen4/ginkgo \
-    gcc/13.2.1.genoa/zen4/eigen \
     gcc/13.2.1.genoa/zen4/ninja \
     gcc/13.2.1.genoa/zen4/py-matplotlib \
     gcc/13.2.1.genoa/zen4/py-xarray \
