+++ toolchains/mi250.hipcc.adastra.spack/prepare.sh
@@ -61,7 +61,6 @@ pdiplugin-set-value%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 pdiplugin-trace%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 pdiplugin-mpi%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 ginkgo%gcc@13.2.1.mi250+openmp+rocm~shared amdgpu_target=gfx90a arch=linux-rhel8-zen3
-eigen%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 py-matplotlib%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 py-xarray%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
 py-h5py%gcc@13.2.1.mi250 arch=linux-rhel8-zen3
