### 2025-06-17 65adfa1
Remove CMAKE_INSTALL_PREFIX variable
