### 2025-06-17 91fe140
Remove comment about HIP RDC
