### 2025-06-17 2c0a89e
Use new AMD GPU naming scheme
