### 2025-06-17 b33872f
Remove old SLL_BUILD_TESTING variable
