+++ toolchains/genoa.gcc.adastra.spack/toolchain.cmake
@@ -34,6 +34,3 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -isystem $ENV{CRAY_LIBSCI_PREFIX}/includ
 set(BLAS_LIBRARIES "$ENV{CRAY_LIBSCI_PREFIX}/lib/libsci_gnu_mp.so;" CACHE STRING "")
 # set(LAPACK_LIBRARIES "$ENV{CRAY_LIBSCI_PREFIX}/lib/libsci_gnu_mp.so;" CACHE STRING "")
 # set(LAPACKE_LIBRARIES "$ENV{CRAY_LIBSCI_PREFIX}/lib/libsci_gnu_mp.so;" CACHE STRING "")
-
-# FIXME: SLL compiler crash
-set(SLL_BUILD_TESTING OFF CACHE BOOL "")
