### 2025-06-18 cf2e215
Fix module ordering in Adastra toolchains (#338)
Turns out that the spack products' modules define ROCM environment
variable. This ends up overriding the variables define by other
fundamental modules, such as the system's ROCM.

The Spack configuration provided by cines is based on ROCM 6.1.2. It is
with this version that we build all the deps.

But GYSELA is known to trigger a GPU global variable bug in HIP. Using
ROCM 6.3.3 fixes that issue. So we build the deps using ROCm 6.1.2 and
build gyselalib using 6.3.3.

This PR swaps the module ordering to prevent any overriding done by
spack module
[…]
