+++ toolchains/mi250.hipcc.adastra.spack/environment.sh
@@ -1,18 +1,8 @@
+#!/bin/bash
+
 ENVIRONMENT_ROOT_DIRECTORY="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd -P)"
 
 module purge
-
-module load cpe/24.07
-# FIXME:
-# craype-accel-amd-gfx90a Error with the cray wrappers clang: error: unsupported option '-fopenmp-targets=' for language mode 'HIP'
-module load craype-x86-trento
-# NOTE: Force 6.3.3 due to startup failures (https://github.com/gyselax/gyselalibxx/pull/198#issuecomment-2943081411)
-module load PrgEnv-cray-amd amd-mixed/6.3.3
-
-module load rocm/6.3.3
-module load cray-fftw
-module load cray-hdf5-parallel
-module load cray-python
 
 SPACK_USER_VERSION="spack-user-4.0.0"
 
@@ -34,5 +24,18 @@ module load \
     gcc/13.2.1.mi250/zen3/py-matplotlib \
     gcc/13.2.1.mi250/zen3/py-xarray \
     gcc/13.2.1.mi250/zen3/py-h5py
+
+unset HIPCC_COMPILE_FLAGS_APPEND
+
+module load cpe/24.07
+# FIXME:
+# craype-accel-amd-gfx90a Error with the cray wrappers clang: error: unsupported option '-fopenmp-targets=' for language mode 'HIP'
+module load craype-x86-trento
+# NOTE: Force 6.3.3 due to startup failures (https://github.com/gyselax/gyselalibxx/pull/198#issuecomment-2943081411)
+module load PrgEnv-cray-amd amd-mixed/6.3.3
+module load rocm/6.3.3
+module load cray-fftw
+module load cray-hdf5-parallel
+module load cray-python
 
 module list
