### 2025-06-18 d09757a
Fix inconsistency with dependency compiler and gyselialib compiler
