### 2025-06-18 23db104
Remove cray modules and libraries from the gcc + hipcc gysela MI250 build. (#339)
We were loading incorrect modules (cray compiler). I changed the cray
module to gcc and thus also changed the blas libraries we link against.

This is more cosmetic than a functional change, though.

---

Please complete the checklist to ensure that all tasks are completed
before marking your pull request as ready for review.

### All Submissions

- [x] Have you ensured that all lines changed in this PR are justified
by a comment found in the description ?
- [ ] Have you updated the
[CHANGELOG.md](https://github.
[…]
