### 2025-06-18 9f5d016
Disable Kokkos Kernels ETI
