+++ toolchains/genoa.gcc.adastra.spack/prepare.sh
@@ -18,70 +18,37 @@ export SPACK_USER_PREFIX="${SHAREDWORKDIR}/gyselalibxx-spack-install-GENOA/Confi
 export SPACK_USER_CACHE_PATH="${SPACK_USER_PREFIX}/cache"
 
 module purge
-module load develop
 module load "${SPACK_USER_VERSION}"
 module list
 
-spack config --scope user update --yes-to-all packages
-spack config --scope user update --yes-to-all config
-spack config --scope user add 'packages:all:permissions:read:world'
-spack config --scope user add 'packages:all:permissions:write:group'
-spack config --scope user add 'concretizer:unify:true'
-# spack config --scope user add 'modules:default:tcl:hash_length:4'
+which spack
 spack debug report
 
-# Inject PDI recipes into our repos.
 mkdir -p -- "${SPACK_USER_CONFIG_PATH}/external-repositories"
+
+# Inject recent PDI recipes into our repository.
 git clone https://github.com/pdidev/spack "${SPACK_USER_CONFIG_PATH}/external-repositories/pdi.spack" || true
 cd -- "${SPACK_USER_CONFIG_PATH}/external-repositories/pdi.spack" && git fetch && git checkout 5483cfea7d2d39d654c2962114248f597b3ecf46 && cd -
-spack repo add --scope user "${SPACK_USER_CONFIG_PATH}/external-repositories/pdi.spack" || true
 
-# NOTE: A sparse checkout would be great.
 git clone https://github.com/spack/spack spack.spack || true
 cd spack.spack && git fetch && git checkout 8e7489b && cd ..
-# NOTE: We may be overriding some CINES modified recipes.
 cp -rf -- spack.spack/var/spack/repos/builtin/packages/ginkgo "${SPACK_USER_CONFIG_PATH}/local-repo/packages"
 
-# FIXME: Add a recent Ginkgo to its package.py
 sed -i '/args.append("-DGINKGO_HIP_AMDGPU={0}".format(arch_str))/a\                args.append("-DCMAKE_HIP_ARCHITECTURES={0}".format(arch_str))' "${SPACK_USER_CONFIG_PATH}/local-repo/packages/ginkgo/package.py"
+
+spack env remove --yes-to-all gyselalibxx-spack-environment
+spack env create gyselalibxx-spack-environment "${TOOLCHAIN_ROOT_DIRECTORY}/gyselalibxx-spack-environment.yaml"
+
+spack --env gyselalibxx-spack-environment repo add "${SPACK_USER_CONFIG_PATH}/external-repositories/pdi.spack"
 
 echo "Preparing the Spack environment..."
 
-# # Either do not use the source mirrors or use --no-check-signature. There is an
-# # ongoing issue a CINES to fix the issue.
-# echo "# Disabled" >"${SPACK_USER_CONFIG_PATH}/mirrors.yaml"
+spack --env gyselalibxx-spack-environment concretize --force
 
-# We use GCC as a base compiler (c/c++/fortran) and implicitly, ROCm's hipcc when the +rocm variant is specified.
-PRODUCT_SPEC_LIST="
-ninja%gcc@13.2.1.genoa arch=linux-rhel8-zen4
-libyaml%gcc@13.2.1.genoa arch=linux-rhel8-zen4
-paraconf%gcc@13.2.1.genoa arch=linux-rhel8-zen4
-pdi%gcc@13.2.1.genoa+python arch=linux-rhel8-zen4
-pdiplugin-decl-hdf5%gcc@13.2.1.genoa arch=linux-rhel8-zen4
-pdiplugin-set-value%gcc@13.2.1.genoa arch=linux-rhel8-zen4
[…]
