+++ toolchains/mi250.hipcc.adastra.spack/environment.sh
@@ -1,41 +1,32 @@
 #!/bin/bash
 
-ENVIRONMENT_ROOT_DIRECTORY="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd -P)"
-
-module purge
+TOOLCHAIN_ROOT_DIRECTORY="$(dirname -- "$(readlink -f -- "${BASH_SOURCE[0]:-${0}}")")"
 
 SPACK_USER_VERSION="spack-user-4.0.0"
 
 export SPACK_USER_PREFIX="${SHAREDWORKDIR}/gyselalibxx-spack-install-MI250/Configuration.${SPACK_USER_VERSION}"
-module use "${SPACK_USER_PREFIX}/modules/tcl/linux-rhel8-zen3"
+export SPACK_USER_CACHE_PATH="${SPACK_USER_PREFIX}/cache"
 
-module load \
-    cmake \
-    gcc/13.2.1.mi250/zen3/libyaml \
-    gcc/13.2.1.mi250/zen3/paraconf \
-    gcc/13.2.1.mi250/zen3/pdi \
-    gcc/13.2.1.mi250/zen3/pdiplugin-decl-hdf5 \
-    gcc/13.2.1.mi250/zen3/pdiplugin-mpi \
-    gcc/13.2.1.mi250/zen3/pdiplugin-set-value \
-    gcc/13.2.1.mi250/zen3/pdiplugin-trace \
-    gcc/13.2.1.mi250/zen3/ginkgo \
-    gcc/13.2.1.mi250/zen3/eigen \
-    gcc/13.2.1.mi250/zen3/ninja \
-    gcc/13.2.1.mi250/zen3/py-matplotlib \
-    gcc/13.2.1.mi250/zen3/py-xarray \
-    gcc/13.2.1.mi250/zen3/py-h5py
-
-unset HIPCC_COMPILE_FLAGS_APPEND
+module purge
+module load "${SPACK_USER_VERSION}"
 
 module load cpe/24.07
-# FIXME:
-# craype-accel-amd-gfx90a Error with the cray wrappers clang: error: unsupported option '-fopenmp-targets=' for language mode 'HIP'
-module load craype-x86-trento
+module load craype-x86-trento craype-accel-amd-gfx90a
 # NOTE: Force 6.3.3 due to startup failures (https://github.com/gyselax/gyselalibxx/pull/198#issuecomment-2943081411)
 module load PrgEnv-gnu-amd amd-mixed/6.3.3
 module load rocm/6.3.3
 module load cray-fftw
-module load cray-hdf5-parallel
 module load cray-python
 
 module list
+
+which spack
+spack debug report
+
+eval "$(spack env activate --prompt --sh gyselalibxx-spack-environment)"
+
+unalias despacktivate
+unset despacktivate
+function despacktivate() {
+    eval "$(spack env deactivate --sh)"
+}

[…]
