### 2025-06-19 1058146
Remove non necessary products.
