+++ toolchains/mi250.hipcc.adastra.spack/environment.sh
@@ -10,12 +10,14 @@ export SPACK_USER_CACHE_PATH="${SPACK_USER_PREFIX}/cache"
 module purge
 module load "${SPACK_USER_VERSION}"
 
+module load cmake
 module load cpe/24.07
 module load craype-x86-trento craype-accel-amd-gfx90a
 # NOTE: Force 6.3.3 due to startup failures (https://github.com/gyselax/gyselalibxx/pull/198#issuecomment-2943081411)
 module load PrgEnv-gnu-amd amd-mixed/6.3.3
 module load rocm/6.3.3
 module load cray-fftw
+module load cray-hdf5-parallel
 module load cray-python
 
 module list
