+++ toolchains/mi250.hipcc.adastra.spack/gyselalibxx-spack-environment.yaml
@@ -1,26 +1,26 @@
 spack:
   definitions:
   - packages:
-    - cmake
+    # - cmake
     - ginkgo+openmp+rocm~shared amdgpu_target=gfx90a
     - libyaml
     - ninja
     - paraconf
     - pdi
     - pdiplugin-decl-hdf5
-    - pdiplugin-decl-netcdf
+    # - pdiplugin-decl-netcdf
     - pdiplugin-mpi
     - pdiplugin-set-value
     - pdiplugin-trace
 
-    - py-dask
+    # - py-dask
     - py-h5py
-    - py-imageio
+    # - py-imageio
     - py-matplotlib
-    - py-netcdf4
-    - py-numpy
-    - py-scipy
-    - py-sympy
+    # - py-netcdf4
+    # - py-numpy
+    # - py-scipy
+    # - py-sympy
     - py-xarray
   - compilers:
     - gcc@13.2.1.mi250
