+++ toolchains/genoa.gcc.adastra.spack/environment.sh
@@ -8,7 +8,14 @@ export SPACK_USER_PREFIX="${SHAREDWORKDIR}/gyselalibxx-spack-install-GENOA/Confi
 export SPACK_USER_CACHE_PATH="${SPACK_USER_PREFIX}/cache"
 
 module purge
+
 module load "${SPACK_USER_VERSION}"
+which spack
+spack debug report
+# Spack must work in a clean, purged environment so it can load modules without
+# having to purge itself (which it does not do..). When we spack env activate,
+# the same constraint applies.
+eval -- "$(spack env activate --prompt --sh gyselalibxx-spack-environment)"
 
 module load cmake
 module load cpe/24.07
@@ -21,11 +28,6 @@ module load cray-hdf5-parallel
 module load cray-python
 
 module list
-
-which spack
-spack debug report
-
-eval "$(spack env activate --prompt --sh gyselalibxx-spack-environment)"
 
 unalias despacktivate
 unset despacktivate
