### 2025-06-19 129824d
Enable all the products
