### 2025-06-20 1d1893a
update README.md for gyro average
