### 2025-06-20 61d1584
use build/evaluator as template parameters
