+++ tests/gyroaverage/gyroaverage_circular.cpp
@@ -56,6 +56,33 @@ struct GridBatch : UniformGridBase<Batch>
 {
 };
 
+template <class ExecutionSpace>
+using SplineRThetaBuilderType = ddc::SplineBuilder2D<
+        ExecutionSpace,
+        typename ExecutionSpace::memory_space,
+        BSplinesR,
+        BSplinesTheta,
+        GridR,
+        GridTheta,
+        SplineRBoundary, // boundary at r=0
+        SplineRBoundary, // boundary at rmax
+        SplineThetaBoundary,
+        SplineThetaBoundary,
+        ddc::SplineSolver::LAPACK>;
+
+template <class ExecutionSpace>
+using SplineRThetaEvaluatorNullBoundType = ddc::SplineEvaluator2D<
+        ExecutionSpace,
+        typename ExecutionSpace::memory_space,
+        BSplinesR,
+        BSplinesTheta,
+        GridR,
+        GridTheta,
+        ddc::NullExtrapolationRule, // boundary at r=0
+        ddc::NullExtrapolationRule, // boundary at rmax
+        ddc::PeriodicExtrapolationRule<Theta>,
+        ddc::PeriodicExtrapolationRule<Theta>>;
+
 using CoordR = ddc::Coordinate<R>;
 using CoordTheta = ddc::Coordinate<Theta>;
 using CoordRTheta = ddc::Coordinate<R, Theta>;
@@ -92,11 +119,12 @@ void initialise(
         double const kx,
         double const ky)
 {
-    IdxRangeRTheta const rtheta_mesh = get_idx_range<GridR, GridTheta>(A);
+    IdxRangeRTheta const rtheta_idx_range = get_idx_range<GridR, GridTheta>(A);
 
     // Initialise R, Z and rho_L
     ddc::parallel_for_each(
-            rtheta_mesh,
+            Kokkos::DefaultExecutionSpace(),
+            rtheta_idx_range,
             KOKKOS_LAMBDA(IdxRTheta const irtheta) {
                 IdxR const ir(irtheta);
                 IdxTheta const itheta(irtheta);
@@ -111,9 +139,10 @@ void initialise(
             });
 
     // Initialise A with Fourier mode
-    IdxRangeRThetaBatch const rthetabatch_mesh = get_idx_range<GridR, GridTheta, GridBatch>(A);
+    IdxRangeRThetaBatch const rthetabatch_idx_range = get_idx_range(A);
     ddc::parallel_for_each(
-            rthetabatch_mesh,
+            Kokkos::DefaultExecutionSpace(),
+            rthetabatch_idx_range,
             KOKKOS_LAMBDA(IdxRThetaBatch const irthetabatch) {
                 IdxR const ir(irthetabatch);
                 IdxTheta const itheta(irthetabatch);
@@ -233,32 +262,47 @@ struct CartesianToPolar
 
 TEST_P(GyroAverageCircularParamTests, TestPeriodicity)
 {
-    DConstFieldRThetaBatch A = get_const_field(this->m_A_alloc);
-    DFieldRThetaBatch A_bar = get_field(this->m_A_bar_alloc);
+    DConstFieldRThetaBatch A = get_const_field(m_A_alloc);
+    DFieldRThetaBatch A_bar = get_field(m_A_bar_alloc);
+
+    using SplineRThetaBuilder = SplineRThetaBuilderType<Kokkos::DefaultExecutionSpace>;
+    using SplineRThetaEvaluatorNullBound
+            = SplineRThetaEvaluatorNullBoundType<Kokkos::DefaultExecutionSpace>;
+
+    IdxRangeRTheta const rtheta_idx_range = get_idx_range<GridR, GridTheta>(A_bar);
[…]
