+++ src/gyroaverage/gyroaverage_operator.hpp
@@ -18,60 +18,28 @@ namespace detail {
  * The class uses 2D B-spline interpolation to evaluate the field at off-grid points along the orbit.
  * The operation is performed in parallel over the (r, theta) grid.
  *
- * @tparam ExecutionSpace      The Kokkos execution space.
- * @tparam GridRminor          The radial grid type.
- * @tparam GridTheta           The poloidal angle grid type.
+ * @tparam SplineRThetaBuilder      The type of the spline builder for the rtheta interpolation
+ * @tparam SplineRThetaEvaluator    The type of the spline evaluator for the rtheta interpolation
  * @tparam IdxRangeRminorThetaBatch The index range over R, Theta and Batch directions.
- * @tparam BSplinesRminor      The B-spline basis for the radial direction.
- * @tparam BSplinesTheta       The B-spline basis for the theta direction.
  * @tparam CoordinateTransformFunction The Coordinate transform Functor.
  */
 template <
-        class ExecutionSpace,
-        class GridRminor,
-        class GridTheta,
+        class SplineRThetaBuilder,
+        class SplineRThetaEvaluator,
         class IdxRangeRminorThetaBatch,
-        class BSplinesRminor,
-        class BSplinesTheta,
         class CoordinateTransformFunction>
 class GyroAverageOperator
 {
-    /**
-     * @brief R grid.
-     */
-    using GridRminorType = GridRminor;
-    /**
-     * @brief Theta grid.
-     */
-    using GridThetaType = GridTheta;
+    using ExecutionSpace = typename SplineRThetaBuilder::exec_space;
 
-    using DimensionRminorType = typename GridRminorType::continuous_dimension_type;
-    using DimensionThetaType = typename GridThetaType::continuous_dimension_type;
+    using GridRminor = typename SplineRThetaBuilder::interpolation_discrete_dimension_type1;
+    using GridTheta = typename SplineRThetaBuilder::interpolation_discrete_dimension_type2;
 
-    using SplineRThetaBuilder = ddc::SplineBuilder2D<
-            ExecutionSpace,
-            typename ExecutionSpace::memory_space,
-            BSplinesRminor,
-            BSplinesTheta,
-            GridRminor,
-            GridTheta,
-            ddc::BoundCond::GREVILLE, // boundary at r=0
-            ddc::BoundCond::GREVILLE, // boundary at rmax
-            ddc::BoundCond::PERIODIC,
-            ddc::BoundCond::PERIODIC,
-            ddc::SplineSolver::LAPACK>;
+    using Rminor = typename GridRminor::continuous_dimension_type;
+    using Theta = typename GridTheta::continuous_dimension_type;
 
-    using SplineRThetaEvaluatorNullBound = ddc::SplineEvaluator2D<
-            ExecutionSpace,
-            typename ExecutionSpace::memory_space,
-            BSplinesRminor,
-            BSplinesTheta,
-            GridRminor,
-            GridTheta,
-            ddc::NullExtrapolationRule, // boundary at r=0
-            ddc::NullExtrapolationRule, // boundary at rmax
-            ddc::PeriodicExtrapolationRule<DimensionThetaType>,
-            ddc::PeriodicExtrapolationRule<DimensionThetaType>>;
+    using BSplinesRminor = typename SplineRThetaBuilder::builder_type1::bsplines_type;
+    using BSplinesTheta = typename SplineRThetaBuilder::builder_type2::bsplines_type;
 
 public:
     using IdxRangeRminor = IdxRange<GridRminor>;
@@ -94,13 +62,23 @@ public:
     using DConstFieldRminorTheta = ConstField<double, IdxRangeRminorTheta>;
     using DConstFieldRminorThetaBatch = ConstField<double, IdxRangeRminorThetaBatch>;
 
[…]
