### 2025-06-20 7ac3913
All products as part of the spack environment + spack loads
