+++ toolchains/genoa.gcc.adastra.spack/environment.sh
@@ -13,21 +13,46 @@ spack debug report
 # Spack must work in a clean, purged environment so it can load modules without
 # having to purge itself or clearing environment variables (which it does not
 # do..). When we spack env activate, the same constraint applies.
-eval -- "$(spack env activate --prompt --sh gyselalibxx-spack-environment)"
+# Use spack load instead of an environment activation as it should limit the
+# inode produced by the environment's view.
+# eval -- "$(spack env activate --prompt --sh gyselalibxx-spack-environment)"
+# unalias despacktivate
+# unset despacktivate
+# function despacktivate() {
+#     eval "$(spack env deactivate --sh)"
+# }
 
-module load cmake
+eval -- "$(
+    spack \
+        --env gyselalibxx-spack-environment \
+        load --sh \
+        cmake \
+        fftw \
+        ginkgo \
+        googletest \
+        ninja \
+        paraconf \
+        pdi \
+        pdiplugin-decl-hdf5 \
+        pdiplugin-decl-netcdf \
+        pdiplugin-mpi \
+        pdiplugin-set-value \
+        pdiplugin-trace \
+        python \
+        py-dask \
+        py-h5py \
+        py-imageio \
+        py-matplotlib \
+        py-netcdf4 \
+        py-numpy \
+        py-scipy \
+        py-sympy \
+        py-xarray \
+        py-pyyaml
+)"
+
 module load cpe/24.07
 module load craype-x86-genoa
-
 module load PrgEnv-gnu
-module load cray-fftw
-module load cray-hdf5-parallel
-module load cray-python
 
 module list
-
-unalias despacktivate
-unset despacktivate
-function despacktivate() {
-    eval "$(spack env deactivate --sh)"
-}
