### 2025-06-20 90ca56f
Calculate o-point in DiscreteToCartesian constructor
