### 2025-06-20 4997a3d
Ensure o_point is a Coord<X, Y>. Add o_point method to curvilinear mappings
