+++ src/coord_transformations/czarny_to_cartesian.hpp
@@ -82,8 +82,7 @@ public:
 private:
     double m_epsilon;
     double m_e;
-    double m_x0;
-    double m_y0;
+    Coord<X, Y> m_o_point;
 
 public:
     /**
@@ -101,12 +100,10 @@ public:
     explicit KOKKOS_FUNCTION CzarnyToCartesian(
             double epsilon,
             double e,
-            double x0 = 0.0,
-            double y0 = 0.0)
+            Coord<X, Y> o_point = Coord<X,Y>(0.0, 0.0))
         : m_epsilon(epsilon)
         , m_e(e)
-        , m_x0(x0)
-        , m_y0(y0)
+        , m_o_point(o_point)
     {
     }
 
@@ -173,6 +170,19 @@ public:
     }
 
     /**
+     * @brief Get the O-point in Cartesian coordinates.
+     *
+     * Get the O-point of this mapping in Cartesian coordinates. This is calculated in
+     * the constructor.
+     *
+     * @return The O-point.
+     */
+    KOKKOS_INLINE_FUNCTION Coord<X, Y> o_point() const
+    {
+        return m_o_point;
+    }
+
+    /**
      * @brief Convert the @f$ (r, \theta) @f$ coordinate to the equivalent (x,y) coordinate.
      *
      * @param[in] coord The coordinate to be converted.
@@ -186,10 +196,10 @@ public:
         const double tmp1
                 = Kokkos::sqrt(m_epsilon * (m_epsilon + 2.0 * r * Kokkos::cos(theta)) + 1.0);
 
-        const double x = (1.0 - tmp1) / m_epsilon + m_x0;
+        const double x = (1.0 - tmp1) / m_epsilon + ddc::get<X>(m_o_point);
         const double y = m_e * r * Kokkos::sin(theta)
                                  / (Kokkos::sqrt(1.0 - 0.25 * m_epsilon * m_epsilon) * (2.0 - tmp1))
-                         + m_y0;
+                         + ddc::get<Y>(m_o_point);
 
         return Coord<X, Y>(x, y);
     }
@@ -403,7 +413,7 @@ public:
      */
     KOKKOS_INLINE_FUNCTION CartesianToCzarny<X, Y, R, Theta> get_inverse_mapping() const
     {
-        return CartesianToCzarny<X, Y, R, Theta>(m_epsilon, m_e, m_x0, m_y0);
+        return CartesianToCzarny<X, Y, R, Theta>(m_epsilon, m_e, m_o_point);
     }
 };
 
