### 2025-06-20 84365ad
Statically type arguments
