+++ src/coord_transformations/cartesian_to_circular.hpp
@@ -66,19 +66,16 @@ public:
     using Theta_cov = typename Theta::Dual;
 
 private:
-    double m_x0;
-    double m_y0;
+    Coord<X, Y> m_o_point;
 
 public:
     /**
      * @brief Instantiate a CartesianToCircular from parameters.
      *
-     * @param[in] x0 The x-coordinate of the centre of the circle (0 by default).
-     * @param[in] y0 The y-coordinate of the centre of the circle (0 by default).
+     * @param[in] o_point The (x,y)-coordinate of the centre of the circle ((0,0) by default).
      */
-    explicit KOKKOS_FUNCTION CartesianToCircular(double x0 = 0.0, double y0 = 0.0)
-        : m_x0(x0)
-        , m_y0(y0)
+    explicit KOKKOS_FUNCTION CartesianToCircular(Coord<X, Y> o_point = Coord<X,Y>(0.0, 0.0))
+        : m_o_point(o_point)
     {
     }
 
@@ -129,8 +126,8 @@ public:
      */
     KOKKOS_FUNCTION Coord<R, Theta> operator()(Coord<X, Y> const& coord) const
     {
-        const double x = ddc::get<X>(coord) - m_x0;
-        const double y = ddc::get<Y>(coord) - m_y0;
+        const double x = ddc::get<X>(coord) - ddc::get<X>(m_o_point);
+        const double y = ddc::get<Y>(coord) - ddc::get<Y>(m_o_point);
         const double r = Kokkos::sqrt(x * x + y * y);
         const double theta_pi_to_pi(Kokkos::atan2(y, x));
         const double theta = theta_pi_to_pi * (theta_pi_to_pi >= 0)
@@ -148,8 +145,8 @@ public:
      */
     KOKKOS_FUNCTION double jacobian(Coord<X, Y> const& coord)
     {
-        const double x = ddc::get<X>(coord) - m_x0;
-        const double y = ddc::get<Y>(coord) - m_y0;
+        const double x = ddc::get<X>(coord) - ddc::get<X>(m_o_point);
+        const double y = ddc::get<Y>(coord) - ddc::get<Y>(m_o_point);
         return 1. / Kokkos::sqrt(x * x + y * y);
     }
 
@@ -168,8 +165,8 @@ public:
             Coord<X, Y> const& coord) const
 
     {
-        const double x = ddc::get<X>(coord) - m_x0;
-        const double y = ddc::get<Y>(coord) - m_y0;
+        const double x = ddc::get<X>(coord) - ddc::get<X>(m_o_point);
+        const double y = ddc::get<Y>(coord) - ddc::get<Y>(m_o_point);
         DTensor<VectorIndexSet<R, Theta>, VectorIndexSet<X_cov, Y_cov>> jacobian_matrix;
         const double r2 = x * x + y * y;
         const double r = Kokkos::sqrt(r2);
@@ -194,8 +191,8 @@ public:
         static_assert(ddc::in_tags_v<IndexTag1, VectorIndexSet<R, Theta>>);
         static_assert(ddc::in_tags_v<IndexTag2, VectorIndexSet<X_cov, Y_cov>>);
 
-        const double x = ddc::get<X>(coord) - m_x0;
-        const double y = ddc::get<Y>(coord) - m_y0;
+        const double x = ddc::get<X>(coord) - ddc::get<X>(m_o_point);
+        const double y = ddc::get<Y>(coord) - ddc::get<Y>(m_o_point);
         if constexpr (std::is_same_v<IndexTag1, R> && std::is_same_v<IndexTag2, X_cov>) {
             // Component (1,1), i.e dr/dx
[…]
