+++ src/coord_transformations/circular_to_cartesian.hpp
@@ -75,8 +75,7 @@ public:
     /**
      * @brief Instantiate a CircularToCartesian from parameters.
      *
-     * @param[in] x0 The x-coordinate of the centre of the circle (0 by default).
-     * @param[in] y0 The y-coordinate of the centre of the circle (0 by default).
+     * @param[in] o_point The (x,y)-coordinate of the centre of the circle ((0,0) by default).
      */
     explicit KOKKOS_FUNCTION CircularToCartesian(Coord<X, Y> o_point = Coord<X,Y>(0.0, 0.0))
         : m_o_point(o_point)
