### 2025-06-20 2417ccb
Update toolchains/mi250.hipcc.adastra.spack/environment.sh
Co-authored-by: Thomas Padioleau <thomas.padioleau@cea.fr>
