+++ toolchains/mi250.hipcc.adastra.spack/environment.sh
@@ -52,7 +52,7 @@ eval -- "$(
 )"
 
 # Due to https://github.com/gyselax/gyselalibxx/pull/198#issuecomment-2943081411
-# we have a different spack to build dependencies and to build gyselalib. It is
+# we use a different HIP compiler to build dependencies and to build gyselalib. It is
 # fine except we must unset some environment variables defined by spack's HIP
 # packages so they do not interfere with HIPCC/clang in the 6.3.3 version.
 # Notably:
