### 2025-06-20 0337368
Update environment.sh
