### 2025-06-21 f7230b4
Use Kokkos provided by Spack
