+++ CMakeLists.txt
@@ -82,15 +82,15 @@ if("${GYSELALIBXX_INCLUDE_TESTING_DEPENDENCIES}")
   endif()
 endif()
 
-find_package(Kokkos 4.4.1 QUIET)
-if(NOT "${Kokkos_FOUND}")
-  ## Use Kokkos from `vendor/`
-  if("${Kokkos_ENABLE_CUDA}")
-    option(Kokkos_ENABLE_CUDA_CONSTEXPR "Whether to activate experimental relaxed constexpr functions" ON)
-    option(Kokkos_ENABLE_CUDA_RELOCATABLE_DEVICE_CODE "Whether to enable relocatable device code (RDC) for CUDA" ON)
-  endif()
-  add_subdirectory("vendor/kokkos/" "kokkos" SYSTEM)
-endif()
+find_package(Kokkos 4.4.1 REQUIRED)
+# if(NOT "${Kokkos_FOUND}")
+#   ## Use Kokkos from `vendor/`
+#   if("${Kokkos_ENABLE_CUDA}")
+#     option(Kokkos_ENABLE_CUDA_CONSTEXPR "Whether to activate experimental relaxed constexpr functions" ON)
+#     option(Kokkos_ENABLE_CUDA_RELOCATABLE_DEVICE_CODE "Whether to enable relocatable device code (RDC) for CUDA" ON)
+#   endif()
+#   add_subdirectory("vendor/kokkos/" "kokkos" SYSTEM)
+# endif()
 
 find_package(KokkosKernels 4.5.1 QUIET)
 if(NOT "${KokkosKernels_FOUND}")
