### 2025-06-23 5cdd242
Stop naming 'break points' as 'knots' (#341)
Fix break points incorrectly labelled as knots. Fixes #69
