### 2025-06-23 e668cef
Use typing information
