+++ tests/coord_transformations/metric_tensor_evaluator.cpp
@@ -88,8 +88,8 @@ TEST_P(InverseMetricTensor3D, InverseMatrixToroidalMap)
     using ToroidalMapping = ToroidalToCylindrical<Mapping2D, Zeta, Phi>;
     using CylindricalMapping = CylindricalToCartesian<R, Z, Zeta, X, Y>;
     using Mapping = CombinedMapping<CylindricalMapping, ToroidalMapping, Coord<Rho, Theta, Phi>>;
-    double major_radius = 6.2;
-    Mapping2D polar_to_RZ(major_radius);
+    Coord<R, Z> o_point(6.2, 0.0);
+    Mapping2D polar_to_RZ(o_point);
     ToroidalMapping toroidal_to_cylindrical(polar_to_RZ);
     CylindricalMapping cylindrical_to_cartesian;
     Mapping mapping(cylindrical_to_cartesian, toroidal_to_cylindrical);
