### 2025-06-23 ad22566
Fix Spack environment
Add changelog entry
