### 2025-06-23 51fbc0b
Clang formatting
