+++ src/interpolation/polar_splines/polar_bsplines.hpp
@@ -54,10 +54,10 @@ public:
 
 
     /// The tag for the radial direction of the B-splines.
-    using DimR = typename BSplinesR::continuous_dimension_type;
+    using R = typename BSplinesR::continuous_dimension_type;
 
     /// The tag for the poloidal direction of the B-splines.
-    using DimTheta = typename BSplinesTheta::continuous_dimension_type;
+    using Theta = typename BSplinesTheta::continuous_dimension_type;
 
 public:
     /// The continuity enforced by the B-splines at the singular point.
@@ -248,23 +248,22 @@ public:
         {
             static_assert(is_accessible_v<Kokkos::DefaultHostExecutionSpace, DiscreteMapping>);
             static_assert(std::is_same_v<MemorySpace, Kokkos::HostSpace>);
-            using DimX = typename DiscreteMapping::cartesian_tag_x;
-            using DimY = typename DiscreteMapping::cartesian_tag_y;
+            using X = typename DiscreteMapping::cartesian_tag_x;
+            using Y = typename DiscreteMapping::cartesian_tag_y;
             using mapping_tensor_product_index_type
                     = Idx<typename DiscreteMapping::BSplineR,
                           typename DiscreteMapping::BSplineTheta>;
             if constexpr (C > -1) {
-                const Coord<DimX, DimY> pole
-                        = curvilinear_to_cartesian(Coord<DimR, DimTheta>(0.0, 0.0));
-                const double x0 = ddc::get<DimX>(pole);
-                const double y0 = ddc::get<DimY>(pole);
+                const Coord<X, Y> pole = curvilinear_to_cartesian(Coord<R, Theta>(0.0, 0.0));
+                const double x0 = ddc::get<X>(pole);
+                const double y0 = ddc::get<Y>(pole);
                 double tau = 0.0;
                 for (std::size_t i(0); i < ddc::discrete_space<BSplinesTheta>().size(); ++i) {
-                    const Coord<DimX, DimY> point = curvilinear_to_cartesian.control_point(
+                    const Coord<X, Y> point = curvilinear_to_cartesian.control_point(
                             mapping_tensor_product_index_type(1, i));
 
-                    const double c_x = ddc::get<DimX>(point);
-                    const double c_y = ddc::get<DimY>(point);
+                    const double c_x = ddc::get<X>(point);
+                    const double c_y = ddc::get<Y>(point);
 
                     double tau1 = -2.0 * (c_x - x0);
                     double tau2 = c_x - x0 - sqrt(3.0) * (c_y - y0);
@@ -274,11 +273,11 @@ public:
                     tau = tau > tau3 ? tau : tau3;
                 }
                 // Determine the corners for the barycentric coordinates
-                const Coord<DimX, DimY> corner1(x0 + tau, y0);
-                const Coord<DimX, DimY> corner2(x0 - 0.5 * tau, y0 + 0.5 * tau * sqrt(3.0));
-                const Coord<DimX, DimY> corner3(x0 - 0.5 * tau, y0 - 0.5 * tau * sqrt(3.0));
+                const Coord<X, Y> corner1(x0 + tau, y0);
+                const Coord<X, Y> corner2(x0 - 0.5 * tau, y0 + 0.5 * tau * sqrt(3.0));
+                const Coord<X, Y> corner3(x0 - 0.5 * tau, y0 - 0.5 * tau * sqrt(3.0));
 
[…]
