### 2025-06-23 3ca609a
Break dependency between EvalSpace for DiscreteToCartesian mapping and for the constructor of BSplines
