### 2025-06-23 f451108
KOKKOS_LAMBDA cannot be in anonymous function
