### 2025-06-23 de4ee1b
KOKKOS_LAMBDA cannot be in anonymous function
