+++ src/coord_transformations/czarny_to_cartesian.hpp
@@ -92,8 +92,7 @@ public:
      * 			The @f$ \epsilon @f$ parameter in the definition of the mapping CzarnyToCartesian.
      * @param[in] e
      * 			The @f$ e @f$ parameter in the definition of the mapping CzarnyToCartesian.
-     * @param[in] x0 The x-coordinate of the centre of the circle (0 by default).
-     * @param[in] y0 The y-coordinate of the centre of the circle (0 by default).
+     * @param[in] o_point The (x,y)-coordinate of the centre of the circle ((0,0) by default).
      *
      * @see CzarnyToCartesian
      */
