+++ tests/multipatch/interface_derivatives/interface_derivative_matrix_test.cpp
@@ -1,0 +1,1141 @@
+// SPDX-License-Identifier: MIT
+#include <ddc/ddc.hpp>
+#include <ddc/kernels/splines.hpp>
+
+#include <gtest/gtest.h>
+
+// #include "2patches_2d_onion_shape_non_uniform.hpp"
+// #include "../../test_utils.hpp"
+
+#include "9patches_2d_periodic_strips_non_uniform.hpp"
+#include "ddc_alias_inline_functions.hpp"
+#include "ddc_helper.hpp"
+// #include "global_2d_onion_shape_non_uniform.hpp"
+
+
+#include <typeinfo>
+
+#include <boost/type_index.hpp>
+
+#include "interface.hpp"
+#include "interface_derivative_matrix.hpp"
+#include "mesh_builder.hpp"
+#include "non_uniform_interpolation_points.hpp"
+#include "single_interface_derivatives_calculator.hpp"
+
+
+/*
+    Test: 
+    * 3 patches
+    * compute derivatives x1, x2, x1x2? (or just function + derivatives?)
+    * which interpolation methods to choose? Additional interpolation points? 
+*/
+
+namespace {
+// Multi-patch ---
+using namespace periodic_strips_non_uniform_2d_9patches;
+// using namespace onion_shape_non_uniform_2d_global;
+
+// Equivalent global mesh ---
+struct Xg
+{
+    static bool constexpr PERIODIC = true;
+    static bool constexpr IS_COVARIANT = true;
+    static bool constexpr IS_CONTRAVARIANT = true;
+    using Dual = Xg;
+};
+
+struct Yg
+{
+    static bool constexpr PERIODIC = false;
+    static bool constexpr IS_COVARIANT = true;
+    static bool constexpr IS_CONTRAVARIANT = true;
+    using Dual = Yg;
+};
+
+struct GridXg : NonUniformGridBase<Xg>
+{
+};
+struct GridYg : NonUniformGridBase<Yg>
+{
+};
+
+struct BSplinesXg : ddc::NonUniformBSplines<Xg, 3>
+{
+};
+struct BSplinesYg : ddc::NonUniformBSplines<Yg, 3>
+{
+};
+
+
+using HostExecSpace = Kokkos::DefaultHostExecutionSpace;
+
+
+// Interpolation points type for the patches
+template <std::size_t PatchIdx>
+using SplineInterpPointsX = ddcHelper::NonUniformInterpolationPoints<
+        BSplinesX<PatchIdx>,
+        ddc::BoundCond::HERMITE,
+        ddc::BoundCond::HERMITE>;
+
+template <std::size_t PatchIdx, ddc::BoundCond BoundCondMin, ddc::BoundCond BoundCondMax>
+using SplineInterpPointsY
+        = ddcHelper::NonUniformInterpolationPoints<BSplinesY<1>, BoundCondMin, BoundCondMax>;
+
+// Interpolation points type for the equivalent global spline.
+using SplineInterpPointsXg = ddcHelper::NonUniformInterpolationPoints<
+        BSplinesXg,
+        ddc::BoundCond::PERIODIC,
+        ddc::BoundCond::PERIODIC>;
+using SplineInterpPointsYg = ddcHelper::NonUniformInterpolationPoints<
+        BSplinesYg,
+        ddc::BoundCond::GREVILLE,
+        ddc::BoundCond::GREVILLE>;
+
+
+// Operators on the equivalent global spline
+using SplineRThetagBuilder = ddc::SplineBuilder2D<
+        HostExecSpace,
+        typename HostExecSpace::memory_space,
+        BSplinesXg,
+        BSplinesYg,
+        GridXg,
[…]
