+++ tests/multipatch/geometries/9patches_2d_periodic_strips_non_uniform.hpp
@@ -1,0 +1,222 @@
+// SPDX-License-Identifier: MIT
+
+/*
+    Geometry defined here: 9 * 2D patches.
+        - for all patches: 
+            - non-periodic on X,Y;
+            - non_uniform cubic splines;
+            - non_uniform Grid1 and Grid2; 
+        - across patches:
+            - periodic on X;
+            - non-periodic on Y;
+
+
+      1  |  2  |  3
+    -----------------
+      4  |  5  |  6
+    -----------------
+      7  |  8  |  9
+*/
+#pragma once
+
+#include <ddc/ddc.hpp>
+#include <ddc/kernels/splines.hpp>
+
+#include "connectivity.hpp"
+#include "edge.hpp"
+#include "interface.hpp"
+#include "patch.hpp"
+
+namespace periodic_strips_non_uniform_2d_9patches {
+
+int constexpr BSplineDegree = 3;
+
+// CONTINUOUS DIMENSIONS -------------------------------------------------------------------------
+/**
+ * @brief First continuous dimension of patch PatchIdx.
+ * @tparam PatchIdx Index of the patch. 
+ */
+struct X
+{
+    /// @brief Non periodic dimension.
+    static bool constexpr PERIODIC = false;
+    /// A boolean indicating if dimension describes a covariant coordinate.
+    static bool constexpr IS_COVARIANT = true;
+    /// A boolean indicating if dimension describes a contravariant coordinate.
+    static bool constexpr IS_CONTRAVARIANT = true;
+    /// A type-alias mapping to the covariant counterpart.
+    using Dual = X;
+};
+
+/**
+ * @brief Second continuous dimension of patch PatchIdx.
+ * @tparam PatchIdx Index of the patch. 
+ */
+struct Y
+{
+    /// @brief Non periodic dimension.
+    static bool constexpr PERIODIC = false;
+    /// A boolean indicating if dimension describes a covariant coordinate.
+    static bool constexpr IS_COVARIANT = true;
+    /// A boolean indicating if dimension describes a contravariant coordinate.
+    static bool constexpr IS_CONTRAVARIANT = true;
+    /// A type-alias mapping to the covariant counterpart.
+    using Dual = Y;
+};
+
+
+
+// DISCRETE DIMENSIONS ---------------------------------------------------------------------------
+/**
+ * @brief Points sequence on the second logical dimension of patch 1.
+ * @tparam PatchIdx Index of the patch. 
+ */
+template <int PatchIdx>
+struct GridX : NonUniformGridBase<X>
+{
+};
+
+/**
+ * @brief Points sequence on the second logical dimension of patch 1.
+ * @tparam PatchIdx Index of the patch. 
+ */
+template <int PatchIdx>
+struct GridY : NonUniformGridBase<Y>
+{
+};
+
+
+
+// SPLINE DIMENSIONS -----------------------------------------------------------------------------
+/**
+ * @brief First spline dimension of patch PatchIdx.
+ * @tparam PatchIdx Index of the patch. 
+ */
+template <int PatchIdx>
+struct BSplinesX : ddc::NonUniformBSplines<X, BSplineDegree>
+{
+};
+
+/**
+ * @brief Second spline dimension of patch PatchIdx.
+ * @tparam PatchIdx Index of the patch. 
+ */
+template <int PatchIdx>
+struct BSplinesY : ddc::NonUniformBSplines<Y, BSplineDegree>
+{
+};
[…]
