### 2025-06-24 cd07336
Apply suggestions from code review
Co-authored-by: Emily Bourne <emily.bourne@epfl.ch>
