+++ docs/joss_paper/paper.bib
@@ -74,12 +74,20 @@ author = {Emily Bourne and Yann Munschy and Virginie Grandgirard and Michel Mehr
 keywords = {Semi-Lagrangian scheme, B-splines, Non-equidistant meshes, Vlasov-Poisson equations, Plasma sheath simulation, GPUs},
 }
 
-@conference{trilaksono2023,
-    author = {Jordy Trilaksono and Philipp Ulbl and Andreas Stegmeir and Frank Jenko},
-    title = {A Directive-Based, Fortran/C++ Interoperable Approach to GPU Offloading of the High Performance Gyrokinetic Turbulence Code GENE-X},
-    booktitle = {Platform for Advanced Scientific Computing (PASC) Conference},
-    year = {2023},
-    note = {"https://pasc23.pasc-conference.org/wp-content/uploads/2023/07/pos146-P18-Directive-Based-FortranC-Interoperable-Approach-to-GPU-Offloading-of-the-High-Performance-Gyrokinetic-Turbulence-Code-GENE-X.pdf"}
+@inproceedings{trilaksono2025,
+  author = {Trilaksono, Jordy and Ulbl, Philipp and Williams, Jeremy and Pfeiler, Carl-Martin and Finkbeiner, Marion and Dannert, Tilman and Laure, Erwin and Markidis, Stefano and Jenko, Frank},
+  title = {OpenACC and OpenMP-Accelerated Fortran/C++ Gyrokinetic Fusion Code GENE-X for Heterogeneous Architectures},
+  year = {2025},
+  isbn = {9798400718861},
+  publisher = {Association for Computing Machinery},
+  url = {https://doi.org/10.1145/3732775.3733587},
+  doi = {10.1145/3732775.3733587},
+  booktitle = {Proceedings of the Platform for Advanced Scientific Computing Conference},
+  pages = {1–12},
+  numpages = {12},
+  keywords = {GPU, OpenACC, OpenMP, MPI, CUDA-aware MPI, NVIDIA, fortran, C++, fusion plasma, gyrokinetics, turbulence},
+  location = {FHNW University of Applied Sciences and Arts Northwestern Switzerland, Brugg-Windisch, Switzerland},
+  series = {PASC '25}
 }
 
 @article{trott2022,
