### 2025-06-24 c868276
Work with Interfaces "not well oriented".
