+++ tests/multipatch/interface_derivatives/interface_derivative_matrix_test.cpp
@@ -22,6 +22,7 @@
 #include "mesh_builder.hpp"
 #include "non_uniform_interpolation_points.hpp"
 #include "single_interface_derivatives_calculator.hpp"
+#include "interface_derivatives_test_utils.hpp"
 
 
 /*
@@ -34,7 +35,6 @@
 namespace {
 // Multi-patch ---
 using namespace periodic_strips_non_uniform_2d_9patches;
-// using namespace onion_shape_non_uniform_2d_global;
 
 // Equivalent global mesh ---
 struct Xg
@@ -120,66 +120,68 @@ using SplineRThetagEvaluator = ddc::SplineEvaluator2D<
         ddc::ConstantExtrapolationRule<Yg, Xg>>;
 
 
-/**
- *  @brief Get interpolation points from the break points by placing 
- * the interpolation points on the break points and adding one on the
- * left boundary cell at 2/3 of the cell. 
- */
-template <class CoordType>
-std::vector<CoordType> get_interpolation_points_add_one_on_left(
-        std::vector<CoordType> const& break_points)
-{
-    CoordType additional_point(break_points[0] * 2. / 3. + break_points[1] * 1. / 3.);
-    std::vector<CoordType> interpolation_points(break_points);
-    interpolation_points.insert(interpolation_points.begin() + 1, additional_point);
-    return interpolation_points;
-}
+// TODO: Add the functions defined here in another file. 
 
-/**
- *  @brief Get interpolation points from the break points by placing 
- * the interpolation points on the break points and adding one on the
- * right boundary cell at 1/3 of the cell. 
- */
-template <class CoordType>
-std::vector<CoordType> get_interpolation_points_add_one_on_right(
-        std::vector<CoordType> const& break_points)
-{
-    int n_bpoints = break_points.size();
-    CoordType additional_point(
-            break_points[n_bpoints - 1] * 2. / 3. + break_points[n_bpoints - 2] * 1. / 3.);
-    std::vector<CoordType> interpolation_points(break_points);
-    interpolation_points.insert(interpolation_points.end() - 1, additional_point);
-    return interpolation_points;
-}
+// /**
+//  *  @brief Get interpolation points from the break points by placing 
+//  * the interpolation points on the break points and adding one on the
+//  * left boundary cell at 2/3 of the cell. 
+//  */
+// template <class CoordType>
+// std::vector<CoordType> get_interpolation_points_add_one_on_left(
+//         std::vector<CoordType> const& break_points)
+// {
+//     CoordType additional_point(break_points[0] * 2. / 3. + break_points[1] * 1. / 3.);
+//     std::vector<CoordType> interpolation_points(break_points);
+//     interpolation_points.insert(interpolation_points.begin() + 1, additional_point);
+//     return interpolation_points;
+// }
 
-/**
- * @brief Fill in a vector of points for the equivalent global mesh 
- * by conserving the same order of the given points.
- */
-template <class CoordTypeG, class CoordTypeP>
-void fill_in(std::vector<CoordTypeG>& points_global, std::vector<CoordTypeP> const& points_patch)
-{
-    for (CoordTypeP pt : points_patch) {
-        points_global.push_back(CoordTypeG {double(pt)});
-    }
-}
+// /**
+//  *  @brief Get interpolation points from the break points by placing 
+//  * the interpolation points on the break points and adding one on the
+//  * right boundary cell at 1/3 of the cell. 
+//  */
+// template <class CoordType>
+// std::vector<CoordType> get_interpolation_points_add_one_on_right(
+//         std::vector<CoordType> const& break_points)
+// {
+//     int n_bpoints = break_points.size();
[…]
