+++ tests/multipatch/interface_derivatives/single_interface_derivatives_calculator_test.cpp
@@ -18,6 +18,8 @@
 #include "mesh_builder.hpp"
 #include "non_uniform_interpolation_points.hpp"
 #include "single_interface_derivatives_calculator.hpp"
+#include "interface_derivatives_test_utils.hpp"
+
 
 
 /*
@@ -132,66 +134,66 @@ using SplineRThetagEvaluator = ddc::SplineEvaluator2D<
         ddc::PeriodicExtrapolationRule<Thetag>>;
 
 
-/**
- *  @brief Get interpolation points from the break points by placing 
- * the interpolation points on the break points and adding one on the
- * left boundary cell at 2/3 of the cell. 
- */
-template <class CoordType>
-std::vector<CoordType> get_interpolation_points_add_one_on_left(
-        std::vector<CoordType> const& break_points)
-{
-    CoordType additional_point(break_points[0] * 2. / 3. + break_points[1] * 1. / 3.);
-    std::vector<CoordType> interpolation_points(break_points);
-    interpolation_points.insert(interpolation_points.begin() + 1, additional_point);
-    return interpolation_points;
-}
+// /**
+//  *  @brief Get interpolation points from the break points by placing 
+//  * the interpolation points on the break points and adding one on the
+//  * left boundary cell at 2/3 of the cell. 
+//  */
+// template <class CoordType>
+// std::vector<CoordType> get_interpolation_points_add_one_on_left(
+//         std::vector<CoordType> const& break_points)
+// {
+//     CoordType additional_point(break_points[0] * 2. / 3. + break_points[1] * 1. / 3.);
+//     std::vector<CoordType> interpolation_points(break_points);
+//     interpolation_points.insert(interpolation_points.begin() + 1, additional_point);
+//     return interpolation_points;
+// }
 
-/**
- *  @brief Get interpolation points from the break points by placing 
- * the interpolation points on the break points and adding one on the
- * right boundary cell at 1/3 of the cell. 
- */
-template <class CoordType>
-std::vector<CoordType> get_interpolation_points_add_one_on_right(
-        std::vector<CoordType> const& break_points)
-{
-    int n_bpoints = break_points.size();
-    CoordType additional_point(
-            break_points[n_bpoints - 1] * 2. / 3. + break_points[n_bpoints - 2] * 1. / 3.);
-    std::vector<CoordType> interpolation_points(break_points);
-    interpolation_points.insert(interpolation_points.end() - 1, additional_point);
-    return interpolation_points;
-}
+// /**
+//  *  @brief Get interpolation points from the break points by placing 
+//  * the interpolation points on the break points and adding one on the
+//  * right boundary cell at 1/3 of the cell. 
+//  */
+// template <class CoordType>
+// std::vector<CoordType> get_interpolation_points_add_one_on_right(
+//         std::vector<CoordType> const& break_points)
+// {
+//     int n_bpoints = break_points.size();
+//     CoordType additional_point(
+//             break_points[n_bpoints - 1] * 2. / 3. + break_points[n_bpoints - 2] * 1. / 3.);
+//     std::vector<CoordType> interpolation_points(break_points);
+//     interpolation_points.insert(interpolation_points.end() - 1, additional_point);
+//     return interpolation_points;
+// }
 
-/**
- * @brief Fill in a vector of points for the equivalent global mesh 
- * by conserving the same order of the given points.
- */
-template <class CoordTypeG, class CoordTypeP>
-void fill_in(std::vector<CoordTypeG>& points_global, std::vector<CoordTypeP> const& points_patch)
-{
-    for (CoordTypeP pt : points_patch) {
[…]

[…]
