### 2025-06-24 f052046
Activate MPICH_GPU_SUPPORT_ENABLED
