+++ toolchains/mi250.hipcc.adastra.spack/environment.sh
@@ -36,3 +36,5 @@ module load \
     gcc/13.2.1.mi250/zen3/py-h5py
 
 module list
+
+export MPICH_GPU_SUPPORT_ENABLED=1
