+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Curvilinear coordinate change classes have an O-point method to retrieve the O-point in the non-curvilinear coordinates.
 - Add a batched `operator()` to `DiscreteToCartesian` allowing a field of coordinates to be converted.
+- Add a `control_points` method to `DiscreteToCartesian` to allow all control points to be retrieved at once.
 
 ### Fixed
 
@@ -30,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generalise `compute_coeffs_on_mapping` to work with any mapping.
 - Rely on GPU-aware MPI to allow GPU-direct MPI for `MPITransposeAllToAll`.
 - Curvilinear coordinate change classes take a `Coord` type to specify the O-point in the constructor.
+- Allow `init_discrete_space` to be used to initialise `PolarBSplines` with a `DiscreteToCartesian` coordinate change, accessible only on GPU.
 
 ### Deprecated
 
