### 2025-06-24 #8304 @EmilyBourne
@kyleniemeyer Hello, I just wanted to check in on this submission, which was moved to your track (3, PE) about a month ago. I understand that things are busy, but I wanted to make sure no notifications went astray. Please let me know if there’s anything I could be doing on my end. Also, if you have a rough idea of the typical queue length for Physics track submissions, that would be helpful to know so I don’t follow up unnecessarily early. Thanks!
