### 2025-06-24 #8304 @kyleniemeyer
Hi @EmilyBourne - sorry for the delay. We do have a fairly substantial queue in the track (nearly 40 submissions), and not many available editors at the moment. It's hard to really estimate how long it will take, but I'm going to be working on pushing things along.
