### 2025-06-25 e2d4352
Use EXPECT_DOUBLE_EQ instead of EXPECT_TRUE that expects true equality
