+++ tests/coord_transformations/jacobian_matrix_coef.cpp
@@ -55,7 +55,18 @@ TEST_P(JacobianMatrixAndJacobianCoefficients, MatrixCircMap)
         ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)
                 = mapping.template jacobian_component<Y, Theta_cov>(coords(irtheta));
 
-        EXPECT_TRUE(Jacobian_matrix == Jacobian_matrix_coeff);
+        EXPECT_DOUBLE_EQ(
+                (ddcHelper::get<X, R_cov>(Jacobian_matrix)),
+                (ddcHelper::get<X, R_cov>(Jacobian_matrix_coeff)));
+        EXPECT_DOUBLE_EQ(
+                (ddcHelper::get<X, Theta_cov>(Jacobian_matrix)),
+                (ddcHelper::get<X, Theta_cov>(Jacobian_matrix_coeff)));
+        EXPECT_DOUBLE_EQ(
+                (ddcHelper::get<Y, R_cov>(Jacobian_matrix)),
+                (ddcHelper::get<Y, R_cov>(Jacobian_matrix_coeff)));
+        EXPECT_DOUBLE_EQ(
+                (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix)),
+                (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)));
     });
 
     // --- for the inverse Jacobian matrix:
@@ -220,7 +231,18 @@ TEST_P(JacobianMatrixAndJacobianCoefficients, MatrixDiscCzarMap)
             ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)
                     = mapping.template jacobian_component<Y, Theta_cov>(coord_rtheta);
 
-            EXPECT_TRUE(Jacobian_matrix == Jacobian_matrix_coeff);
+            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<X, R_cov>(Jacobian_matrix)),
+                    (ddcHelper::get<X, R_cov>(Jacobian_matrix_coeff)));
+            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<X, Theta_cov>(Jacobian_matrix)),
+                    (ddcHelper::get<X, Theta_cov>(Jacobian_matrix_coeff)));
+            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<Y, R_cov>(Jacobian_matrix)),
+                    (ddcHelper::get<Y, R_cov>(Jacobian_matrix_coeff)));
+            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix)),
+                    (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)));
 
             // --- for the inverse Jacobian matrix:
             Tensor inv_Jacobian_matrix = inv_jacobian(coord_rtheta);
@@ -238,7 +260,18 @@ TEST_P(JacobianMatrixAndJacobianCoefficients, MatrixDiscCzarMap)
             ddcHelper::get<Theta, Y_cov>(inv_Jacobian_matrix_coeff)
                     = inv_jacobian.template inv_jacobian_component<Theta, Y_cov>(coord_rtheta);
 
-            EXPECT_TRUE(inv_Jacobian_matrix == inv_Jacobian_matrix_coeff);
+            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<R, X_cov>(inv_Jacobian_matrix)),
+                    (ddcHelper::get<R, X_cov>(inv_Jacobian_matrix_coeff)));
+            EXPECT_DOUBLE_EQ(
[…]
