### 2025-06-25 #8304 @EmilyBourne
> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

I would like to suggest:
- [paulwilhelmvlasov](https://github.com/paulwilhelmvlasov)
- [pratikvn](https://github.com/pratikvn)
- [joglekara](https://github.com/joglekara)

If you are also open to suggestions of potential reviewers who are not necessarily familiar with JOSS then I would also like to suggest:
- [jordytrilaksono](https://github.com/jordytrilaksono)

I can provide his email address if necessary.
