### 2025-06-26 4371860
Use EXPECT_NEAR instead of EXPECT_DOUBLE_EQ
