+++ tests/coord_transformations/jacobian_matrix_coef.cpp
@@ -55,18 +55,22 @@ TEST_P(JacobianMatrixAndJacobianCoefficients, MatrixCircMap)
         ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)
                 = mapping.template jacobian_component<Y, Theta_cov>(coords(irtheta));
 
-        EXPECT_DOUBLE_EQ(
+        EXPECT_NEAR(
                 (ddcHelper::get<X, R_cov>(Jacobian_matrix)),
-                (ddcHelper::get<X, R_cov>(Jacobian_matrix_coeff)));
-        EXPECT_DOUBLE_EQ(
+                (ddcHelper::get<X, R_cov>(Jacobian_matrix_coeff)),
+                1e-15);
+        EXPECT_NEAR(
                 (ddcHelper::get<X, Theta_cov>(Jacobian_matrix)),
-                (ddcHelper::get<X, Theta_cov>(Jacobian_matrix_coeff)));
-        EXPECT_DOUBLE_EQ(
+                (ddcHelper::get<X, Theta_cov>(Jacobian_matrix_coeff)),
+                1e-15);
+        EXPECT_NEAR(
                 (ddcHelper::get<Y, R_cov>(Jacobian_matrix)),
-                (ddcHelper::get<Y, R_cov>(Jacobian_matrix_coeff)));
-        EXPECT_DOUBLE_EQ(
+                (ddcHelper::get<Y, R_cov>(Jacobian_matrix_coeff)),
+                1e-15);
+        EXPECT_NEAR(
                 (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix)),
-                (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)));
+                (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)),
+                1e-15);
     });
 
     // --- for the inverse Jacobian matrix:
@@ -231,18 +235,22 @@ TEST_P(JacobianMatrixAndJacobianCoefficients, MatrixDiscCzarMap)
             ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)
                     = mapping.template jacobian_component<Y, Theta_cov>(coord_rtheta);
 
-            EXPECT_DOUBLE_EQ(
+            EXPECT_NEAR(
                     (ddcHelper::get<X, R_cov>(Jacobian_matrix)),
-                    (ddcHelper::get<X, R_cov>(Jacobian_matrix_coeff)));
-            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<X, R_cov>(Jacobian_matrix_coeff)),
+                    1e-15);
+            EXPECT_NEAR(
                     (ddcHelper::get<X, Theta_cov>(Jacobian_matrix)),
-                    (ddcHelper::get<X, Theta_cov>(Jacobian_matrix_coeff)));
-            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<X, Theta_cov>(Jacobian_matrix_coeff)),
+                    1e-15);
+            EXPECT_NEAR(
                     (ddcHelper::get<Y, R_cov>(Jacobian_matrix)),
-                    (ddcHelper::get<Y, R_cov>(Jacobian_matrix_coeff)));
-            EXPECT_DOUBLE_EQ(
+                    (ddcHelper::get<Y, R_cov>(Jacobian_matrix_coeff)),
+                    1e-15);
+            EXPECT_NEAR(
                     (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix)),
-                    (ddcHelper::get<Y, Theta_cov>(Jacobian_matrix_coeff)));
[…]
