### 2025-06-27 154c2b5
Avoid misuse of Levi-Civita tensor. Fixes #348
