+++ src/data_types/static_tensors.hpp
@@ -211,6 +211,8 @@ public:
         if constexpr (is_covariant_vector_index_set_v<ValidIndexSet>) {
             m_coeff = jacobian;
         } else {
+            // Evaluating at a singular point will lead to NaN
+            assert(fabs(jacobian) >= 1e-19);
             m_coeff = 1.0 / jacobian;
         }
     }
@@ -287,6 +289,9 @@ KOKKOS_INLINE_FUNCTION constexpr double get(
 } // namespace ddcHelper
 
 namespace detail {
+
+template <class ElementType, class ValidIndexSetRow, class ValidIndexSetCol>
+inline constexpr bool enable_tensor_type<IdentityTensor<ElementType, ValidIndexSetRow, ValidIndexSetCol>> = true;
 
 template <class ElementType, class ValidIndexSet>
 inline constexpr bool enable_tensor_type<LeviCivitaTensor<ElementType, ValidIndexSet>> = true;
