### 2025-06-27 32ed70b
Fix minimum version requirement of Kokkos and tolerance of floating point comparisons in some tests (#344)
- Fix minimum version requirement of Kokkos.
- Fix tolerance of floating point comparisons in
JacobianMatrixAndJacobianCoefficients and MultipatchSplineEvaluatorTest
tests

---------

Co-authored-by: Emily Bourne <emily.bourne@epfl.ch>
