### 2025-06-27 36276f2
Improve docstrings following suggestions
