### 2025-06-27 01d5aa1
(x_0,y_0) is not the o_point
