+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Curvilinear coordinate change classes have an o-point method to retrieve the o-point in the non-curvilinear coordinates.
+- Curvilinear coordinate change classes have an O-point method to retrieve the O-point in the non-curvilinear coordinates.
 - Add a batched `operator()` to `DiscreteToCartesian` allowing a field of coordinates to be converted.
 
 ### Fixed
@@ -28,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Integration of `ddc::StridedDiscreteDomain` by making `IdxRangeSlice` a type alias.
 - The parameter `iter_start` has been removed from the constructor of `RestartInitialisation`.
 - Generalise `compute_coeffs_on_mapping` to work with any mapping.
-- Curvilinear coordinate change classes take a `Coord` type to specify the o-point in the constructor.
+- Curvilinear coordinate change classes take a `Coord` type to specify the O-point in the constructor.
 
 ### Deprecated
 
