### 2025-06-27 c98d22d
Call constructor
