### 2025-06-27 2faa47f
Use GPU-direct MPI (#187)
Use GPU-direct MPI. GPU-direct MPI does not need to be enabled with a
specific command. It is sufficient to pass a pointer allocated with
`cudaMalloc` or `hipMalloc` etc to MPI.
