+++ src/multipatch/data_types/types.hpp
@@ -57,6 +57,8 @@ using DField1OnPatch = DField<typename Patch::IdxRange1>;
 template <class Patch>
 using DConstField1OnPatch = DConstField<typename Patch::IdxRange1>;
 
+template <class Patch>
+using DConstFieldOnPatch_host = host_t<DConstFieldOnPatch<Patch>>; 
 
 // VECTOR FIELDS ---------------------------------------------------------------------------------
 
@@ -209,3 +211,58 @@ using ConstDeriv2_OnPatch_2D
 template <class Patch>
 using ConstDeriv12_OnPatch_2D
         = DConstField<IdxRange<ddc::Deriv<typename Patch::Dim1>, ddc::Deriv<typename Patch::Dim2>>>;
+
+/**
+ * @brief Type for MultipatchType: A field of the n-th derivatives in the direction of Patch's first
+ * logical grid defined on the index range of n and the Patch's second logical grid.
+ */
+template <class Patch>
+using Deriv1_OnPatch_2D = DField<IdxRange<ddc::Deriv<typename Patch::Dim1>, typename Patch::Grid2>>;
+
+/**
+ * @brief Type for MultipatchType: A field of the n-th derivatives in the direction of Patch's second
+ * logical grid defined on the index range of n and the Patch's first logical grid.
+ */
+template <class Patch>
+using Deriv2_OnPatch_2D = DField<IdxRange<typename Patch::Grid1, ddc::Deriv<typename Patch::Dim2>>>;
+
+/**
+ * @brief Type for MultipatchType: A field of the derivatives @f$ \partial_1^(n)\partial_2^(m) f(x) @f$
+ * defined on the index ranges of valid n and m.
+ */
+template <class Patch>
+using Deriv12_OnPatch_2D
+        = DField<IdxRange<ddc::Deriv<typename Patch::Dim1>, ddc::Deriv<typename Patch::Dim2>>>;
+
+
+/**
+ * @brief Type for MultipatchType: A field defined on host of the n-th derivatives in the direction of Patch's first
+ * logical grid defined on the index range of n and the Patch's second logical grid.
+ */
+template <class Patch>
+using Deriv1_OnPatch_2D_host = host_t<Deriv1_OnPatch_2D<Patch>>;
+
+/**
+ * @brief Type for MultipatchType: A field defined on host of the n-th derivatives in the direction of Patch's second
+ * logical grid defined on the index range of n and the Patch's first logical grid.
+ */
+template <class Patch>
+using Deriv2_OnPatch_2D_host = host_t<Deriv2_OnPatch_2D<Patch>>;
+
+/**
+ * @brief Type for MultipatchType: A field defined on host of the derivatives @f$ \partial_1^(n)\partial_2^(m) f(x) @f$
+ * defined on the index ranges of valid n and m.
+ */
+template <class Patch>
+using Deriv12_OnPatch_2D_host = host_t<Deriv12_OnPatch_2D<Patch>>;
+
+
+
+template <class Patch>
+using ConstDeriv1_OnPatch_2D_host = host_t<ConstDeriv1_OnPatch_2D<Patch>>; 
+
+template <class Patch>
+using ConstDeriv2_OnPatch_2D_host = host_t<ConstDeriv2_OnPatch_2D<Patch>>; 
+
+template <class Patch>
+using ConstDeriv12_OnPatch_2D_host = host_t<ConstDeriv12_OnPatch_2D<Patch>>; 
