### 2025-06-27 71909a9
Expose O-point from curvilinear mappings (#345)
Add a `o_point` method to curvilinear mappings to help locate the
singularity regardless of the CPU/GPU accessibility of the class. The
O-point is a coordinate and should therefore be statically typed. The
constructor arguments are updated to reflect this.
