+++ assets/.doxy/gyselalibxx/gyselalibxx/czarny__to__cartesian_8hpp_source.md
@@ -45,19 +45,16 @@ public:
 private:
     double m_epsilon;
     double m_e;
-    double m_x0;
-    double m_y0;
+    Coord<X, Y> m_coordinate_origin;
 
 public:
     explicit KOKKOS_FUNCTION CzarnyToCartesian(
             double epsilon,
             double e,
-            double x0 = 0.0,
-            double y0 = 0.0)
+            Coord<X, Y> coordinate_origin = Coord<X, Y>(0.0, 0.0))
         : m_epsilon(epsilon)
         , m_e(e)
-        , m_x0(x0)
-        , m_y0(y0)
+        , m_coordinate_origin(coordinate_origin)
     {
     }
 
@@ -81,6 +78,15 @@ public:
         return m_e;
     }
 
+    KOKKOS_INLINE_FUNCTION Coord<X, Y> o_point() const
+    {
+        const Coord<X> x(
+                (1.0 - Kokkos::sqrt(m_epsilon * m_epsilon + 1.0)) / m_epsilon
+                + ddc::get<X>(m_coordinate_origin));
+        const Coord<Y> y(m_coordinate_origin);
+        return Coord<X, Y>(x, y);
+    }
+
     KOKKOS_FUNCTION Coord<X, Y> operator()(Coord<R, Theta> const& coord) const
     {
         const double r = ddc::get<R>(coord);
@@ -88,10 +94,10 @@ public:
         const double tmp1
                 = Kokkos::sqrt(m_epsilon * (m_epsilon + 2.0 * r * Kokkos::cos(theta)) + 1.0);
 
-        const double x = (1.0 - tmp1) / m_epsilon + m_x0;
+        const double x = (1.0 - tmp1) / m_epsilon + ddc::get<X>(m_coordinate_origin);
         const double y = m_e * r * Kokkos::sin(theta)
                                  / (Kokkos::sqrt(1.0 - 0.25 * m_epsilon * m_epsilon) * (2.0 - tmp1))
-                         + m_y0;
+                         + ddc::get<Y>(m_coordinate_origin);
 
         return Coord<X, Y>(x, y);
     }
@@ -251,7 +257,7 @@ public:
 
     KOKKOS_INLINE_FUNCTION CartesianToCzarny<X, Y, R, Theta> get_inverse_mapping() const
     {
-        return CartesianToCzarny<X, Y, R, Theta>(m_epsilon, m_e, m_x0, m_y0);
+        return CartesianToCzarny<X, Y, R, Theta>(m_epsilon, m_e, m_coordinate_origin);
     }
 };
 
