+++ assets/.doxy/gyselalibxx/gyselalibxx/coord__transformation__tools_8hpp.md
@@ -53,7 +53,7 @@
 |  constexpr bool | [**has\_singular\_o\_point\_inv\_jacobian\_v**](#variable-has_singular_o_point_inv_jacobian_v)   = `/* multi line expression */`<br> |
 |  constexpr bool | [**is\_accessible\_v**](#variable-is_accessible_v)   = `/* multi line expression */`<br> |
 |  constexpr bool | [**is\_analytical\_mapping\_v**](#variable-is_analytical_mapping_v)   = `mapping\_detail::IsAnalyticalMapping&lt;Mapping&gt;::value`<br> |
-|  constexpr bool | [**is\_curvilinear\_2d\_mapping\_v**](#variable-is_curvilinear_2d_mapping_v)   = `/* multi line expression */`<br> |
+|  constexpr bool | [**is\_curvilinear\_2d\_mapping\_v**](#variable-is_curvilinear_2d_mapping_v)   = `/* multi line expression */`<br>_Indicates that a coordinate change operator is 2D with a curvilinear mapping showing an O-point._  |
 |  constexpr bool | [**is\_mapping\_v**](#variable-is_mapping_v)   = `mapping\_detail::IsMapping&lt;Mapping&gt;::value`<br> |
 
 
@@ -184,6 +184,7 @@ constexpr bool is_analytical_mapping_v;
 
 ### variable is\_curvilinear\_2d\_mapping\_v 
 
+_Indicates that a coordinate change operator is 2D with a curvilinear mapping showing an O-point._ 
 ```C++
 constexpr bool is_curvilinear_2d_mapping_v;
 ```
