### 2025-06-27 dfa97cd
Initialise the derivatives with the global spline to have two spline representations identical.
