+++ tests/multipatch/interface_derivatives/interface_derivative_matrix_test.cpp
@@ -120,6 +120,13 @@ using SplineRThetagEvaluator = ddc::SplineEvaluator2D<
         ddc::ConstantExtrapolationRule<Yg, Xg>>;
 
 
+Coord<Xg, Yg> get_global_coord(Coord<X, Y> const& local_coord)
+{
+    double const x = ddc::select<X>(local_coord);
+    double const y = ddc::select<Y>(local_coord);
+    return Coord<Xg, Yg>(x, y);
+}
+
 /// @brief Initialise the function with f(x,y) = cos(2/3*pi*x)sin(y).
 template <class Grid1, class Grid2>
 void initialise_2D_function(host_t<DField<IdxRange<Grid1, Grid2>>> function)
@@ -132,56 +139,43 @@ void initialise_2D_function(host_t<DField<IdxRange<Grid1, Grid2>>> function)
     });
 }
 
-/// @brief Initialise the function with dxf(x,y) = -2/3*pi*sin(2/3*pi*x)sin(y).
-template <class Grid1, class Grid2>
-void initialise_2D_x_derivative(
-        host_t<DField<IdxRange<ddc::Deriv<typename Grid1::continuous_dimension_type>, Grid2>>>
-                deriv_x,
-        Idx<Grid1> const& idx_x)
-{
-    double const xg = ddc::coordinate(idx_x);
-    ddc::for_each(
-            get_idx_range(deriv_x),
-            [&](Idx<ddc::Deriv<typename Grid1::continuous_dimension_type>, Grid2> idx) {
-                // Get the coordinate on the equivalent global mesh.
-                double const yg = ddc::coordinate(Idx<Grid2>(idx));
-                deriv_x(idx) = -2. / 3. * M_PI * Kokkos::sin(xg * 2. / 3. * M_PI) * Kokkos::sin(yg);
-            });
-}
 
-
-/// @brief Initialise the function with dyf(x,y) = cos(2/3*pi*x)cos(y).
+/// @brief Initialise the y-derivatives from the global spline.
 template <class Grid1, class Grid2>
 void initialise_2D_y_derivative(
         host_t<DField<IdxRange<Grid1, ddc::Deriv<typename Grid2::continuous_dimension_type>>>>
                 deriv_y,
-        Idx<Grid2> const& idx_y)
+        Idx<Grid2> const& idx_y,
+        SplineRThetagEvaluator const& evaluator_g,
+        host_t<DConstField<IdxRange<BSplinesXg, BSplinesYg>>> const& function_g_coef)
 {
-    double const yg = ddc::coordinate(idx_y);
     ddc::for_each(
             get_idx_range(deriv_y),
             [&](Idx<Grid1, ddc::Deriv<typename Grid2::continuous_dimension_type>> idx) {
                 // Get the coordinate on the equivalent global mesh.
-                double const xg = ddc::coordinate(Idx<Grid1>(idx));
-                deriv_y(idx) = Kokkos::cos(xg * 2. / 3. * M_PI) * Kokkos::cos(yg);
+                Idx<Grid1, Grid2> idx_xy(Idx<Grid1>(idx), idx_y);
+                Coord<Xg, Yg> interface_coord(get_global_coord(ddc::coordinate(idx_xy)));
+                deriv_y(idx) = evaluator_g.deriv_dim_2(interface_coord, function_g_coef);
             });
 }
 
 
-/// @brief Initialise the function with dxyf(x,y) = -2/3*pi*sin(2/3*pi*x)cos(y).
+/// @brief Initialise the cross derivatives from the global spline.
 template <class Grid1, class Grid2>
 void initialise_2D_xy_derivative(
         host_t<DField<IdxRange<
                 ddc::Deriv<typename Grid1::continuous_dimension_type>,
[…]
