### 2025-06-30 d32cc77
Clang formatting
