+++ src/data_types/static_tensors.hpp
@@ -291,9 +291,10 @@ KOKKOS_INLINE_FUNCTION constexpr double get(
 namespace detail {
 
 template <class ElementType, class ValidIndexSetRow, class ValidIndexSetCol>
-inline constexpr bool enable_tensor_type<IdentityTensor<ElementType, ValidIndexSetRow, ValidIndexSetCol>> = true;
+inline constexpr bool
+        enable_tensor_type<IdentityTensor<ElementType, ValidIndexSetRow, ValidIndexSetCol>> = true;
 
 template <class ElementType, class ValidIndexSet>
 inline constexpr bool enable_tensor_type<LeviCivitaTensor<ElementType, ValidIndexSet>> = true;
 
-}
+} // namespace detail
