+++ src/data_types/vector_field.hpp
@@ -1,12 +1,9 @@
 // SPDX-License-Identifier: MIT
-
 #pragma once
-#include "ddc_alias_inline_functions.hpp"
+#include <ddc/ddc.hpp>
+
 #include "ddc_aliases.hpp"
-#include "ddc_helper.hpp"
-#include "vector_field_mem.hpp"
-#include "vector_index_tools.hpp"
-
+#include "vector_field_common.hpp"
 
 template <
         class ElementType,
@@ -14,511 +11,30 @@ template <
         class VectorIndexSetType,
         class MemorySpace = Kokkos::DefaultExecutionSpace::memory_space,
         class LayoutStridedPolicy = Kokkos::layout_right>
-class VectorField;
-
-template <
-        class ElementType,
-        class IdxRangeType,
-        class VectorIndexSetType,
-        class MemorySpace,
-        class LayoutStridedPolicy>
-inline constexpr bool enable_vector_field<VectorField<
-        ElementType,
-        IdxRangeType,
-        VectorIndexSetType,
-        MemorySpace,
-        LayoutStridedPolicy>> = true;
-
-template <
-        class ElementType,
-        class IdxRangeType,
-        class VectorIndexSetType,
-        class MemorySpace,
-        class LayoutStridedPolicy>
-inline constexpr bool enable_data_access_methods<VectorField<
-        ElementType,
-        IdxRangeType,
-        VectorIndexSetType,
-        MemorySpace,
-        LayoutStridedPolicy>> = true;
-
-template <
-        class ElementType,
-        class IdxRangeType,
-        class VectorIndexSetType,
-        class MemorySpace,
-        class LayoutStridedPolicy>
-inline constexpr bool enable_borrowed_vector_field<VectorField<
-        ElementType,
-        IdxRangeType,
-        VectorIndexSetType,
-        MemorySpace,
-        LayoutStridedPolicy>> = true;
-
-
-/**
- * @brief A class which holds multiple (scalar) fields in order to represent a vector field.
- *
- * @tparam ElementType The data type of a scalar element of the vector field.
- * @tparam IdxRangeType
- * @tparam VectorIndexSetType A VectorIndexSet describing the dimensions described by the scalar elements of a vector field element.
- * @tparam MemorySpace The memory space (CPU/GPU).
- * @tparam LayoutStridedPolicy The memory layout. See DDC.
- */
-template <
-        class ElementType,
-        class IdxRangeType,
-        class VectorIndexSetType,
-        class MemorySpace,
-        class LayoutStridedPolicy>
-class VectorField
-    : public VectorFieldCommon<
-              Field<ElementType, IdxRangeType, MemorySpace, LayoutStridedPolicy>,
-              VectorIndexSetType>
-{
-public:
-    /**
-     * @brief Type describing the object which can be extracted from this VectorField using the get<> function.
-     */
-    using field_type = Field<ElementType, IdxRangeType, MemorySpace, LayoutStridedPolicy>;
-
-private:
-    using base_type = VectorFieldCommon<field_type, VectorIndexSetType>;
-
-public:
-    /// The type of an element in one of the Fields comprising the VectorField
-    using element_type = typename base_type::element_type;
-
-    using typename base_type::NDTypeTag;
-
-    using typename base_type::chunk_span_type;
-    using typename base_type::chunk_view_type;
-
-public:
-    /**
-     * @brief A type which can hold a modifiable reference to a VectorFieldMem.
-     *
-     * A type which can hold a reference to a VectorFieldMem. If this object is modifiable then
-     * so is the span type.
-     * This is a DDC keyword used to make this class interchangeable with Field.
-     */
-    using span_type = VectorField<
-            ElementType,
[…]
