+++ src/data_types/vector_field_mem.hpp
@@ -1,23 +1,84 @@
 // SPDX-License-Identifier: MIT
-
 #pragma once
 #include <ddc/ddc.hpp>
 
-#include "ddc_alias_inline_functions.hpp"
 #include "ddc_aliases.hpp"
-#include "ddc_helper.hpp"
+#include "tensor.hpp"
 #include "vector_field_common.hpp"
-#include "vector_index_tools.hpp"
 
-/**
- * @brief Pre-declaration of VectorFieldMem.
- */
 template <
         class ElementType,
         class IdxRangeType,
         class VectorIndexSetType,
         class MemSpace = Kokkos::DefaultExecutionSpace::memory_space>
-class VectorFieldMem;
+class VectorFieldMem
+    : private FieldMem<
+              ElementType,
+              ddc::cartesian_prod_t<
+                      IdxRange<detail::VecIdxSetWrapper<VectorIndexSetType>>,
+                      IdxRangeType>,
+              MemSpace>
+{
+    static_assert(is_vector_index_set_v<VectorIndexSetType>);
+
+private:
+    using base_type = FieldMem<
+            ElementType,
+            ddc::cartesian_prod_t<
+                    IdxRange<detail::VecIdxSetWrapper<VectorIndexSetType>>,
+                    IdxRangeType>,
+            MemSpace>;
+
+public:
+    using idx_range_type = IdxRangeType;
+    using Allocator = ddc::KokkosAllocator<ElementType, MemSpace>;
+    /// @brief The type of the elements in the fields.
+    using element_type = Tensor<std::remove_const_t<ElementType>, VectorIndexSetType>;
+    using base_type::layout_type;
+
+private:
+    using internal_idx_range_type = typename base_type::discrete_domain_type;
+
+public:
+    /// Construct a labeled VectorFieldMem on a domain with uninitialized values
+    explicit VectorFieldMem(
+            std::string const& label,
+            idx_range_type const& domain,
+            Allocator allocator = Allocator())
+        : base_type(
+                label,
+                internal_idx_range_type(detail::make_idx_range<VectorIndexSetType>(), domain),
+                allocator)
+    {
+    }
+
+    /// Construct a VectorFieldMem on a domain with uninitialized values
+    explicit VectorFieldMem(idx_range_type const& domain, Allocator allocator = Allocator())
+        : VectorFieldMem("no-label", domain, std::move(allocator))
+    {
+    }
+
+    idx_range_type idx_range() const
+    {
+        return idx_range_type(this->domain());
+    }
+
+    /** Element access using a multi-dimensional Idx
+     * @param delems discrete coordinates
+     * @return copy of this element
+     */
+    template <class... ODDims, typename T, T... ints>
+    element_type const operator()(Idx<ODDims...> const& delems, std::integer_sequence<T, ints...>)
+            const noexcept
+    {
+        return element_type((base_type::m_values[ints](delems))...);
+    }
+
+
+
+    using base_type::span_view;
+    using base_type::span_cview;
+};
 
 
 template <class ElementType, class IdxRangeType, class DimSeq, class MemSpace>
@@ -32,313 +93,14 @@ template <class ElementType, class IdxRangeType, class DimSeq, class Allocator>
 inline constexpr bool
         enable_mem_type<VectorFieldMem<ElementType, IdxRangeType, DimSeq, Allocator>> = true;
 
-/**
- * @brief Pre-declaration of VectorField.
- */
-template <class, class, class, class, class>
-class VectorField;
-
-/**
- * @brief A class which describes the storage for a vector field.
- *
- * A class which describes the storage for a vector field. In other words a class which maps a position
- * on an index range to a vector (x,y,z,...). This is done by storing the values at the positions in
- * individual FieldMems.
- *
- * @tparam ElementType The data type of a scalar element of the vector field.
[…]

[…]
