### 2025-06-30 792f198
Add methods to get the fields that are found in polar_splines
