### 2025-06-30 6707627
Add test on the spline representations + cleaning a bit.
