+++ src/multipatch/interface_derivatives/interface_derivative_matrix.hpp
@@ -37,31 +37,41 @@
  *      - Deal with periodic case here or in another operator?
  */
 
+ /**
+  * @brief 
+  */
 template <
         class Connectivity,
         class Grid1D,
         template <typename P>
         typename ValuesOnPatch,
-        bool PERIODIC,
         ddc::BoundCond LowerBound = ddc::BoundCond::HERMITE,
         ddc::BoundCond UpperBound = ddc::BoundCond::HERMITE,
         class ExecSpace = Kokkos::DefaultHostExecutionSpace,
         class... Patches>
 class InterfaceDerivativeMatrix
 {
-    using all_interface_collection = typename Connectivity::interface_collection; // TypeSeq
+    using all_interface_collection = typename Connectivity::interface_collection;
     using interface_collection =
             typename Connectivity::get_all_interfaces_along_direction_t<Grid1D>;
 
-    using all_patches = typename Connectivity::all_patches; // TypeSeq
+    using all_patches = typename Connectivity::all_patches;
 
     static constexpr std::size_t number_of_interfaces = ddc::type_seq_size_v<interface_collection>;
+
+    static_assert(
+            ((LowerBound == ddc::BoundCond::PERIODIC) && (UpperBound == ddc::BoundCond::PERIODIC))
+                    || ((LowerBound != ddc::BoundCond::PERIODIC)
+                        && (UpperBound != ddc::BoundCond::PERIODIC)),
+            "If one boundary is periodic, the other boundary should be too.");
+
+    static constexpr bool is_periodic = (LowerBound == ddc::BoundCond::PERIODIC);
 
     // TODO: remove Interfaces with OutsideEdge
     // Remove all the interfaces with an OutsideEdge.
     // Rely on the get_all_interfaces_along_direction_t that orders the interfaces.
     using outer_interface_collection = std::conditional_t<
-            PERIODIC,
+            is_periodic,
             ddc::detail::TypeSeq<>,
             ddc::detail::TypeSeq<
                     ddc::type_seq_element_t<0, interface_collection>,
@@ -76,13 +86,20 @@ class InterfaceDerivativeMatrix
             inner_interface_collection>;
 
     using FirstPatch = std::conditional_t<
-            PERIODIC,
+            is_periodic,
             find_patch_t<Grid1D, all_patches>,
             typename FirstInterface::Edge2::associated_patch>;
     using Grid1DSeq = collect_grids_on_dim_t<
             find_patch_t<Grid1D, all_patches>,
             Grid1D,
             interface_collection>;
+
+    static constexpr std::size_t n_inner_interfaces
+            = ddc::type_seq_size_v<inner_interface_collection>;
+
+    static_assert(
+            ddc::type_seq_size_v<Grid1DSeq> == n_inner_interfaces + !is_periodic,
+            "The number of 1D grids and inner interfaces should fit.");
 
 
     template <typename Patch>
@@ -128,48 +145,11 @@ class InterfaceDerivativeMatrix
             typename ExecSpace::memory_space>;
 
     // HELPFUL ALIASES ===========================================================================
-    template <typename Grid>
-    using get_patch_on_grid = find_patch_t<Grid, all_patches>;
-
-
-    // Get the grid tag from a tag templated on the grid.
-    template <typename T>
-    struct get_grid;
-
-    template <template <typename P> typename T, typename Grid>
-    struct get_grid<T<Grid>>
-    {
-        using type = Grid;
-    };
-
-    template <typename T>
-    using get_grid_t = typename get_grid<T>::type;
-
-    // Get the Patch defined with the given IdxRange tag.
-    template <typename IdxRange1D>
-    using get_patch_of_idx_range_1d = get_patch_on_grid<get_grid_t<IdxRange1D>>;
-
[…]

[…]
