### 2025-06-30 76b2eb5
Remove unnecessary std::move
