### 2025-07-01 986002e
Update CHANGELOG.md
Co-authored-by: Emily Bourne <emily.bourne@epfl.ch>
