### 2025-07-01 03da68f
Remove unnecessary std::move (#353)
Remove unnecessary std::move

closes #48 

---------

Co-authored-by: Emily Bourne <emily.bourne@epfl.ch>
