### 2025-07-01 7766a71
Add a cast function to transit incrementally
