### 2025-07-01 d74b15e
Add field interface to evaluator
