### 2025-07-01 89b8974
Use DFieldMem in polarpoissonlikesolver
