+++ .github/workflows/cpu_tests.yml
@@ -106,7 +106,7 @@ jobs:
       fail-fast: false
     container:
       image: ${{ needs.docker_update.outputs.image_tag }}
-      options: --user root
+      options: --user ci
     name: CPU Test (${{ matrix.toolchain }}, ${{ matrix.compiler.name }})
     env:
       CC: ${{ matrix.compiler.c_compiler }}
