+++ .github/workflows/static_analyses.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ubuntu-latest
     container:
       image: ghcr.io/gyselax/gyselalibxx_env
-      options: --user 1001
+      options: --user ci
     steps:
       - uses: actions/checkout@v4
       - shell: bash
@@ -56,7 +56,7 @@ jobs:
     runs-on: ubuntu-latest
     container:
       image: ghcr.io/gyselax/gyselalibxx_env
-      options: --user 1001
+      options: --user ci
     steps:
       - uses: actions/checkout@v4
         with:
