### 2025-07-01 1f0b51f
Clean files, add some documentation. Remove modification MultipatchConnectivity. Remove test file.
