+++ src/multipatch/interface_derivatives/interface_derivative_matrix.hpp
@@ -21,13 +21,6 @@
  *      - Constructor
  *          - Take a list of "parallel" interfaces.
  *          - For each "conforming" lines, get the coefficients a and b.
- *          - Fill in the matrix.
- *      - operator()
- *          - get the coefficient c.
- *          - Fill in the vector.
- *          - Inverse the matrix system. (See how it is done in the SplineBuilder.)
- *          - update derivatives. How to assert values?
- *
  *  - Remarks:
  *      - Should we restrain to the conforming case first?
  *      - And then the non-conforming case in another operator?
@@ -37,8 +30,27 @@
  *      - Deal with periodic case here or in another operator?
  */
 
- /**
-  * @brief 
+
+/**
+  * @brief Class to compute the interface derivatives along a given direction 
+  * on all the interfaces.  
+  * 
+  * During the instantiation of the class, the matrix (I-M) is computed are stored in the class. 
+  * When we call the operator .solve(), 
+  *     - InterfaceDerivativeMatrix computes the vector C from the given values. 
+  *     - It inverses the matrix system: (I-M)S = C. 
+  *     - The interface derivatives are stored in the vector S. 
+  *     - It updates the given derivatives field with the values of the vector S. 
+  * 
+  * 
+  * @tparam Connectivity A MultipatchConnectivity class describing all the patch connections.
+  * @tparam Grid1D A given direction.
+  * @tparam ValuesOnPatch Type of the const Field  where the function values are defined. 
+  *         It is templated on the patch. 
+  * @tparam LowerBound Lower/left boundary condition of the first local spline along the given direction. 
+  * @tparam UpperBound Upper/right boundary condition of the last local spline along the given direction. 
+  * @tparam ExecSpace Execution space.
+  * @tparam Patches List of patches containing all the involved patches in the given direction. 
   */
 template <
         class Connectivity,
@@ -51,10 +63,14 @@ template <
         class... Patches>
 class InterfaceDerivativeMatrix
 {
+    // All the interfaces given as input to the MultipatchConnectivity class.
+    // We expect the parameters defined on these interfaces.
     using all_interface_collection = typename Connectivity::interface_collection;
+    // All the sorted interfaces with the correct orientation.
     using interface_collection =
             typename Connectivity::get_all_interfaces_along_direction_t<Grid1D>;
 
+    // All the paches of the geometry.
     using all_patches = typename Connectivity::all_patches;
 
     static constexpr std::size_t number_of_interfaces = ddc::type_seq_size_v<interface_collection>;
@@ -77,31 +93,49 @@ class InterfaceDerivativeMatrix
                     ddc::type_seq_element_t<0, interface_collection>,
                     ddc::type_seq_element_t<number_of_interfaces - 1, interface_collection>>>;
 
+    // All the interfaces in the Grid1D direction without the interfaces with OutsideEdge.
     using inner_interface_collection
             = ddc::type_seq_remove_t<interface_collection, outer_interface_collection>;
+
+    // Number of inner interfaces. It fixes the number of element in the matrix and vectors.
+    static constexpr std::size_t n_inner_interfaces
+            = ddc::type_seq_size_v<inner_interface_collection>;
 
     using FirstInterface = ddc::type_seq_element_t<0, inner_interface_collection>;
     using LastInterface = ddc::type_seq_element_t<
             ddc::type_seq_size_v<inner_interface_collection> - 1,
             inner_interface_collection>;
 
+    // First patch of the sorted interfaces.
     using FirstPatch = std::conditional_t<
             is_periodic,
             find_patch_t<Grid1D, all_patches>,
[…]
