+++ tests/multipatch/interface_derivatives/test_with_ginkgo.cpp
@@ -1,160 +1,0 @@
-#include <ddc/ddc.hpp>
-// #include <ddc/kernels/splines.hpp>
-
-#include <ginkgo/ginkgo.hpp>
-
-
-// Fill in the matrix.
-void generate_stencil_matrix(gko::matrix::Dense<>* matrix)
-{
-    const auto nb_points = matrix->get_size()[0];
-    const double coefs[] = {-1, 2, -1};
-
-    for (int i = 0; i < nb_points; ++i) {
-        for (auto dofs : {-1, 0, 1}) {
-            if (0 <= i + dofs && i + dofs < nb_points) {
-                matrix->at(i, i + dofs) = coefs[dofs + 1];
-            }
-        }
-    }
-}
-
-
-void generate_stencil_matrix(gko::matrix::Csr<>* matrix)
-{
-    Kokkos::View<
-            int*,
-            Kokkos::DefaultHostExecutionSpace> // , Kokkos::MemoryTraits<Kokkos::Unmanaged>>
-            rows(matrix->get_row_ptrs(), matrix->get_size()[0]);
-    Kokkos::View<
-            int*,
-            Kokkos::DefaultHostExecutionSpace> // , Kokkos::MemoryTraits<Kokkos::Unmanaged>>
-            cols(matrix->get_col_idxs(), matrix->get_num_stored_elements());
-    Kokkos::View<
-            double*,
-            Kokkos::DefaultHostExecutionSpace> // , Kokkos::MemoryTraits<Kokkos::Unmanaged>>
-            vals(matrix->get_values(), matrix->get_num_stored_elements());
-
-
-    const auto nb_points = matrix->get_size()[0];
-    const double coefs[] = {-1, 2, -1};
-
-    for (int i = 0; i < nb_points; ++i) {
-        for (auto dofs : {-1, 0, 1}) {
-            if (0 <= i + dofs && i + dofs < nb_points) {
-                vals(3*i + dofs) = coefs[dofs + 1];
-                cols(3*i + dofs) = i + dofs;
-                rows(i) = i-1;
-            }
-        }
-    }
-}
-
-
-
-int main(int argc, char* argv[])
-{
-    const unsigned int nb_points = 10;
-
-    using mtx_dense = gko::matrix::Dense<double>;
-    using mtx_csr = gko::matrix::Csr<double>;
-    // using mtx = gko::matrix::Coo<double>;
-
-    const auto exec = gko::ReferenceExecutor::create();
-
-    // DENSE MATRIX -----------------------------------------------------------
-    // Create matrix nb_points x nb_points
-    std::shared_ptr<mtx_dense> matrix; 
-    matrix = gko::share(mtx_dense::create(exec, gko::dim<2>(nb_points)));
-    // std::unique_ptr<mtx_dense> matrix; 
-    // matrix = std::make_unique(mtx_dense::create(exec, gko::dim<2>(nb_points)));
-    // auto matrix = gko::share(mtx_dense::create(exec, gko::dim<2>(nb_points)));
-    // auto vec = gko::share(mtx_dense::create(exec, gko::dim<2>(nb_points, 1)));
-
-    auto vec = mtx_dense::create(exec, gko::dim<2>(nb_points, 1));
-    auto sol = mtx_dense::create(exec, gko::dim<2>(nb_points, 1));
-
-    // Use smart pointer.
-    generate_stencil_matrix(gko::lend(matrix));
-    // generate_stencil_matrix(matrix);
-
-    for (int i = 0; i < nb_points; ++i) {
-        for (int j = 0; j < nb_points; ++j) {
-            std::cout << matrix->get_values()[i * (nb_points) + j] << "   ";
-        }
[…]

[…]
