### 2025-07-01 7ba15f8
typo detected in README
