### 2025-07-01 4b842e0
Add a page to the docs about how to debug
