### 2025-07-01 5a1540f
Complete sentence
