### 2025-07-01 49f9ba4
Use correct control points
