### 2025-07-01 cd8a72e
README: add documentation in the README.md + add some in the source file.
