+++ src/multipatch/interface_derivatives/README.md
@@ -21,6 +21,8 @@ the boundary cells.
 - [Relation between derivatives on the boundaries of two connected patches](#relation-between-derivatives-on-the-boundaries-of-two-connected-patches): It documents the `SingleInterfaceDerivativesCalculator` operator.
   - [How to use the `SingleInterfaceDerivativesCalculator` operator?](#how-to-use-the-singleinterfacederivativescalculator-operator): It documents how to use the operator in the code.
   - [Formulae](#formulae): It details the formulae applies in the operator.
+- [Relation between the interface derivatives along one direction](#relation-between-the-interface-derivatives-along-one-direction): It documents the `InterfaceDerivativeMatrix` operator.
+  - [How to use the `InterfaceDerivativeMatrix` operator?](#how-to-use-the-interfacederivativematrix-operator): It documents how to use the operator in the code.
 
 ## Relation between derivatives on the boundaries of two connected patches
 
@@ -469,6 +471,235 @@ with the weights given by,
 and $`a^I_{1,1} = -\frac{1}{2} \frac{\Delta x^-}{\Delta x^+ +\Delta x^-}`$
 and $`b^I_{1,1} =  -\frac{1}{2} \frac{\Delta x^+}{\Delta x^+ +\Delta x^-}`$.
 
+## Relation between the interface derivatives along one direction
+
+We consider now a set of patch connected via different interfaces. 
+We want to compute the derivatives at all the interfaces. 
+
+To compute all the interface derivatives along one direction, we collect all the relations between 
+three consecutive derivatives for each interface. 
+(For a given interface, the relation is defined in the previous section,
+see [Relation between derivatives on the boundaries of two connected patches](#relation-between-derivatives-on-the-boundaries-of-two-connected-patches)). 
+
+All these relations can be stored in a matrix system as follows, 
+
+```math
+\begin{bmatrix}
+    s'(\mathcal{X}_1) \\
+    \vdots \\
+    s'(\mathcal{X}_I) \\
+    \vdots \\
+    s'(\mathcal{X}_{N_I-1}) \\
+\end{bmatrix}
+= 
+\begin{bmatrix}
+    1 & -a^1 & \dots & 0\\
+    -b^2 & 1 & -a^2 & \vdots \\
+    \vdots & \ddots & \ddots & -a^{N_I -2}\\
+    0 & \dots & -b^{N_I -1}& 1 \\
+\end{bmatrix}^{-1}
+\begin{bmatrix}
+    c^1 + b^1 s'(\mathcal{X}_0) \\
+    \vdots \\
+    c^I \\
+    \vdots \\
+    c^{N_I-1} + a^{N_I-1} s'(\mathcal{X}_{N_I}) \\
+\end{bmatrix}, 
+```
+
+with 
+* $`a^I, b^I \text{ and } c^I`$ the coefficients computed with `SingleInterfaceDerivativeCalculator` 
+for a given interface *I* (and given number of cells on the right and the left patches that we do not 
+precise here to lighten the notation). 
+* $`\{\mathcal{X}_I\}_I`$ a set of parallel interfaces. 
+* $`s'(\mathcal{X})`$ the derivative at the *I*th interface.
+
+We simply refer it as 
+```math
+    S = (\mathbb{I}-M)^{-1}C.
+```
+
+> **Remark:** This matrix is given in the case where the equivalent global spline passing through all 
+> the patches in the given direction use **Hermite boundary conditions**. 
+> 
+> If the equivalent global spline uses **additional interpolation points as closure condition** 
[…]
