### 2025-07-01 0fcd4c5
TypSeq -> TypeSeq to avoid to call it TypoSeq
