+++ tests/multipatch/interface_derivatives/single_interface_derivatives_calculator_test.cpp
@@ -15,10 +15,10 @@
 #include "ddc_helper.hpp"
 #include "global_2d_onion_shape_non_uniform.hpp"
 #include "interface.hpp"
+#include "interface_derivatives_test_utils.hpp"
 #include "mesh_builder.hpp"
 #include "non_uniform_interpolation_points.hpp"
 #include "single_interface_derivatives_calculator.hpp"
-#include "interface_derivatives_test_utils.hpp"
 
 
 
