+++ tests/multipatch/interface_derivatives/interface_derivatives_test_utils.hpp
@@ -1,6 +1,6 @@
 
 // SPDX-License-Identifier: MIT
-#pragma once 
+#pragma once
 #include <ddc/ddc.hpp>
 #include <ddc/kernels/splines.hpp>
 
