+++ src/multipatch/interface_derivatives/single_interface_derivatives_calculator.hpp
@@ -118,7 +118,7 @@ class SingleInterfaceDerivativesCalculator
     using IdxRangeBreakPt2 = IdxRange<GridBreakPt2>;
 
 public:
-    /// @brief Interface between the two involved patches. 
+    /// @brief Interface between the two involved patches.
     using associated_interface = InterfaceType;
 
 private:
