+++ src/multipatch/interface_derivatives/README.md
@@ -19,7 +19,7 @@ the boundary cells.
 ## Contents
 
 - [Relation between derivatives on the boundaries of two connected patches](#relation-between-derivatives-on-the-boundaries-of-two-connected-patches): It documents the `SingleInterfaceDerivativesCalculator` operator.
-  - [How to use the `SingleInterfaceDerivativesCalculator` operator?](#how-to-use-the-singleinterfacederivativescalculator-operator): It documents how to use the operator in the code.
+  - [How to use the `SingleInterfaceDerivativesCalculator` operator?](#how-to-use-the-singleinterfacederivativescalculator-operator): It documents how to use the operator in the code. 
   - [Formulae](#formulae): It details the formulae applies in the operator.
 - [Relation between the interface derivatives along one direction](#relation-between-the-interface-derivatives-along-one-direction): It documents the `InterfaceDerivativeMatrix` operator.
   - [How to use the `InterfaceDerivativeMatrix` operator?](#how-to-use-the-interfacederivativematrix-operator): It documents how to use the operator in the code.
@@ -184,15 +184,15 @@ of the interpolating function values. In the code, we store the weights in front
 as they depend only on the grids. For the same reason, we store $`\gamma_i`$ as a vector,
 
 ```math
-{\bf \Gamma_i} 
-    = \left[({\bf \Gamma_i})_0, ({\bf \Gamma_i})_1, ({\bf \Gamma_i})_2\right] 
-    = \frac{3}{2} \frac{1}{\Delta x^+_{i} +\Delta x^-_{i}} 
+{\bf \Gamma_i}
+    = \left[({\bf \Gamma_i})_0, ({\bf \Gamma_i})_1, ({\bf \Gamma_i})_2\right]
+    = \frac{3}{2} \frac{1}{\Delta x^+_{i} +\Delta x^-_{i}}
         \left[
-            - \frac{\Delta x^+_{i}}{\Delta x^-_{i}}, 
+            - \frac{\Delta x^+_{i}}{\Delta x^-_{i}},
             \left(
                 \frac{\Delta x^+_{i}}{\Delta x^-_{i}}
                 - \frac{\Delta x^-_{i}}{\Delta x^+_{i}}
-            \right), 
+            \right),
             \frac{\Delta x^-_{i}}{\Delta x^+_{i}}
         \right].
 ```
@@ -221,8 +221,8 @@ We start the recursion on patch 2. We initialise it with,
 
 ```math
 \begin{aligned}
-    & \omega_{k, 1,1}^i = \gamma_{i,k}, 
-    && \omega_{k, 1,2}^i = \frac{1}{1 - \alpha_i  \beta_{i+1}} \omega_{k, 1,1}^i 
+    & \omega_{k, 1,1}^i = \gamma_{i,k},
+    && \omega_{k, 1,2}^i = \frac{1}{1 - \alpha_i  \beta_{i+1}} \omega_{k, 1,1}^i
                         + \frac{\alpha_i}{1 - \alpha_i  \beta_{i+1}} \gamma_{i+1,k},
 \end{aligned}
 ```
@@ -322,7 +322,7 @@ and for $`m = 2, ..., N^L-1`$,
 
 ```math
 \begin{aligned}
-    & \omega_{k, m+1,N^R}^i = 
+    & \omega_{k, m+1,N^R}^i =
     \frac{1}{1 - \alpha_{i-m} \frac{b^{i}_{m,N^R}}{b^{i}_{m-1,N^R}}}
             \left[
                 \omega_{k, m,N^R}^i
@@ -473,15 +473,15 @@ and $`b^I_{1,1} =  -\frac{1}{2} \frac{\Delta x^+}{\Delta x^+ +\Delta x^-}`$.
 
 ## Relation between the interface derivatives along one direction
 
[…]
