### 2025-07-02 f087601
Introduce a tag to represent gyro motion
