### 2025-07-02 bd5e9af
remove unused variables
