### 2025-07-02 f83a2ae
Correct available control points
