### 2025-07-02 9d309be
Lie poisson improvements (#352)
Improve the Lie-Poisson bracket by adding an overload of the
`operator()` which allows the `dg` argument to be a 2D tensor instead of
a vector. Fixes #327. Additionally add an assertion in
`LeviCivitaTensor` to ensure that an error is raised if the tensor is
evaluated at a singular point (i.e. a point where the Jacobian is 0).
Fixes #348.
Add missing `enable_tensor_type` for `IdentityTensor` so
`is_tensor_v<MyIdentityTensor>` returns `true`.
