### 2025-07-02 68eb4e1
Add a PDI_expose_vector_field method
