+++ src/io/pdi_helper.hpp
@@ -1,7 +1,11 @@
 // SPDX-License-Identifier: MIT
 #pragma once
+#include <ddc/ddc.hpp>
+#include <ddc/pdi.hpp>
 
 #include <pdi.h>
+
+#include "vector_field.hpp"
 
 namespace detail {
 
@@ -42,6 +46,12 @@ auto get_vector_tuple(TupleType input_args, std::integer_sequence<size_t, I...>)
     return std::tie(std::get<I * 2 + 1>(input_args)...);
 }
 
+template <class TupleType, class StringTupleType, size_t... I>
+void expose_arrays(TupleType out_obj, StringTupleType names, std::integer_sequence<size_t, I...>)
+{
+    ((ddc::expose_to_pdi(std::get<I>(names).c_str(), std::get<I>(out_obj))), ...);
+}
+
 } // namespace detail
 
 /**
@@ -80,4 +90,18 @@ void PDI_expose_idx_range(IdxRange<Grids...> index_range, std::string name)
     }
     PDI_expose((name + "_starts").c_str(), starts_s_arr.data(), PDI_OUT);
     PDI_expose((name + "_extents").c_str(), extents_s_arr.data(), PDI_OUT);
+}
+
+template <class ElementType, class IdxRangeType, class... IndexTag, class... Args>
+void PDI_expose_vector_field(
+        std::string name_stem,
+        VectorConstField<ElementType, IdxRangeType, VectorIndexSet<IndexTag...>, Kokkos::HostSpace>
+                out_vector,
+        Args const&... name_suffixes)
+{
+    static_assert(sizeof...(Args) == sizeof...(IndexTag));
+    static_assert((std::is_convertible_v<Args, std::string> && ...));
+    std::tuple names = {(name_stem + name_suffixes)...};
+    std::tuple out_obj = std::make_tuple(ddcHelper::get<IndexTag>(out_vector)...);
+    detail::expose_arrays(out_obj, names, std::make_index_sequence<sizeof...(Args)> {});
 }
