### 2025-07-02 db3bd7f
Test how to use DerivField with a SplineBuilder2D
