+++ tests/data_types/use_deriv_field_with_builder.cpp
@@ -1,0 +1,401 @@
+// SPDX-License-Identifier: MIT
+#include <ddc/ddc.hpp>
+#include <ddc/kernels/splines/deriv.hpp>
+
+#include <gtest/gtest.h>
+
+#include "ddc_alias_inline_functions.hpp"
+#include "ddc_aliases.hpp"
+#include "ddc_helper.hpp"
+#include "derivative_field.hpp"
+#include "derivative_field_mem.hpp"
+#include "grid_builder.hpp"
+#include "view.hpp"
+
+namespace {
+struct X
+{
+    static bool constexpr PERIODIC = false;
+    static bool constexpr IS_COVARIANT = true;
+    static bool constexpr IS_CONTRAVARIANT = true;
+    using Dual = X;
+};
+
+struct Y
+{
+    static bool constexpr PERIODIC = false;
+    static bool constexpr IS_COVARIANT = true;
+    static bool constexpr IS_CONTRAVARIANT = true;
+    using Dual = Y;
+};
+
+struct GridX : UniformGridBase<X>
+{
+};
+
+struct BSplinesX : ddc::UniformBSplines<X, 3>
+{
+};
+
+struct GridY : UniformGridBase<Y>
+{
+};
+
+struct BSplinesY : ddc::UniformBSplines<Y, 3>
+{
+};
+
+using dX = ddc::Deriv<GridX>;
+using DerivX = ddc::Deriv<X>;
+using IdxX = Idx<GridX>;
+using IdxStepX = IdxStep<GridX>;
+using IdxRangeX = IdxRange<GridX>;
+
+using dY = ddc::Deriv<GridY>;
+using DerivY = ddc::Deriv<Y>;
+using IdxY = Idx<GridY>;
+using IdxStepY = IdxStep<GridY>;
+using IdxRangeY = IdxRange<GridY>;
+
+
+using IdxXY = Idx<GridX, GridY>;
+using IdxStepXY = IdxStep<GridX, GridY>;
+using IdxRangeXY = IdxRange<GridX, GridY>;
+
+using IdxdXX = Idx<dX, GridX>;
+using IdxdYX = Idx<dY, GridX>;
+using IdxdYY = Idx<dY, GridY>;
+using IdxdXY = Idx<dX, GridY>;
+
+using IdxdXdYXY = Idx<dX, dY, GridX, GridY>;
+using IdxRangedXdYXY = IdxRange<dX, dY, GridX, GridY>;
+
+using SplineInterpPointsX = ddc::
+        KnotsAsInterpolationPoints<BSplinesX, ddc::BoundCond::HERMITE, ddc::BoundCond::HERMITE>;
+using SplineInterpPointsY = ddc::
+        KnotsAsInterpolationPoints<BSplinesY, ddc::BoundCond::HERMITE, ddc::BoundCond::HERMITE>;
+
+using DerivFieldMemdXdYXY = DerivFieldMem<double, IdxRangedXdYXY, 2>;
+using DerivFielddXdYXY = DerivField<double, IdxRangedXdYXY>;
+
+} // namespace
+
+
+
+TEST(DerivFieldTest, SplineBuilderUse)
+{
+    // Define meshes -----------------------------------------------------------------------------
+    const Coord<X> xmin(0);
+    const Coord<X> xmax(1);
+    const IdxStepX x_ncells(10);
+
+    const Coord<Y> ymin(0);
+    const Coord<Y> ymax(1);
+    const IdxStepY y_ncells(10);
+
+    ddc::init_discrete_space<BSplinesX>(xmin, xmax, x_ncells);
+    ddc::init_discrete_space<BSplinesY>(ymin, ymax, y_ncells);
+
[…]
