### 2025-07-02 9e05791
Improve type name
