### 2025-07-02 bbed90b
Add a PDI output for VectorFields (#359)
Add a PDI output for VectorFields. Fixes #356
