+++ assets/.doxy/gyselalibxx/gyselalibxx/pdi__helper_8hpp.md
@@ -10,7 +10,10 @@
 
 
 
+* `#include <ddc/ddc.hpp>`
+* `#include <ddc/pdi.hpp>`
 * `#include <pdi.h>`
+* `#include "vector_field.hpp"`
 
 
 
@@ -53,6 +56,7 @@
 | Type | Name |
 | ---: | :--- |
 |  void | [**PDI\_expose\_idx\_range**](#function-pdi_expose_idx_range) (IdxRange&lt; Grids... &gt; index\_range, std::string name) <br> |
+|  void | [**PDI\_expose\_vector\_field**](#function-pdi_expose_vector_field) (std::string const & name\_stem, [**VectorConstField**](classVectorField.md)&lt; ElementType, IdxRangeType, VectorIndexSet&lt; IndexTag... &gt;, Kokkos::HostSpace &gt; out\_vector, StringType const &... name\_suffixes) <br>_A helper function to expose a vector field to PDI ready for output to a file._  |
 |  void | [**PDI\_get\_arrays**](#function-pdi_get_arrays) (std::string const & event\_name, std::string const & name, std::vector&lt; [**T**](structT.md) &gt; & out\_vector, Args &... input\_args) <br> |
 
 
@@ -99,6 +103,38 @@ void PDI_expose_idx_range (
 
 
 
+
+<hr>
+
+
+
+### function PDI\_expose\_vector\_field 
+
+_A helper function to expose a vector field to PDI ready for output to a file._ 
+```C++
+template<class ElementType, class IdxRangeType, class... IndexTag, class... StringType>
+void PDI_expose_vector_field (
+    std::string const & name_stem,
+    VectorConstField < ElementType, IdxRangeType, VectorIndexSet< IndexTag... >, Kokkos::HostSpace > out_vector,
+    StringType const &... name_suffixes
+) 
+```
+
+
+
+
+
+**Parameters:**
+
+
+* `name_stem` The prefix for the names of the elements of the vector field. 
+* `out_vector` The vector field to be exposed. 
+* `name_suffixes` The suffixes for the names of the elements of the vector field. There must be the same number of suffixes as there are dimensions in the vector field. 
+
+
+
+
+        
 
 <hr>
 
