+++ assets/.doxy/gyselalibxx/gyselalibxx/pdi__helper_8hpp_source.md
@@ -10,8 +10,12 @@
 ```C++
 // SPDX-License-Identifier: MIT
 #pragma once
+#include <ddc/ddc.hpp>
+#include <ddc/pdi.hpp>
 
 #include <pdi.h>
+
+#include "vector_field.hpp"
 
 namespace detail {
 
@@ -52,6 +56,12 @@ auto get_vector_tuple(TupleType input_args, std::integer_sequence<size_t, I...>)
     return std::tie(std::get<I * 2 + 1>(input_args)...);
 }
 
+template <class TupleType, class StringTupleType, size_t... I>
+void expose_arrays(TupleType out_obj, StringTupleType names, std::integer_sequence<size_t, I...>)
+{
+    ((ddc::expose_to_pdi(std::get<I>(names).c_str(), std::get<I>(out_obj))), ...);
+}
+
 } // namespace detail
 
 template <class T, class... Args>
@@ -87,6 +97,20 @@ void PDI_expose_idx_range(IdxRange<Grids...> index_range, std::string name)
     }
     PDI_expose((name + "_starts").c_str(), starts_s_arr.data(), PDI_OUT);
     PDI_expose((name + "_extents").c_str(), extents_s_arr.data(), PDI_OUT);
+}
+
+template <class ElementType, class IdxRangeType, class... IndexTag, class... StringType>
+void PDI_expose_vector_field(
+        std::string const& name_stem,
+        VectorConstField<ElementType, IdxRangeType, VectorIndexSet<IndexTag...>, Kokkos::HostSpace>
+                out_vector,
+        StringType const&... name_suffixes)
+{
+    static_assert(sizeof...(StringType) == sizeof...(IndexTag));
+    static_assert((std::is_convertible_v<StringType, std::string> && ...));
+    std::tuple names = {(name_stem + name_suffixes)...};
+    std::tuple out_obj = std::make_tuple(ddcHelper::get<IndexTag>(out_vector)...);
+    detail::expose_arrays(out_obj, names, std::make_index_sequence<sizeof...(StringType)> {});
 }
 ```
 
