+++ tests/gyroaverage/gyroaverage_circular.cpp
@@ -176,57 +176,54 @@ protected:
         m_gyroradius = GetParam();
 
         // Start of the domain of interest in the R dimension
-        double const r_start = 0.;
-        double const r_end = m_r0;
+        CoordR const r_start(0.);
+        CoordR const r_end(m_r0);
 
         // Number of discretization points in the R dimension
         std::size_t const nb_r_points = 200;
 
         // Initialisation of the global domain in R
-        ddc::init_discrete_space<BSplinesR>(CoordR(r_start), CoordR(r_end), nb_r_points);
+        ddc::init_discrete_space<BSplinesR>(r_start, r_end, nb_r_points);
         ddc::init_discrete_space<GridR>(SplineInterpPointsR::get_sampling<GridR>());
 
         IdxRangeR const r_domain = SplineInterpPointsR::get_domain<GridR>();
 
         // Start of the domain of interest in the Theta dimension
-        double const theta_start = 0.;
-        double const theta_end = M_PI * 2.0;
+        CoordTheta const theta_start(0.);
+        CoordTheta const theta_end(M_PI * 2.0);
 
         // Number of discretization points in the Theta dimension
         std::size_t const nb_theta_points = 200;
 
         // Initialisation of the global domain in Theta
-        ddc::init_discrete_space<
-                BSplinesTheta>(CoordTheta(theta_start), CoordTheta(theta_end), nb_theta_points);
+        ddc::init_discrete_space<BSplinesTheta>(theta_start, theta_end, nb_theta_points);
         ddc::init_discrete_space<GridTheta>(SplineInterpPointsTheta::get_sampling<GridTheta>());
 
         IdxRangeTheta const theta_domain = SplineInterpPointsTheta::get_domain<GridTheta>();
 
         // Start of the domain of interest in the batch dimension (likely a mixture of
         // phi and vpar)
-        double const batch_start = 0.;
-        double const batch_end = M_PI * 2.0;
+        CoordBatch const batch_start(0.);
+        CoordBatch const batch_end(M_PI * 2.0);
 
         // Number of discretization points in the batch dimension
         std::size_t const nb_batch_points = 2;
 
         // Initialisation of the global domain in batch
-        IdxRangeBatch const batch_domain
-                = ddc::init_discrete_space<GridBatch>(GridBatch::init<GridBatch>(
-                        CoordBatch(batch_start),
-                        CoordBatch(batch_end),
-                        IdxStep<GridBatch>(nb_batch_points)));
+        IdxRangeBatch const batch_domain = ddc::init_discrete_space<GridBatch>(
+                GridBatch::init<
+                        GridBatch>(batch_start, batch_end, IdxStep<GridBatch>(nb_batch_points)));
 
         // Allocate members with move assign operator
         IdxRangeRTheta const rtheta_mesh = IdxRangeRTheta(r_domain, theta_domain);
         IdxRangeRThetaBatch const rthetabatch_mesh
                 = IdxRangeRThetaBatch(r_domain, theta_domain, batch_domain);
-        m_Rcoord_alloc = DFieldMemRTheta(rtheta_mesh, ddc::DeviceAllocator<double>());
[…]
