+++ src/gyroaverage/gyroaverage_operator.hpp
@@ -9,8 +9,6 @@
 #include "ddc_aliases.hpp"
 #include "geometry_pseudo_cartesian.hpp"
 
-namespace detail {
-
 /**
  * @brief Operator to compute the gyroaverage of a field in (r, theta) coordinates.
  *
@@ -24,19 +22,15 @@ namespace detail {
  * @tparam SplineRThetaBuilder      The type of the spline builder for the rtheta interpolation
  * @tparam SplineRThetaEvaluator    The type of the spline evaluator for the rtheta interpolation
  * @tparam IdxRangeRminorThetaBatch The index range over R, Theta and Batch directions.
- * @tparam CoordinateTransformFunction The Coordinate transform Functor.
+ * @tparam ToLogicalCoordTransform  Function to convert (R, Z) to (r, theta).
  */
 template <
         class SplineRThetaBuilder,
         class SplineRThetaEvaluator,
         class IdxRangeRminorThetaBatch,
-        class CoordinateTransformFunction>
+        class ToLogicalCoordTransform>
 class GyroAverageOperator
 {
-    // FIXME
-    // Need to add a static assert to check evaluator is addmissible to builder
-    // using ddc::is_evaluator_admissible
-    // This will be available from DDC 0.9.0
     using ExecutionSpace = typename SplineRThetaBuilder::exec_space;
 
     using GridRminor = typename SplineRThetaBuilder::interpolation_discrete_dimension_type1;
@@ -83,14 +77,25 @@ public:
     using DFieldRminorTheta = DField<IdxRangeRminorTheta>;
     using DFieldRminorThetaBatch = DField<IdxRangeRminorThetaBatch>;
     using DFieldBSRminorTheta = DField<IdxRangeBSRminorTheta>;
-    using DConstFieldRminorTheta = ConstField<double, IdxRangeRminorTheta>;
-    using DConstFieldRminorThetaBatch = ConstField<double, IdxRangeRminorThetaBatch>;
+    using DConstFieldRminorTheta = DConstField<IdxRangeRminorTheta>;
+    using DConstFieldRminorThetaBatch = DConstField<IdxRangeRminorThetaBatch>;
 
     using CoordRminorTheta = Coord<Rminor, Theta>;
     using CoordR_gyroTheta_gyro = Coord<R_gyro, Theta_gyro>;
-    using Rmajor = X_pC;
-    using Z = Y_pC;
+    using Rmajor = typename ToLogicalCoordTransform::cartesian_tag_x;
+    using Z = typename ToLogicalCoordTransform::cartesian_tag_y;
     using CoordRZ = Coord<Rmajor, Z>;
+
+    // FIXME
+    // Need to add a static assert to check evaluator is addmissible to builder
+    // using ddc::is_evaluator_admissible
+    // This will be available from DDC 0.9.0
+    static_assert(
+            is_mapping_v<ToLogicalCoordTransform>,
+            "CoordinateTransformFunction must be a mapping");
+    static_assert(std::is_same_v<typename ToLogicalCoordTransform::CoordArg, CoordRZ>);
+    static_assert(std::is_same_v<typename ToLogicalCoordTransform::CoordResult, CoordRminorTheta>);
+    static_assert(is_accessible_v<ExecutionSpace, ToLogicalCoordTransform>);
 
 private:
     /**
@@ -111,7 +116,7 @@ private:
     /**
      * @brief coordinate_transform Function to convert (R, Z) to (r, theta).
      */
-    CoordinateTransformFunction m_coordinate_transform;
+    ToLogicalCoordTransform m_coordinate_transform;
 
     /**
      * @brief Number of points to use in the gyroaverage integration (default: 8)
@@ -131,7 +136,7 @@ public:
[…]
