+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Gyroaverage operator for circular geometry
+- Add a Gyroaverage operator with tests for circular geometry.
 - Curvilinear coordinate change classes have an O-point method to retrieve the O-point in the non-curvilinear coordinates.
 - Add a batched `operator()` to `DiscreteToCartesian` allowing a field of coordinates to be converted.
 - Add a `LiePoissonBracket::operator()` overload which takes a 2D tensor as the second argument to the bracket.
