### 2025-07-02 9469a87
make internal types private
