### 2025-07-03 c30e6a6
cpu prepare.sh : set standalone spack paths
