### 2025-07-03 22a1f9e
cpu prepare.sh : use spack external find
