+++ tests/data_types/use_deriv_field_with_builder.cpp
@@ -136,32 +136,32 @@ TEST(DerivFieldTest, SplineBuilderUse)
     Idx<dX> first_dx(1);
     Idx<dY> first_dy(1);
 
-    DField<IdxRange<GridX, GridY>, Kokkos::HostSpace, Kokkos::layout_stride> function_exacted
+    DField<IdxRange<GridX, GridY>, Kokkos::HostSpace, Kokkos::layout_stride> function_extracted
             = function_and_derivs.get_values_field();
 
-    DField<IdxRange<GridY>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_xmin_exacted
+    DField<IdxRange<GridY>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_xmin_extracted
             = function_and_derivs[IdxdXX(first_dx, idx_range_slice_dx.front())];
-    DField<IdxRange<GridY>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_xmax_exacted
+    DField<IdxRange<GridY>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_xmax_extracted
             = function_and_derivs[IdxdXX(first_dx, idx_range_slice_dx.back())];
 
-    DField<IdxRange<GridX>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_ymin_exacted
+    DField<IdxRange<GridX>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_ymin_extracted
             = function_and_derivs[IdxdYY(first_dy, idx_range_slice_dy.front())];
-    DField<IdxRange<GridX>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_ymax_exacted
+    DField<IdxRange<GridX>, Kokkos::HostSpace, Kokkos::layout_stride> derivs_ymax_extracted
             = function_and_derivs[IdxdYY(first_dy, idx_range_slice_dy.back())];
 
     /*
         EXAMPLE OF USE FOR THE FIRST DERIVATIVES. 
 
     // Defined on (dX, GridX, slice GridY)
-    detail::ViewNDMaker<3, double, false>::type derivs_x_exacted
+    detail::ViewNDMaker<3, double, false>::type derivs_x_extracted
             = function_and_derivs.get_mdspan(deriv_block_x);
     // Defined on (dY, slice GridX, GridY)
-    detail::ViewNDMaker<3, double, false>::type derivs_y_exacted
+    detail::ViewNDMaker<3, double, false>::type derivs_y_extracted
             = function_and_derivs.get_mdspan(deriv_block_y);
     */
 
     // Defined on (dX, dY, slice GridX, GridY)
-    detail::ViewNDMaker<4, double, false>::type derivs_xy_exacted
+    detail::ViewNDMaker<4, double, false>::type derivs_xy_extracted
             = function_and_derivs.get_mdspan(deriv_block_xy);
 
 
@@ -169,23 +169,23 @@ TEST(DerivFieldTest, SplineBuilderUse)
     ddc::for_each(idx_range_xy, [&](IdxXY idx_xy) {
         double const x = ddc::select<X>(ddc::coordinate(idx_xy));
         double const y = ddc::select<Y>(ddc::coordinate(idx_xy));
-        function_exacted(idx_xy) = x * x + y * y + 2 * x * y;
+        function_extracted(idx_xy) = x * x + y * y + 2 * x * y;
     });
 
     ddc::for_each(idx_range_y, [&](IdxY idx_y) {
         double const y = ddc::coordinate(idx_y);
         double const x_min(xmin);
         double const x_max(xmax);
[…]
