### 2025-07-03 5eade22
Add gyroaverage operator for circular geometry (#309)
Resolves #22 
This PR aims at adding a gyroaverage operator for a circular geometry.

1. Currently works only for a circular geometry
2. Sequential along batch dimension based on the discussion last week

For the detail of implementation and testing, see the
[reference](https://gitlab.maisondelasimulation.fr/gysela-developpers/gysela/-/blob/master/doc/GYSELA_gyroaverage/GYSELA_gyroaverage.pdf)
