+++ assets/.doxy/gyselalibxx/gyselalibxx/structGyroAverageOperator_1_1R__gyro.md
@@ -1,0 +1,121 @@
+
+
+# Struct GyroAverageOperator::R\_gyro
+
+
+
+[**ClassList**](annotated.md) **>** [**R\_gyro**](structGyroAverageOperator_1_1R__gyro.md)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Public Types
+
+| Type | Name |
+| ---: | :--- |
+| typedef R\_gyro\_cov | [**Dual**](#typedef-dual)  <br> |
+
+
+
+
+
+
+## Public Static Attributes
+
+| Type | Name |
+| ---: | :--- |
+|  constexpr bool | [**PERIODIC**](#variable-periodic)   = `false`<br> |
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Public Types Documentation
+
+
+
+
+### typedef Dual 
+
+```C++
+using GyroAverageOperator< SplineRThetaBuilder, SplineRThetaEvaluator, IdxRangeRminorThetaBatch, ToLogicalCoordTransform >::R_gyro::Dual =  R_gyro_cov;
+```
+
+
+
+
+<hr>
+## Public Static Attributes Documentation
+
+
+
+
+### variable PERIODIC 
+
+```C++
+constexpr bool GyroAverageOperator< SplineRThetaBuilder, SplineRThetaEvaluator, IdxRangeRminorThetaBatch, ToLogicalCoordTransform >::R_gyro::PERIODIC;
+```
+
+
+
+
+<hr>
+
+------------------------------
+The documentation for this class was generated from the following file `/home/runner/work/gyselalibxx/gyselalibxx/code_branch/src/gyroaverage/gyroaverage_operator.hpp`
+
