+++ assets/.doxy/gyselalibxx/gyselalibxx/class_member_functions.md
@@ -177,6 +177,7 @@
 * **get\_sampling** ([**GaussLegendre**](classGaussLegendre.md), [**ddcHelper::NonUniformInterpolationPoints**](classddcHelper_1_1NonUniformInterpolationPoints.md))
 * **get\_sampling\_on\_cell** ([**GaussLegendre**](classGaussLegendre.md))
 * **Gradient** ([**Gradient**](classGradient.md))
+* **GyroAverageOperator** ([**GyroAverageOperator**](classGyroAverageOperator.md))
 * **get\_amplitudes** ([**KrookSourceAdaptive**](classKrookSourceAdaptive.md))
 * **getclosest** ([**Lagrange**](classLagrange.md))
 * **getclosest\_binsearch** ([**Lagrange**](classLagrange.md))
@@ -317,7 +318,7 @@
 
 ## o
 
-* **operator()** ([**AdvectionFieldFinder**](classAdvectionFieldFinder.md), [**BarycentricToCartesian**](classBarycentricToCartesian.md), [**BslAdvection1D**](classBslAdvection1D.md), [**BslAdvectionPolar**](classBslAdvectionPolar.md), [**BslAdvectionSpatial**](classBslAdvectionSpatial.md), [**BslAdvectionVelocity**](classBslAdvectionVelocity.md), [**BslExplicitPredCorrRTheta**](classBslExplicitPredCorrRTheta.md), [**BslImplicitPredCorrRTheta**](classBslImplicitPredCorrRTheta.md), [**BslPredCorrRTheta**](classBslPredCorrRTheta.md), [**BumpontailEquilibrium**](classBumpontailEquilibrium.md), [**CartesianToBarycentric**](classCartesianToBarycentric.md), [**CartesianToCircular**](classCartesianToCircular.md), [**CartesianToCylindrical**](classCartesianToCylindrical.md), [**CartesianToCzarny**](classCartesianToCzarny.md), [**CentralFDMPartialDerivative**](classCentralFDMPartialDerivative.md), [**CentralFDMPartialDerivativeWithBValue**](classCentralFDMPartialDerivativeWithBValue.md), [**ChargeDensityCalculator**](classChargeDensityCalculator.md), [**CircularToCartesian**](classCircularToCartesian.md), [**CollisionOperator**](classCollisionOperator.md), [**CollisionsInter**](classCollisionsInter.md), [**CollisionsIntra**](classCollisionsIntra.md), [**CombinedMapping**](classCombinedMapping.md), [**ConstantExtrapolationRuleOnion**](structConstantExtrapolationRuleOnion.md), [**ConstantPartialDerivative**](classConstantPartialDerivative.md), [**CylindricalToCartesian**](classCylindricalToCartesian.md), [**CzarnyToCartesian**](classCzarnyToCartesian.md), [**DerivField&lt; ElementType, IdxRange&lt; DDims... &gt;, MemorySpace, LayoutStridedPolicy &gt;**](classDerivField_3_01ElementType_00_01IdxRange_3_01DDims_8_8_8_01_4_00_01MemorySpace_00_01LayoutStridedPolicy_01_4.md), [**DerivFieldMem&lt; ElementType, IdxRange&lt; DDims... &gt;, NDerivs, MemSpace &gt;**](classDerivFieldMem_3_01ElementType_00_01IdxRange_3_01DDims_8_8_8_01_4_00_01NDerivs_00_01MemSpace_01_4.md), [**DiscreteToCartesian**](classDiscreteToCartesian.md), [**DiscreteToCartesianBuilder**](classDiscreteToCartesianBuilder.
[…]
