### 2025-07-03 #8304 @kyleniemeyer
@EmilyBourne yes, JOSS reviewers will look at the whole repo, but only a specific version of the software. The reviewers are not expected to review unreleased code in a development branch, just the version that you submitted (v0.1.0).

It's fine to have an open PR with the paper, though typically comments on the paper just come to the review issue here.
