### 2025-07-03 #8304 @EmilyBourne
v0.1.0 is a tagged version which can be found here: https://github.com/gyselax/gyselalibxx/releases/tag/v0.1.0

and on the `main` branch.

We will be releasing v0.2.0 shortly. Is that a problem given that at the end of the JOSS review I will be asked to tag and deposit on zenodo a version that was reviewed plus any changes requested by reviewers? If I follow my usual workflow the tagged version is likely to be v0.2.1
