### 2025-07-03 #8304 @kyleniemeyer
If you would like the JOSS review to focus on v0.2.0, then we should wait to assign reviewers until that version is released.
