### 2025-07-03 10ffd5c
Allow a GPU spline evaluator to be used to initialise polar bsplines (#346)
Allow the discrete mapping passed to the constructor of `PolarBSplines`
to be defined on either CPU or GPU. Fixes #342 . The method `o_point` is
used to retrieve the coordinate of the o-point and the new method
`control_points` is used to retrieve the control points. The mapping is
not used for any other calculations.

**Commit Summary**
- Use a coordinate transformation defined on GPU to initialise
`PolarBSplines` when the CPU object is unused in the rest of the code
- Add a `control_points` method to `DiscreteToCarte
[…]
