+++ src/coord_transformations/discrete_to_cartesian.hpp
@@ -342,17 +342,17 @@ public:
      * identified by its control points @f$ \{(c_{x,k}, c_{y,k})\}_{k}  @f$ where
      * @f$ c_{x,k} @f$ and @f$ c_{y,k} @f$ are the B-splines coefficients:
      *
-     * @f$ x(r,\theta) = \sum_{k=0}^{N_r\times N_{\theta}-1} c_{x, k} B_k{r,\theta} @f$,
+     * @f$ x(r,\theta) = \sum_{k=0}^{N_r\times N_{\theta}-1} c_{x, k} B_k(r,\theta) @f$,
      *
-     * @f$ y(r,\theta) = \sum_{k=0}^{N_r\times N_{\theta}-1} c_{y, k} B_k{r,\theta} @f$,
+     * @f$ y(r,\theta) = \sum_{k=0}^{N_r\times N_{\theta}-1} c_{y, k} B_k(r,\theta) @f$,
      *
      * where @f$ N_r\times N_{\theta} @f$ is the number of B-splines.
      *
      * The control points can be obtained by interpolating the mapping on interpolation
      * points (see GrevilleInterpolationPoints or KnotsAsInterpolationPoints).
      * We can also note that the first control points @f$ \{(c_{x,k}, c_{y,k})\}_{k=0}^{N_{\theta}} @f$
-     * are equal to the pole @f$ (c_{x,k}, c_{y,k}) = (x_0,y_0) @f$, @f$ \forall k = 0, ..., N_{\theta}-1 @f$
-     * where @f$ x(0,\theta), y(0,\theta) = (x_0,y_0) @f$ @f$ \forall \theta @f$.
+     * are equal to the pole @f$ (c_{x,k}, c_{y,k}) = (x_0,y_0) , \ \forall k = 0, ..., N_{\theta}-1 @f$
+     * where @f$ x(0,\theta), y(0,\theta) = (x_0,y_0) \ \forall \theta @f$.
      *
      *
      * @param[in] el
@@ -367,6 +367,48 @@ public:
             Idx<BSplineR, BSplineTheta> const& el) const
     {
         return Coord<X, Y>(m_x_spline_representation(el), m_y_spline_representation(el));
+    }
+
+    /**
+     * @brief Get a set of control points of the mapping on B-splines.
+     *
+     * The mapping @f$ (r,\theta) \mapsto (x,y) @f$ decomposed on B-splines can be
+     * identified by its control points @f$ \{(c_{x,k}, c_{y,k})\}_{k}  @f$ where
+     * @f$ c_{x,k} @f$ and @f$ c_{y,k} @f$ are the B-splines coefficients:
+     *
+     * @f$ x(r,\theta) = \sum_{k=0}^{N_r\times N_{\theta}-1} c_{x, k} B_k(r,\theta) @f$,
+     *
+     * @f$ y(r,\theta) = \sum_{k=0}^{N_r\times N_{\theta}-1} c_{y, k} B_k(r,\theta) @f$,
+     *
+     * where @f$ N_r\times N_{\theta} @f$ is the number of B-splines.
+     *
+     * The control points can be obtained by interpolating the mapping on interpolation
+     * points (see GrevilleInterpolationPoints or KnotsAsInterpolationPoints).
+     * We can also note that the first control points @f$ \{(c_{x,k}, c_{y,k})\}_{k=0}^{N_{\theta}} @f$
+     * are equal to the pole @f$ (c_{x,k}, c_{y,k}) = (x_0,y_0) , \ \forall k = 0, ..., N_{\theta}-1 @f$
+     * where @f$ x(0,\theta), y(0,\theta) = (x_0,y_0) \ \forall \theta @f$.
+     *
+     *
+     * @param[in] exec_space
[…]
