+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a batched `operator()` to `DiscreteToCartesian` allowing a field of coordinates to be converted.
 - Add a `LiePoissonBracket::operator()` overload which takes a 2D tensor as the second argument to the bracket.
 - Add a function `PDI_expose_vector_field` to output a vector field via PDI.
+- Add a `control_points` method to `DiscreteToCartesian` to allow all control points to be retrieved at once.
 
 ### Fixed
 
@@ -36,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generalise `compute_coeffs_on_mapping` to work with any mapping.
 - Rely on GPU-aware MPI to allow GPU-direct MPI for `MPITransposeAllToAll`.
 - Curvilinear coordinate change classes take a `Coord` type to specify the O-point in the constructor.
+- Allow `init_discrete_space` to be used to initialise `PolarBSplines` with a GPU-based `DiscreteToCartesian` coordinate change operator.
 
 ### Deprecated
 
