+++ assets/.doxy/gyselalibxx/gyselalibxx/class_members.md
@@ -148,6 +148,7 @@
 * **chunk\_span** ([**DerivFieldCommon&lt; FieldType, IdxRange&lt; DDims... &gt; &gt;**](classDerivFieldCommon_3_01FieldType_00_01IdxRange_3_01DDims_8_8_8_01_4_01_4.md))
 * **chunk\_view** ([**DerivFieldCommon&lt; FieldType, IdxRange&lt; DDims... &gt; &gt;**](classDerivFieldCommon_3_01FieldType_00_01IdxRange_3_01DDims_8_8_8_01_4_01_4.md))
 * **control\_point** ([**DiscreteToCartesian**](classDiscreteToCartesian.md))
+* **control\_points** ([**DiscreteToCartesian**](classDiscreteToCartesian.md))
 * **CoordField** ([**FEM1DPoissonSolver**](classFEM1DPoissonSolver.md))
 * **CoordFieldMem** ([**FEM1DPoissonSolver**](classFEM1DPoissonSolver.md))
 * **CoordPDEDim** ([**FEM1DPoissonSolver**](classFEM1DPoissonSolver.md))
@@ -275,9 +276,9 @@
 * **deriv\_1\_and\_2** ([**MultipatchSplineEvaluator2D**](classMultipatchSplineEvaluator2D.md), [**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
 * **deriv\_dim\_1** ([**MultipatchSplineEvaluator2D**](classMultipatchSplineEvaluator2D.md), [**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
 * **deriv\_dim\_2** ([**MultipatchSplineEvaluator2D**](classMultipatchSplineEvaluator2D.md), [**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
-* **DimR** ([**PolarBSplines**](classPolarBSplines.md), [**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
-* **DimTheta** ([**PolarBSplines**](classPolarBSplines.md), [**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
 * **discrete\_vector\_type** ([**PolarBSplines::Impl**](classPolarBSplines_1_1Impl.md), [**SpeciesInformation::Impl**](classSpeciesInformation_1_1Impl.md), [**TriangularBernsteinPolynomialBasis::Impl**](classTriangularBernsteinPolynomialBasis_1_1Impl.md))
+* **DimR** ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
+* **DimTheta** ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
 * **deriv\_dim\_1\_and\_2** ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
 * **DFieldRTheta** ([**PolarSplineFEMPoissonLikeSolver**](classPolarSplineFEMPoissonLikeSolver.md))
 * **derivative** ([**PolarSplineFEMPoissonLikeSolver::EvalDeriv1DType**](structPolarSplineFEMPoissonLikeSolver_1_1EvalDeriv1DType.md), [**PolarSplineFEMPoissonLikeSolver::EvalDeriv2DType**](structPolarSplineFEMPoissonLikeSolver_1_1EvalDeriv2DType.md))
@@ -1025,7 +1026,7 @@
 
 ## r
 
[…]
