+++ assets/.doxy/gyselalibxx/gyselalibxx/class_member_typedefs.md
@@ -174,9 +174,9 @@
 * **Derivs12OnPatch** ([**MultipatchSplineBuilder2D**](classMultipatchSplineBuilder2D.md))
 * **Derivs1OnPatch** ([**MultipatchSplineBuilder2D**](classMultipatchSplineBuilder2D.md))
 * **Derivs2OnPatch** ([**MultipatchSplineBuilder2D**](classMultipatchSplineBuilder2D.md))
-* **DimR** ([**PolarBSplines**](classPolarBSplines.md), [**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
-* **DimTheta** ([**PolarBSplines**](classPolarBSplines.md), [**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
 * **discrete\_vector\_type** ([**PolarBSplines::Impl**](classPolarBSplines_1_1Impl.md), [**SpeciesInformation::Impl**](classSpeciesInformation_1_1Impl.md), [**TriangularBernsteinPolynomialBasis::Impl**](classTriangularBernsteinPolynomialBasis_1_1Impl.md))
+* **DimR** ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
+* **DimTheta** ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md))
 * **DFieldRTheta** ([**PolarSplineFEMPoissonLikeSolver**](classPolarSplineFEMPoissonLikeSolver.md))
 * **DFieldBS** ([**Spline1DPartialDerivative**](classSpline1DPartialDerivative.md), [**Spline2DPartialDerivative**](classSpline2DPartialDerivative.md))
 * **DFieldBSMem** ([**Spline1DPartialDerivative**](classSpline1DPartialDerivative.md), [**Spline2DPartialDerivative**](classSpline2DPartialDerivative.md))
@@ -489,7 +489,7 @@
 
 ## r
 
-* **R** ([**BslAdvectionPolar**](classBslAdvectionPolar.md), [**IPolarFootFinder**](classIPolarFootFinder.md), [**OnionPatchLocator&lt; MultipatchType&lt; IdxRangeOnPatch, Patches... &gt;, LogicalToPhysicalMapping, PhysicalToLogicalMapping, ExecSpace &gt;**](classOnionPatchLocator_3_01MultipatchType_3_01IdxRangeOnPatch_00_01Patches_8_8_8_01_4_00_01Logicff6c45b073183ccdfc0de0e4a415a7fa.md), [**PolarSplineFEMPoissonLikeSolver**](classPolarSplineFEMPoissonLikeSolver.md), [**RefinedDiscreteToCartesianBuilder**](classRefinedDiscreteToCartesianBuilder.md), [**SplinePolarFootFinder**](classSplinePolarFootFinder.md), [**ToroidalToCylindrical**](classToroidalToCylindrical.md))
[…]

[…]
