### 2025-07-03 51739cd
v0.2.0 (#365)
Co-authored-by: Thomas Padioleau <thomas.padioleau@cea.fr>
Co-authored-by: Virginie Grandgirard <49058376+gdgirard@users.noreply.github.com>
Co-authored-by: gdgirard <gdgirard@login6.head.adastra.cines.fr>
Co-authored-by: geoflow <elkmine31@hotmail.fr>
Co-authored-by: Etienne Malaboeuf <eti.malaboeuf@gmail.com>
Co-authored-by: yasahi-hpc <57478230+yasahi-hpc@users.noreply.github.com>
