+++ src/interpolation/polar_splines/polar_bsplines.hpp
@@ -54,10 +54,10 @@ public:
 
 
     /// The tag for the radial direction of the B-splines.
-    using DimR = typename BSplinesR::continuous_dimension_type;
+    using R = typename BSplinesR::continuous_dimension_type;
 
     /// The tag for the poloidal direction of the B-splines.
-    using DimTheta = typename BSplinesTheta::continuous_dimension_type;
+    using Theta = typename BSplinesTheta::continuous_dimension_type;
 
 public:
     /// The continuity enforced by the B-splines at the singular point.
@@ -243,28 +243,43 @@ public:
          *                                  singular point and determine the Barycentric coordinates which are used to define
          *                                  the new basis splines which cross the singular point.
          */
-        template <class DiscreteMapping>
-        explicit Impl(const DiscreteMapping& curvilinear_to_cartesian)
+        template <class X, class Y, class SplineEvaluator, class EvalMemorySpace>
+        explicit Impl(const DiscreteToCartesian<X, Y, SplineEvaluator, R, Theta, EvalMemorySpace>&
+                              curvilinear_to_cartesian)
         {
-            static_assert(is_accessible_v<Kokkos::DefaultHostExecutionSpace, DiscreteMapping>);
             static_assert(std::is_same_v<MemorySpace, Kokkos::HostSpace>);
-            using DimX = typename DiscreteMapping::cartesian_tag_x;
-            using DimY = typename DiscreteMapping::cartesian_tag_y;
-            using mapping_tensor_product_index_type
-                    = Idx<typename DiscreteMapping::BSplineR,
-                          typename DiscreteMapping::BSplineTheta>;
+            using DiscreteMapping
+                    = DiscreteToCartesian<X, Y, SplineEvaluator, R, Theta, EvalMemorySpace>;
+            static_assert(std::is_same_v<typename DiscreteMapping::BSplineR, BSplinesR>);
+            static_assert(std::is_same_v<typename DiscreteMapping::BSplineTheta, BSplinesTheta>);
+            using EvalExecSpace = std::conditional_t<
+                    std::is_same_v<EvalMemorySpace, Kokkos::HostSpace>,
+                    Kokkos::DefaultHostExecutionSpace,
+                    Kokkos::DefaultExecutionSpace>;
             if constexpr (C > -1) {
-                const Coord<DimX, DimY> pole
-                        = curvilinear_to_cartesian(Coord<DimR, DimTheta>(0.0, 0.0));
-                const double x0 = ddc::get<DimX>(pole);
-                const double y0 = ddc::get<DimY>(pole);
-                double tau = 0.0;
-                for (std::size_t i(0); i < ddc::discrete_space<BSplinesTheta>().size(); ++i) {
-                    const Coord<DimX, DimY> point = curvilinear_to_cartesian.control_point(
-                            mapping_tensor_product_index_type(1, i));
+                IdxRange<BSplinesR> idx_range_r_ctrl_pts(IdxR(0), IdxStepR(std::max(2, C + 1)));
+                IdxRange<BSplinesR, BSplinesTheta> idx_range_ctrl_pts(
+                        idx_range_r_ctrl_pts,
+                        ddc::discrete_space<BSplinesTheta>().full_domain());
+                FieldMem<Coord<X, Y>, IdxRange<BSplinesR, BSplinesTheta>, EvalMemorySpace>
[…]

[…]
