### 2025-07-03 #8304 @piyueh
> @piyueh for the versioning for Zenodo it will probably be easier to work from v0.2.0. I will try to keep changes on devel to a minimum while the review is ongoing.

Just a side note and maybe @kyleniemeyer can correct me if I'm wrong: I think it's not necessary to limit the activities on the `devel` branch. Reviewers are only expected to review the code in the snapshot of [`v0.2.0`](https://github.com/gyselax/gyselalibxx/tree/v0.2.0). Activities in other branches are irrelevant to this JOSS submission. For example, if you add new features in the `devel` branch during the JOSS review, these new features in `devel` will not be covered by the final JOSS paper.
