### 2025-07-03 a578217
Add Baptiste and Yann as authors
