### 2025-07-03 973f34a
prepare.sh : update existing-folder check (-> not home)
