+++ toolchains/cpu.spack.gyselalibxx_env/prepare.sh
@@ -6,29 +6,6 @@ if [[ "${BASH_SOURCE[0]}" != "${0}" ]]; then
     return 1
 fi
 
-if [ -d "${HOME}/.spack" ]; then
-    echo "It is recommended to remove any old spack installation $HOME/.spack before running this script."
-    echo "$HOME/.spack detected, continue anyway? [YyNn]"
-    read ignore_existing
-    if [[ $ignore_existing == y* ]]; then
-        echo "Ignoring $HOME/.spack"
-    else
-        exit 1
-    fi
-fi
-
-set -e
-
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-
-# Download spack
-wget https://github.com/spack/spack/releases/download/v0.23.1/spack-0.23.1.tar.gz
-tar -xf spack-0.23.1.tar.gz
-rm spack-0.23.1.tar.gz
-
-# Bug fix for spack < 0.24 and gcc>=14
-wget https://raw.githubusercontent.com/spack/spack/b369d8b2509794c4f46f62c81f25c247ca58418e/var/spack/repos/builtin/packages/py-netcdf4/package.py -O spack-0.23.1/var/spack/repos/builtin/packages/py-netcdf4/package.py
-
 # set paths for spack to not use home ~/.spack folder
 CURRENT_DIR=$(pwd)
 export SPACK_PATH=${CURRENT_DIR}/spack-0.23.1/
@@ -36,8 +13,31 @@ export SPACK_USER_CONFIG_PATH="${SPACK_PATH}/user_config"
 export SPACK_SYSTEM_CONFIG_PATH="${SPACK_PATH}/sys_config"
 export SPACK_USER_CACHE_PATH="${SPACK_PATH}/user_cache"
 
+set -e
+
+if [ -d "${SPACK_PATH}" ]; then
+    echo "It is recommended to remove any old spack installation ${SPACK_PATH} before running this script."
+    echo "${SPACK_PATH} detected, continue anyway? [YyNn]"
+    read ignore_existing
+    if [[ $ignore_existing == y* ]]; then
+        echo "Using existing ${SPACK_PATH}"
+    else
+        exit 1
+    fi
+else
+    # Download spack
+    wget https://github.com/spack/spack/releases/download/v0.23.1/spack-0.23.1.tar.gz
+    tar -xf spack-0.23.1.tar.gz
+    rm spack-0.23.1.tar.gz
+fi
+
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+
+# Bug fix for spack < 0.24 and gcc>=14
+wget https://raw.githubusercontent.com/spack/spack/b369d8b2509794c4f46f62c81f25c247ca58418e/var/spack/repos/builtin/packages/py-netcdf4/package.py -O spack-0.23.1/var/spack/repos/builtin/packages/py-netcdf4/package.py
+
 # Activate spack
-. spack-0.23.1/share/spack/setup-env.sh
+. ${SPACK_PATH}/share/spack/setup-env.sh
 
 # Reduce the naming scheme of packages to avoid shebang issues.
 # Increase the time out that is by default too short for some packages (like PDI)
