### 2025-07-04 2f41c68
Fix missing equations in docs (#368)
Fix missing equations in docs by pointing at a branch of my fork which
contains all fixes as PRs to MkDoxy are still not merged. Fixes #296
