### 2025-07-07 8a8ea64
prepare.sh : un-find curl and perl after external find
