+++ src/data_types/derivative_field_common.hpp
@@ -584,9 +584,9 @@ public:
      *
      * @returns Field The field on the physical index range.
      */
-    auto get_values_field() const
-    {
-        IdxRange<> no_specified_dims;
-        return get_internal_field(no_specified_dims).span_cview();
-    }
+    // auto get_values_field() const
+    // {
+    //     IdxRange<> no_specified_dims;
+    //     return get_internal_field(no_specified_dims).span_cview();
+    // }
 };
