+++ src/data_types/derivative_field_mem.hpp
@@ -59,6 +59,8 @@ public:
     /// @brief The type of the elements in the chunks.
     using element_type = typename base_type::element_type;
 
+    using memory_space = MemSpace;
+
     /**
      * @brief The IdxRange on which the chunks in this object are defined.
      *
