+++ tests/multipatch/interface_derivatives/interface_derivative_matrix_test.cpp
@@ -8,12 +8,15 @@
 #include "9patches_2d_periodic_strips_non_uniform.hpp"
 #include "ddc_alias_inline_functions.hpp"
 #include "ddc_helper.hpp"
+#include "derivative_field.hpp"
+#include "derivative_field_mem.hpp"
 #include "interface.hpp"
 #include "interface_derivative_matrix.hpp"
 #include "interface_derivatives_test_utils.hpp"
 #include "mesh_builder.hpp"
 #include "non_uniform_interpolation_points.hpp"
 #include "single_interface_derivatives_calculator.hpp"
+#include "view.hpp"
 
 
 /*
@@ -127,6 +130,38 @@ using SplineRThetagEvaluator = ddc::SplineEvaluator2D<
         ddc::PeriodicExtrapolationRule<Xg>,
         ddc::ConstantExtrapolationRule<Yg, Xg>,
         ddc::ConstantExtrapolationRule<Yg, Xg>>;
+
+
+template <class Patch>
+using DerivFieldMemOnPatch = DerivFieldMem<
+        double,
+        IdxRange<
+                ddc::Deriv<typename Patch::Grid1>,
+                ddc::Deriv<typename Patch::Grid2>,
+                typename Patch::Grid1,
+                typename Patch::Grid2>,
+        2>;
+
+template <class Patch>
+using DerivFieldMemOnPatch_host = host_t<DerivFieldMemOnPatch<Patch>>;
+
+template <class Patch>
+using DerivFieldOnPatch = DerivField<
+        double,
+        IdxRange<
+                ddc::Deriv<typename Patch::Grid1>,
+                ddc::Deriv<typename Patch::Grid2>,
+                typename Patch::Grid1,
+                typename Patch::Grid2>>;
+
+template <class Patch>
+using DerivFieldOnPatch_host = host_t<DerivFieldOnPatch<Patch>>;
+
+template <class Patch>
+using IdxRange1SliceOnPatch = IdxRangeSlice<typename Patch::Grid1>;
+
+template <class Patch>
+using IdxRange2SliceOnPatch = IdxRangeSlice<typename Patch::Grid2>;
 
 
 // USEFUL FUNCTIONS ==============================================================================
@@ -299,6 +334,316 @@ void initialise_all_xy_derivatives_max_max(
      ...);
 }
 
+
+template <class Patch>
+void copy_function_values_in_deriv_field_on_patch(
+        DerivFieldOnPatch_host<Patch> function_and_derivs,
+        DFieldOnPatch_host<Patch> const& function)
+{
+    DField<typename Patch::IdxRange12, Kokkos::HostSpace, Kokkos::layout_stride> function_extracted
+            = function_and_derivs.get_values_field();
+    ddc::for_each(get_idx_range(function), [&](typename Patch::Idx12 const idx) {
+        function_extracted(idx) = function(idx);
+    });
+}
+
+template <class... Patches>
+void copy_function_values_in_deriv_field(
+        MultipatchField<DerivFieldOnPatch_host, Patches...> functions_and_derivs,
+        MultipatchField<DFieldOnPatch_host, Patches...> const& functions)
+{
+    (copy_function_values_in_deriv_field_on_patch<Patches>(
+             functions_and_derivs.template get<Patches>(),
+             functions.template get<Patches>()),
+     ...);
+}
+
+
+template <class Patch>
[…]
