+++ src/multipatch/interface_derivatives/interface_derivative_matrix.hpp
@@ -14,6 +14,7 @@
 #include "multipatch_type.hpp"
 #include "single_interface_derivatives_calculator.hpp"
 #include "types.hpp"
+#include "view.hpp"
 
 /**
  * TODO:
@@ -184,6 +185,32 @@ class InterfaceDerivativeMatrix
             IdxRange<ddc::Deriv<typename Patch::Dim1>, ddc::Deriv<typename Patch::Dim2>>,
             typename ExecSpace::memory_space>;
 
+    template <class Patch>
+    using DerivFieldOnPatch = DerivField<
+            double,
+            IdxRange<
+                    ddc::Deriv<typename Patch::Grid1>,
+                    ddc::Deriv<typename Patch::Grid2>,
+                    typename Patch::Grid1,
+                    typename Patch::Grid2>>;
+
+    template <class Patch>
+    using DerivFieldOnPatch_host = host_t<DerivFieldOnPatch<Patch>>;
+
+
+    template <class Patch>
+    using IdxRange1SliceOnPatch = IdxRangeSlice<typename Patch::Grid1>;
+
+    template <class Patch>
+    using IdxRange2SliceOnPatch = IdxRangeSlice<typename Patch::Grid2>;
+
+    struct eval_deriv
+    {
+    };
+    struct eval_cross_deriv
+    {
+    };
+
 
     // HELPFUL ALIASES ===========================================================================
     // Get the type of the interface given to define the Connectivity class.
@@ -301,12 +328,15 @@ public:
      * @param idx_ranges MultipatchType collection of index ranges defined on the given list of patches. 
      * @param derivatives_calculators Tuple of SingleInterfaceDerivativesCalculator containing all the 
      *          interface derivative calculator for each interface in the given Grid1D direction. 
+     * @param reduction_factor Reduction factor for a Gingko dense matrix with a conjugate gradient solver 
+     *          and a Jacobi preconditioner. 
      * @warning The interface derivative calculators have to be ordered in the tuple. 
      *      The order has to be the same as the interfaces in the Grid1D direction. 
      */
     InterfaceDerivativeMatrix(
             MultipatchType<IdxRangeOnPatch, Patches...> const& idx_ranges,
-            SingleInterfaceDerivativesCalculatorTuple const derivatives_calculators)
+            SingleInterfaceDerivativesCalculatorTuple const derivatives_calculators,
+            double const& reduction_factor = 1e-6)
         : m_idx_ranges(idx_ranges)
         , m_derivatives_calculators(derivatives_calculators)
     {
@@ -322,8 +352,9 @@ public:
                                   gko::stop::Iteration::build()
                                           .with_max_iters(n_inner_interfaces)
                                           .on(exec),
-                                  gko::stop::ResidualNorm<>::build().with_reduction_factor(1e-6).on(
-                                          exec))
+                                  gko::stop::ResidualNorm<>::build()
+                                          .with_reduction_factor(reduction_factor)
+                                          .on(exec))
                           .with_preconditioner(
                                   PreconditionerJ::build().with_max_block_size(8u).on(exec))
                           .on(exec);
@@ -415,6 +446,34 @@ public:
                 std::make_integer_sequence<std::size_t, n_inner_interfaces> {});
     }
 
+    template <class IdxPar>
+    void solve_deriv( // Should be useful for non-conforming case (later).
+            MultipatchField<DerivFieldOnPatch_host, Patches...> functions_and_derivs,
[…]
