### 2025-07-07 24a2e2f
Complete definition
