+++ tests/testing_tools/test_math_utils.hpp
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: MIT
 #pragma once
+#include <Kokkos_Core.hpp>
 
-KOKKOS_FUNCTION double cyl_bessel_j();
+KOKKOS_FUNCTION double cyl_bessel_j(const double nu, const double x);
