### 2025-07-07 3d427c9
Use Kokkos maths
