+++ tests/testing_tools/test_math_utils.hpp
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: MIT
 #pragma once
-#include <Kokkos_Core.hpp>
 
-KOKKOS_FUNCTION double cyl_bessel_j(const double nu, const double x);
+double cyl_bessel_j(double nu, double x);
