### 2025-07-07 a4aef83
Add a warning documentation
