### 2025-07-07 fc1ff04
Add assert < 30
