### 2025-07-07 a31e69a
Correct explanation in CHANGELOG
