### 2025-07-08 f224e4b
Put back cmath. Prefer tgamma. Increase loops
