### 2025-07-08 b1d238d
tgamma doesn't like 0
