+++ tests/testing_tools/test_math_utils.cpp
@@ -6,10 +6,10 @@
 double cyl_bessel_j(const double nu, const double x)
 {
     assert(nu < 30);
-    double sum = std::pow(x / 2, nu) / std::tgamma(nu + 1);
-    for (int k = 1; k < 100; ++k) {
+    double sum = 0;
+    for (int k = 0; k < 100; ++k) {
         sum += std::pow(-1, k) * std::pow(x / 2, 2 * k + nu)
-               / (std::tgamma(k) * std::tgamma(nu + k + 1));
+               / (std::tgamma(k + 1) * std::tgamma(nu + k + 1));
     }
     return sum;
 }
