### 2025-07-08 e9d412c
prepare.sh : remove `spack external find`
