### 2025-07-08 1b5a68c
Fix physical_deriv_grids definition
