+++ src/utils/type_seq_tools.hpp
@@ -64,6 +64,15 @@ struct FindGrid<Dim, ddc::detail::TypeSeq<>>
     using type = void;
 };
 
+template <class TypeSeqDims, class TypeSeqGrids>
+struct FindAllGrids;
+
+template <class... TypeSeqDims, class TypeSeqGrids>
+struct FindAllGrids<ddc::detail::TypeSeq<TypeSeqDims...>, TypeSeqGrids>
+{
+    using type = ddc::detail::TypeSeq<typename FindGrid<TypeSeqDims, TypeSeqGrids>::type...>;
+};
+
 template <class CoordType, class IdxRangeType>
 struct FindIdxType;
 
@@ -200,6 +209,10 @@ using type_seq_cat_t = typename detail::TypeSeqCat<TypeSeqs...>::type;
 /// A tool to find the grid that is defined along the specified dimension (e.g. get GridX from X)
 template <class Dim, class TypeSeqGrid>
 using find_grid_t = typename detail::FindGrid<Dim, TypeSeqGrid>::type;
+
+/// A tool to find the grid that is defined along the specified dimension (e.g. get GridX from X)
+template <class TypeSeqDim, class TypeSeqGrid>
+using find_all_grids_t = typename detail::FindAllGrids<TypeSeqDim, TypeSeqGrid>::type;
 
 /**
  * @brief Get a TypeSeq containing all unique types from the original TypeSeq.
