### 2025-07-08 ea761cc
Clang formatting
