### 2025-07-09 a14bee2
Run tests as folder-agnostic cppcheck addons
