+++ bin/ci_tools/check_naming_conventions.py
@@ -1,0 +1,107 @@
+import argparse
+from pathlib import Path
+import re
+import sys
+
+import cppcheckdata
+
+ddc_keyword_map = {'DiscreteElement': 'Idx',
+                   'DiscreteVector' : 'IdxStep',
+                   'DiscreteDomain' : 'IdxRange',
+                   'Chunk'          : 'FieldMem',
+                   'ChunkSpan'      : 'Field',
+                   'ChunkView'      : 'ConstField',
+                   'Coordinate'     : 'Coord',
+                   'span_view'      : 'get_field',
+                   'span_cview'     : 'get_const_field',
+                   'domain'         : 'get_idx_range',
+                   'spline_domain'  : 'get_spline_idx_range',
+                   'UniformPointSampling' : 'UniformGridBase',
+                   'NonUniformPointSampling' : 'NonUniformGridBase'}
+
+HOME_DIR = Path(__file__).parent.parent.parent.absolute()
+
+file_exceptions = [HOME_DIR / 'src' / 'utils' / 'ddc_helper.hpp',
+                   HOME_DIR / 'src' / 'utils' / 'ddc_aliases.hpp',
+                   HOME_DIR / 'src' / 'utils' / 'ddc_alias_inline_functions.hpp']
+
+RE_SNAKE_CASE = re.compile("[a-z0-9]+(?:_[a-z0-9]+)*")
+RE_CAMEL_CASE = re.compile("([A-Z][a-z0-9]*)+")
+
+def reportError(token, severity, msg, errorId):
+    cppcheckdata.reportError(token, severity, msg, 'naming', errorId)
+
+def main():
+    parser = cppcheckdata.ArgumentParser()
+    args = parser.parse_args()
+
+    dump_files, ctu_info_files = cppcheckdata.get_files(args)
+
+    for arg in dump_files:
+        if not arg.endswith('.dump') or any(f'{f}.dump' == arg for f in file_exceptions):
+            continue
+        print(f'Checking {arg}...')
+        data = cppcheckdata.parsedump(arg)
+
+        fname = Path(arg).stem
+        res = re.match(RE_SNAKE_CASE, fname)
+        if not res:
+            reportError(var.typeStartToken, 'style',
+                        f'File {arg} violates snake-case naming convention', 'varname')
+
+        for cfg in data.iterconfigurations():
+            print(f'Checking {arg}, config {cfg.name}...')
+
+            ddc_usage = [tok for tok in cfg.tokenlist if tok.str in ddc_keyword_map]
+            for tok in ddc_usage:
+                key = tok.str
+                new_key = ddc_keyword_map[key]
+                reportError(tok, 'style',
+                            f'Please use the new naming conventions instead of DDC names ({key}->{new_key})',
+                            'DDCNames')
+
+            for var in cfg.variables:
+                if var.nameToken:
+                    res = re.match(RE_SNAKE_CASE, var.nameToken.str)
+                    if not res:
+                        reportError(var.typeStartToken, 'style',
+                                    f'Variable {var.nameToken.str} violates snake-case naming convention',
+                                    'varname')
+                if (var.access is None) or var.access != 'Private':
+                    continue
+
+                if not ((var.nameToken.str.startswith('m_') and not var.isStatic) or
+                        (var.nameToken.str.startswith('s_') and var.isStatic)):
+                    reportError(var.typeStartToken, 'style',
+                                f'Private member variable {var.nameToken.str} violates naming convention',
+                                'privateMemberVariable')
+
+            for scope in cfg.scopes:
+                if scope.type == 'Function':
+                    function = scope.function
+                    if function is not None and function.type in ('Constructor', 'Destructor', 'CopyConstructor', 'MoveConstructor'):
[…]
