### 2025-07-09 d88fedb
Correct assert and put back test
