+++ tests/testing_tools/test_math_utils.cpp
@@ -5,7 +5,7 @@
 
 double cyl_bessel_j(const double nu, const double x)
 {
-    assert(nu < 30);
+    assert(x < 30);
     double sum = 0;
     for (int k = 0; k < 100; ++k) {
         sum += std::pow(-1, k) * std::pow(x / 2, 2 * k + nu)
