+++ tests/gyroaverage/gyroaverage_circular.cpp
@@ -306,90 +306,87 @@ TEST_P(GyroAverageCircularParamTests, TestPeriodicity)
     });
 }
 
-// This test needs tuning to represent realistic physical parameters.
-// In its current state it requires Bessel functions evaluated using arguments
-// for which the function is not very stable.
-//TEST_P(GyroAverageCircularParamTests, TestAnalytical)
-//{
-//    DConstFieldRThetaBatch A = get_const_field(m_A_alloc);
-//    DFieldRThetaBatch A_bar = get_field(m_A_bar_alloc);
-//
-//    using SplineRThetaBuilder = SplineRThetaBuilderType<Kokkos::DefaultExecutionSpace>;
-//    using SplineRThetaEvaluatorNullBound
-//            = SplineRThetaEvaluatorNullBoundType<Kokkos::DefaultExecutionSpace>;
-//
-//    IdxRangeRThetaBatch const rthetabatch_idx_range = get_idx_range(A_bar);
-//    IdxRangeRTheta const rtheta_idx_range(rthetabatch_idx_range);
-//    ddc::NullExtrapolationRule r_extrapolation_rule;
-//    ddc::PeriodicExtrapolationRule<Theta> theta_extrapolation_rule;
-//    SplineRThetaBuilder const spline_builder(rtheta_idx_range);
-//    SplineRThetaEvaluatorNullBound const spline_evaluator(
-//            r_extrapolation_rule,
-//            r_extrapolation_rule,
-//            theta_extrapolation_rule,
-//            theta_extrapolation_rule);
-//
-//    using GyroAverageOperatorType = GyroAverageOperator<
-//            SplineRThetaBuilder,
-//            SplineRThetaEvaluatorNullBound,
-//            IdxRangeRThetaBatch,
-//            CartesianToPolar>;
-//    GyroAverageOperatorType gyroaverage(
-//            get_const_field(m_rho_L_alloc),
-//            spline_builder,
-//            spline_evaluator,
-//            CartesianToPolar(),
-//            m_nb_gyro_points);
-//    gyroaverage(A_bar, A);
-//
-//    auto Rcoord_alloc_host = ddc::create_mirror_and_copy(get_field(m_Rcoord_alloc));
-//    host_t<DFieldRTheta> Rcoord_host = get_field(Rcoord_alloc_host);
-//    auto Zcoord_alloc_host = ddc::create_mirror_and_copy(get_field(m_Zcoord_alloc));
-//    host_t<DFieldRTheta> Zcoord_host = get_field(Zcoord_alloc_host);
-//    auto rho_L_alloc_host = ddc::create_mirror_and_copy(get_field(m_rho_L_alloc));
-//    host_t<DFieldRTheta> rho_L_host = get_field(rho_L_alloc_host);
-//    auto A_bar_alloc_host = ddc::create_mirror_and_copy(A_bar);
-//    host_t<DFieldRThetaBatch> A_bar_host = get_field(A_bar_alloc_host);
-//
-//    IdxRangeTheta const theta_idx_range(rthetabatch_idx_range);
-//
-//    double const r0 = m_r0;
-//    double const R0 = m_R0;
-//    double const kperp = m_kperp;
-//    double const kx = m_kx;
-//    double const ky = m_ky;
-//    int const nb_gyro_points = m_nb_gyro_points;
-//    ddc::for_each(rthetabatch_idx_range, [&](IdxRThetaBatch const irthetabatch) {
-//        IdxR const ir(irthetabatch);
-//        IdxTheta const itheta(irthetabatch);
-//        IdxBatch const ibatch(irthetabatch);
-//        double const r = ddc::coordinate(ir);
-//        double const R0eff = R0;
-//        double const gyroradius = rho_L_host(ir, itheta);
-//        double const kperprho = kperp * gyroradius;
[…]
