### 2025-07-09 501915e
Strip overenthusiastic -I
