### 2025-07-10 b84d3db
Use a DField to ensure correct bounds checking (fix bad memory usage) and to facilitate logic.
