+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -112,11 +112,18 @@ private:
     /// The type of an index range over the polar B-splines
     using IdxRangeBSPolar = IdxRange<PolarBSplinesRTheta>;
     using IdxBSPolar = Idx<PolarBSplinesRTheta>;
+    using IdxStepBSPolar = IdxStep<PolarBSplinesRTheta>;
 
     using IdxRangeBSR = IdxRange<BSplinesR>;
     using IdxRangeBSTheta = IdxRange<BSplinesTheta>;
     using IdxRangeBSRTheta = IdxRange<BSplinesR, BSplinesTheta>;
+
+    using IdxBSR = Idx<BSplinesR>;
+    using IdxBSTheta = Idx<BSplinesTheta>;
     using IdxBSRTheta = Idx<BSplinesR, BSplinesTheta>;
+
+    using IdxStepBSR = IdxStep<BSplinesR>;
+    using IdxStepBSTheta = IdxStep<BSplinesTheta>;
 
     using IdxRangeBatchedBSRTheta
             = ddc::detail::convert_type_seq_to_discrete_domain_t<ddc::type_seq_replace_t<
@@ -289,11 +296,11 @@ public:
         , m_matrix_size(ddc::discrete_space<PolarBSplinesRTheta>().nbasis() - m_nbasis_theta)
         , m_idxrange_fem_non_singular(
                   ddc::discrete_space<PolarBSplinesRTheta>().tensor_bspline_idx_range().remove_last(
-                          IdxStep<PolarBSplinesRTheta> {m_nbasis_theta}))
+                          IdxStepBSPolar {m_nbasis_theta}))
         , m_idxrange_bsplines_r(ddc::discrete_space<BSplinesR>().full_domain().remove_first(
-                  IdxStep<BSplinesR> {m_n_overlap_cells}))
+                  IdxStepBSR {m_n_overlap_cells}))
         , m_idxrange_bsplines_theta(ddc::discrete_space<BSplinesTheta>().full_domain().take_first(
-                  IdxStep<BSplinesTheta> {m_nbasis_theta}))
+                  IdxStepBSTheta {m_nbasis_theta}))
         , m_idxrange_quadrature_r(
                   Idx<QDimRMesh>(0),
                   IdxStep<QDimRMesh>(
@@ -1494,62 +1501,103 @@ public:
      * @brief Fills the nnz data structure by computing the number of non-zero per line.
      * This number is linked to the weak formulation and depends on @f$(r,\theta)@f$ splines.
      * After this function the array will contain:
-     * nnz[0] = 0.
-     * nnz[1] = 0.
-     * nnz[2] = number of non-zero elements in line 0.
-     * nnz[3] = number of non-zero elements in lines 0-1.
-     * ...
-     * nnz[matrix_size] = number of non-zero elements in lines 0-(matrix_size-1).
+     * nnz_per_row[0] = 0.
+     * nnz_per_row[1] = 0.
+     * nnz_per_row[2] = number of non-zero elements in line 0.
+     * nnz_per_row[3] = number of non-zero elements in lines 0-1.
+     * ..._per_row
+     * nnz_per_row[matrix_size] = number of non-zero elements in lines 0-(matrix_size-1).
      *
-     * @param[out]  nnz A 1D Kokkos view of length matrix_size+1 which stores the count of the non-zeros along the lines of the matrix.
+     * @param[out]  nnz_per_row A 1D Kokkos view of length matrix_size+1 which stores the sum of the non-zeros in the matrix on all lines up
+     *                          to the one in.
      */
-    void init_nnz_per_line(Kokkos::View<int*, Kokkos::LayoutRight> nnz) const
[…]
