### 2025-07-10 aca3273
Reduced the computational cost of the gyroaveraging test (and fixed the phase used in the analystical formula)
