### 2025-07-10 ea70dd5
Remove sum loop
