+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -1519,6 +1519,9 @@ public:
         IdxRangeBSPolar polar_bspl_idx_range
                 = ddc::discrete_space<PolarBSplinesRTheta>().full_domain().remove_last(
                         radial_boundary_splines);
+
+        Kokkos::deep_copy(nnz_per_row, 0.);
+
         assert(mat_size == polar_bspl_idx_range.size());
         // nnz per line
         Field<int, IdxRangeBSPolar>
@@ -1534,50 +1537,77 @@ public:
         // left or to its right
         IdxStepBSR n_one_side_overlap(BSplinesR::degree());
 
+        int nnz_for_sigular_rows = n_singular_basis + degree * nbasis_theta_proxy;
+
         // rows representing the bsplines which cover the singular domain
         // These overlap with other singular domain splines and degree radial splines
-        ddc::parallel_for_each(
+        Kokkos::parallel_for(
                 "overlap singular radial",
-                Kokkos::DefaultExecutionSpace(),
-                PolarBSplinesRTheta::template singular_idx_range<PolarBSplinesRTheta>(),
-                KOKKOS_LAMBDA(IdxBSPolar const k) {
-                    nnz(k) = n_singular_basis + degree * nbasis_theta_proxy;
-                });
+                Kokkos::RangePolicy<Kokkos::DefaultExecutionSpace>(1, n_singular_basis + 1),
+                KOKKOS_LAMBDA(const int k) { nnz_per_row(k + 1) = k * nnz_for_sigular_rows; });
+
+        int nnz_sum = nnz_for_sigular_rows * n_singular_basis;
 
         IdxRangeBSR idxrange_bsplines_r_overlapping_singular
                 = m_idxrange_bsplines_r.take_first(n_one_side_overlap);
         IdxRangeBSTheta idxrange_bsplines_theta = m_idxrange_bsplines_theta;
+        IdxRangeBSRTheta idxrange_singular_overlap(
+                idxrange_bsplines_r_overlapping_singular,
+                idxrange_bsplines_theta);
         // rows representing the bsplines which can be represented as a tensor product of
         // radial and poloidal bsplines but which still overlap with the singular domain
-        ddc::parallel_for_each(
+        Kokkos::parallel_for(
                 "inner overlap",
-                Kokkos::DefaultExecutionSpace(),
-                IdxRangeBSRTheta(idxrange_bsplines_r_overlapping_singular, idxrange_bsplines_theta),
-                KOKKOS_LAMBDA(IdxBSRTheta const k_2d) {
-                    IdxBSR k_r(k_2d);
-                    IdxBSPolar k(PolarBSplinesRTheta::template get_polar_index<PolarBSplinesRTheta>(
-                            k_2d));
-                    int nbasis_overlap_in_singular_region = n_singular_basis;
-                    int nsupport_cells_outside_singular_region
-                            = (k_r - idxrange_bsplines_r_overlapping_singular.front()) + 1;
-                    int nradial_basis_overlap_outside_singular_region
-                            = degree + nsupport_cells_outside_singular_region;
-                    nnz(k) = nbasis_overlap_in_singular_region
-                             + nradial_basis_overlap_outside_singular_region * stencil_overlap;
+                Kokkos::TeamPolicy<>(
+                        Kokkos::DefaultExecutionSpace(),
+                        idxrange_bsplines_r_overlapping_singular.size(),
+                        Kokkos::AUTO),
+                KOKKOS_LAMBDA(const Kokkos::TeamPolicy<>::member_type& team) {
+                    IdxStepBSR ir(team.league_rank());
+                    // number of radial bases which overlap with this element
+                    int nr = degree + ir + 1;
+
[…]
