### 2025-07-10 1c0ea96
Work on parallel on a SingleInterfaceDerivativesCalculatorCollection operator.
