### 2025-07-10 fa5a1b7
Emily's trick
