### 2025-07-10 e4022c3
Improve docstring
