### 2025-07-10 178c733
Update static analysis to handle multi-level parallelism
