+++ bin/ci_tools/gyselalib_static_analysis.py
@@ -549,10 +549,15 @@ def check_kokkos_lambda_use(file):
                     else:
                         i+=1
 
-                if config.data[arg_splitters[-1]+1]['str'] == '[':
+                is_thread_parallelism = any('TeamThreadRange' == a['str'] for a in config.data[idx+1:arg_splitters[-1]])
+                if config.data[arg_splitters[-1]+1]['str'] == '[' and not is_thread_parallelism:
                     last_arg = ' '.join(a['str'] for a in config.data[arg_splitters[-1]+1:end_idx])
                     correct_last_arg = 'KOKKOS_LAMBDA '+last_arg[last_arg.find(']')+1:]
                     report_error(FATAL, file, elem.attrib['linenr'], f'The lambda function passed to the function ddc::{p} must be a KOKKOS_LAMBDA\n{last_arg}\nShould be:\n{correct_last_arg}')
+                elif config.data[arg_splitters[-1]+1]['str'].startswith('KOKKOS') and is_thread_parallelism:
+                    last_arg = ' '.join(a['str'] for a in config.data[arg_splitters[-1]+1:end_idx])
+                    correct_last_arg = '[&]'+last_arg[last_arg.find(']')+1:]
+                    report_error(FATAL, file, elem.attrib['linenr'], f'The lambda function passed as thread parallelism to the function ddc::{p} must be a KOKKOS_LAMBDA\n{last_arg}\nShould be:\n{correct_last_arg}')
 
                 lambda_body_start = next(i for i,a in enumerate(config.data[arg_splitters[-1]+1:], arg_splitters[-1]+1) if a['str'] == '{')
                 lambda_body_end_id = config.data[lambda_body_start]['link']
