### 2025-07-10 #8304 @EmilyBourne
> [@EmilyBourne](https://github.com/EmilyBourne) The submission looks good to me. I'll start the process of identifying reviewers.

@piyueh Ok sounds great. In case you missed it I already posted my suggestions for potential reviewers above:
https://github.com/openjournals/joss-reviews/issues/8304#issuecomment-3003941120
