### 2025-07-11 99b307c
Don't use get_idx_range on GPU
