### 2025-07-11 0570ec7
Work round syntaxError in cppcheck
