### 2025-07-11 cdb74c2
Remove filter to catch .hpp files
