### 2025-07-11 4d70404
Add check_for_memory_misuse addon
