+++ bin/ci_tools/check_for_memory_misuse.py
@@ -1,0 +1,161 @@
+import argparse
+import glob
+from pathlib import Path
+import sys
+
+import cppcheckdata
+
+def reportError(token, msg, errorId, severity='error'):
+    cppcheckdata.reportError(token, 'error', msg, 'memory', errorId)
+    if token.file == file:
+        cppcheckdata.reportError(token, 'error', msg, 'memory', errorId)
+
+def main():
+    parser = cppcheckdata.ArgumentParser()
+    args = parser.parse_args()
+
+    dump_files, ctu_info_files = cppcheckdata.get_files(args)
+
+    data_dump = {f.removesuffix('.dump'): cppcheckdata.parsedump(f) for f in dump_files}
+
+    # Detect execution spaces
+    for f, data in data_dump.items():
+        global file
+        file = Path(f)
+        while len(file.suffixes) > 1:
+            file = file.with_suffix('')
+        file = str(file)
+
+        for cfg in data.iterconfigurations():
+            # clang_warnings', 'containers', 'directives', 'functions', 'macro_usage', 'name', 'preprocessor_if_conditions', 'scopes', 'setIdMap', 'set_id_map', 'set_tokens_links', 'standards', 'tokenlist', 'typedefInfo', 'valueflow', 'variables'
+
+            class_scope = None
+
+            # Find execution space for each scope
+            for scope in cfg.scopes:
+                if scope.type not in ('Function', 'Unconditional'):
+                    continue
+                if scope.nestedIn.type == 'Class':
+                    class_scope = scope.nestedIn
+                tok_idx = next(i for i, tok in enumerate(cfg.tokenlist) if tok.Id == scope.bodyStartId)
+                start_idx = next((tok_idx-i for i, tok in enumerate(reversed(cfg.tokenlist[:tok_idx]),1) if tok.str in ('{', '}', ';')), 0)
+                exec_type = next((tok.str for tok in cfg.tokenlist[start_idx:tok_idx] if 'KOKKOS_' in tok.str), 'CPU')
+                scope.exec_type = exec_type
+                scope.exec_space = 'CPU' if exec_type == 'CPU' else 'GPU'
+
+            for scope in cfg.scopes:
+                if scope.type not in ('Class', 'Struct'):
+                    continue
+
+                tok = next(tok for tok in cfg.tokenlist if tok.Id == scope.bodyStartId)
+
+                function_scopes = [s for s in scope.nestedList if s.type == 'Function']
+                functions = [s.function for s in function_scopes]
+                has_gpu_constructor = any(s.exec_space != 'CPU' for f,s in zip(functions, function_scopes) \
+                                        if f.type == 'Constructor')
+
+                nested_scopes = scope.nestedList
+                if any(getattr(s, 'exec_space', None) == 'GPU' for s in nested_scopes) and \
+                        any(v.isReference for v in scope.varlist) and \
+                        not has_gpu_constructor:
+                    reportError(tok,
+                        f'Class {scope.className} contains CPU references to objects.',
+                        'classExecSpace')
+
+                for var in scope.varlist:
+                    tok_start_idx = next(i for i, tok in enumerate(cfg.tokenlist) if tok.Id == var.typeStartTokenId)
+                    tok_end_idx = next(i for i, tok in enumerate(cfg.tokenlist) if tok.Id == var.typeEndTokenId) + 1
+                    type_descr = [tok.str for tok in cfg.tokenlist[tok_start_idx:tok_end_idx]]
+                    if any(t in type_descr for t in ('FieldMem', 'VectorFieldMem', 'Field', 'VectorField', 'DiscreteToCartesian')):
+                        if (s in type_descr for s in ('host_t', 'DefaultHostExecutionSpace', 'HostSpace')):
+                            var.mem_space = 'CPU'
[…]
