### 2025-07-11 a756368
Only keep Python executable
