+++ bin/ci_tools/check_naming_conventions.py
@@ -1,9 +1,12 @@
-import argparse
+"""
+Check that the naming conventions documented in CODING_STANDARD.md are respected.
+This file should be run using ./bin/run_cppcheck
+"""
 from pathlib import Path
 import re
 import sys
 
-import cppcheckdata
+import cppcheckdata #pylint: disable=import-error
 
 ddc_keyword_map = {'DiscreteElement': 'Idx',
                    'DiscreteVector' : 'IdxStep',
@@ -28,20 +31,16 @@ file_exceptions = [HOME_DIR / 'src' / 'utils' / 'ddc_helper.hpp',
 RE_SNAKE_CASE = re.compile("[a-z0-9]+(?:_[a-z0-9]+)*")
 RE_CAMEL_CASE = re.compile("([A-Z][a-z0-9]*)+")
 
-class FileErrInfo:
-    def __init__(self, fname):
-        self.file = fname
-        self.linenr = 0
-        self.column = None
-
 def reportError(token, msg, errorId):
+    """ Report an incorrect naming convention.
+    """
     cppcheckdata.reportError(token, 'style', msg, 'naming', errorId)
 
 def main():
     parser = cppcheckdata.ArgumentParser()
     args = parser.parse_args()
 
-    dump_files, ctu_info_files = cppcheckdata.get_files(args)
+    dump_files, _ = cppcheckdata.get_files(args)
 
     for arg in dump_files:
         if not arg.endswith('.dump') or any(f'{f}.dump' == arg for f in file_exceptions):
@@ -56,6 +55,8 @@ def main():
         #    filename = arg.removesuffix(".dump")
         #    reportError(token,
         #                f'File {filename} violates snake-case naming convention', 'varname')
+
+        # Find any use of the DDC keywords listed in ddc_keyword_map in the raw code.
         ddc_usage = [tok for tok in data.rawTokens if tok.str in ddc_keyword_map]
         for tok in ddc_usage:
             key = tok.str
@@ -64,47 +65,64 @@ def main():
                         f'Please use the new naming conventions instead of DDC names ({key}->{new_key})',
                         'DDCNames')
 
+        # Loop over configurations. A configuration is created by an ifdef
         for cfg in data.iterconfigurations():
             print(f'Checking {arg}, config {cfg.name}...')
 
-
+            # Check that variables respect naming conventions
             for var in cfg.variables:
+                # If anonymous variable then move on to next element
                 if var.nameToken is None:
                     continue
-                if var.nameToken:
-                    res = re.match(RE_SNAKE_CASE, var.nameToken.str)
-                    if not res:
-                        reportError(var.typeStartToken,
-                                    f'Variable {var.nameToken.str} violates snake-case naming convention',
-                                    'varname')
-                if (var.access is None) or var.access != 'Private':
-                    continue
 
-                if not ((var.nameToken.str.startswith('m_') and not var.isStatic) or
-                        (var.nameToken.str.startswith('s_') and var.isStatic)):
+                # Check that variables use snake case by checking against a regex
+                res = re.match(RE_SNAKE_CASE, var.nameToken.str)
+                if not res:
                     reportError(var.typeStartToken,
-                                f'Private member variable {var.nameToken.str} violates naming convention',
-                                'privateMemberVariable')
+                                f'Variable {var.nameToken.str} violates snake-case naming convention',
+                                'varname')
 
+                # For class variables check that member variables start with m_
+                # and that static variables start with s_
+                if var.access in ('Private', 'Protected', 'Public'):
+                    name = var.nameToken.str
+                    if not ((name.startswith('m_') and not var.isStatic) or
+                            (name.startswith('s_') and var.isStatic)):
+                        if var.isStatic:
+                            expected = 's_' + name
+                        else:
[…]
