### 2025-07-14 #8582 @piyueh
### Here are some notes about this JOSS submission:

1. The `paper.md` is in the `joss` branch (https://github.com/gyselax/gyselalibxx/tree/joss)
2. The source code to be reviewed is tagged `v0.2.0` (https://github.com/gyselax/gyselalibxx/tree/v0.2.0)
3. If there are any code changes requested by reviewers, those changes should be made on top of `v0.2.0` as this JOSS paper is only for this particular version.

### Below are general information for reviewers:

As a reviewer, the first step is to create a checklist for your review by entering `@editorialbot generate my checklist` as a new comment in this thread.

These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines. Or you can find them here:
* https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
* https://joss.readthedocs.io/en/latest/review_criteria.html
* https://joss.readthedocs.io/en/latest/review_checklist.html

We aim for reviews to be completed within about 4-6 weeks. Please let me know if any of you require some more time. We can also use `@editorialbot` to set automatic reminders if you know you'll be away for a known period of time.
