### 2025-07-16 b1b88b9
Fix index calculation
