+++ CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
+## [v0.2.0] - 2025-07-16
+
+### Fixed
+
+-Modify ruche.v100/environment.sh file to fix tokamaxi simulation segfault issues.
+
 ## [v0.2.0] - 2025-07-03
 
 ### Added
