### 2025-07-16 e4fd420
fix typo in changelog
