+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Modify ruche.v100/environment.sh file to fix tokamaxi simulation segfault issues.
+- Modify `ruche.v100/environment.sh` file to fix tokamaxi simulation segfault issues.
   
 ### Changed
 
