### 2025-07-16 bb05c9c
Fix Ruche modules (#377)
The aim of this PR is to fix an environment issue encountered during
Gysela-X tokamaxi simulations on Ruche. Indeed, running with more than
one mpi process was leading to a segmentation fault because the version
of MPI that was used was not GPU aware.
The swap command is used to ensure that the correct version of cuda and
openmpi-cuda are used, and to avoid relying on automatically loaded
versions of these modules.
