+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Modify `ruche.v100/environment.sh` file to fix tokamaxi simulation segfault issues.
+  
 ### Changed
 
 ### Deprecated
