### 2025-07-16 e682f6a
Clarify docstring
Co-authored-by: Pauline Vidal  <142578924+PaulineVidal@users.noreply.github.com>
