### 2025-07-16 3cc1933
Fix `DerivField` definition (#370)
As noted while trying to use `DerivField` with a `SplineBuilder` (see
#360). The `DerivField` has a design flaw as it stores the derivatives
as `ddc::Deriv<GridX>` instead of `ddc::Deriv<X>`. This PR fixes this
issue.

-----

Co-authored-by: Pauline Vidal <142578924+PaulineVidal@users.noreply.github.com>
