+++ assets/.doxy/gyselalibxx/gyselalibxx/classDerivField_3_01ElementType_00_01IdxRange_3_01DDims_8_8_8_01_4_00_01MemorySpace_00_01LayoutStridedPolicy_01_4.md
@@ -171,7 +171,7 @@ The values of the field and the derivatives may be defined on different index ra
 
 
 * `ElementType` The type of the elements inside the chunks. 
-* `IdxRange<DDims...>` The index range on which the internal fields are defined. This index range is the physical index range on which the values are defined combined with the index range of the derivatives of interest (e.g. IdxRange&lt;Deriv&lt;IDimX&gt;, IDimX, IDimY&gt;). 
+* `IdxRange<DDims...>` The index range on which the internal fields are defined. This index range is the physical index range on which the values are defined combined with the index range of the derivatives of interest (e.g. IdxRange&lt;Deriv&lt;X&gt;, [**GridX**](structGridX.md), [**GridY**](structGridY.md)&gt;). 
 * `MemorySpace` The memory space where the data is saved (CPU/GPU). 
 * `LayoutStridedPolicy` The way in which the memory is laid out in memory (contiguous in the leading/trailing dimension, strided, etc). 
 
