+++ assets/.doxy/gyselalibxx/gyselalibxx/type__seq__tools_8hpp_source.md
@@ -74,6 +74,15 @@ struct FindGrid<Dim, ddc::detail::TypeSeq<>>
     using type = void;
 };
 
+template <class TypeSeqDims, class TypeSeqGrids>
+struct FindAllGrids;
+
+template <class... TypeSeqDims, class TypeSeqGrids>
+struct FindAllGrids<ddc::detail::TypeSeq<TypeSeqDims...>, TypeSeqGrids>
+{
+    using type = ddc::detail::TypeSeq<typename FindGrid<TypeSeqDims, TypeSeqGrids>::type...>;
+};
+
 template <class CoordType, class IdxRangeType>
 struct FindIdxType;
 
@@ -203,6 +212,9 @@ using type_seq_cat_t = typename detail::TypeSeqCat<TypeSeqs...>::type;
 
 template <class Dim, class TypeSeqGrid>
 using find_grid_t = typename detail::FindGrid<Dim, TypeSeqGrid>::type;
+
+template <class TypeSeqDim, class TypeSeqGrid>
+using find_all_grids_t = typename detail::FindAllGrids<TypeSeqDim, TypeSeqGrid>::type;
 
 template <class StartTypeSeq>
 using type_seq_unique_t = typename detail::GetUnique<StartTypeSeq>::type;
