### 2025-07-16 270a3ef
Reduced the number of points checks; fixed an error in the formula
