+++ tests/gyroaverage/gyroaverage_circular.cpp
@@ -239,7 +239,7 @@ protected:
 
 protected:
     double m_gyroradius;
-    std::size_t const m_nb_gyro_points = 32;
+    std::size_t const m_nb_gyro_points = 6; // Must be odd for the error estimate below to apply
     double const m_B0 = 1.0;
     double const m_R0 = 300.0;
     double const m_r0 = 150.0;
@@ -348,7 +348,19 @@ TEST_P(GyroAverageCircularParamTests, TestAnalytical)
     auto A_bar_alloc_host = ddc::create_mirror_and_copy(A_bar);
     host_t<DFieldRThetaBatch> A_bar_host = get_field(A_bar_alloc_host);
 
+    IdxRangeR const r_idx_range(rthetabatch_idx_range);
     IdxRangeTheta const theta_idx_range(rthetabatch_idx_range);
+    IdxRangeBatch const batch_idx_range(rthetabatch_idx_range);
+
+    IdxStep<GridR, GridTheta, GridBatch> rthetabatch_check_strides(5, 11, 1);
+    IdxStep<GridR, GridTheta, GridBatch> rthetabatch_check_extents(
+            r_idx_range.size() / 5,
+            theta_idx_range.size() / 11,
+            batch_idx_range.size());
+    IdxRangeSlice<GridR, GridTheta, GridBatch> rthetabatch_idx_range_check(
+            rthetabatch_idx_range.front(),
+            rthetabatch_check_extents,
+            rthetabatch_check_strides);
 
     double const r0 = m_r0;
     double const R0 = m_R0;
@@ -356,7 +368,7 @@ TEST_P(GyroAverageCircularParamTests, TestAnalytical)
     double const kx = m_kx;
     double const ky = m_ky;
     int const nb_gyro_points = m_nb_gyro_points;
-    ddc::for_each(rthetabatch_idx_range, [&](IdxRThetaBatch const irthetabatch) {
+    ddc::for_each(rthetabatch_idx_range_check, [&](IdxRThetaBatch const irthetabatch) {
         IdxR const ir(irthetabatch);
         IdxTheta const itheta(irthetabatch);
         IdxBatch const ibatch(irthetabatch);
@@ -364,26 +376,25 @@ TEST_P(GyroAverageCircularParamTests, TestAnalytical)
         // In the outer region, the particle position can be out of small radius (r), where the values are considered to be zero.
         // Also the gyroaveraged value has the periodicity along theta direction (tested separately above)
         // These conditions are not taken into account in the analytical formula, so we do not test these cases.
-        if (ir % 5 == 0 && itheta % 11 == 0 &&
-            r < r0 * 0.8 && itheta < theta_idx_range.back()) {
+        if (r < r0 * 0.8) {
             double const R0eff = R0;
             double const gyroradius = rho_L_host(ir, itheta);
             double const kperprho = kperp * gyroradius;
-            double const phase = kx * (Rcoord_host(ir, itheta) - R0eff) + ky * Zcoord_host(ir, itheta);
-            double const phik = std::atan2(Zcoord_host(ir, itheta), Rcoord_host(ir, itheta) - R0eff);
-            std::size_t nb_gyro_points_half = m_nb_gyro_points / 2;
-            double const sgn = std::pow(-1.0, static_cast<double>(nb_gyro_points_half));
-            double const sgn_tmp = nb_gyro_points % 2 == 0 ? sgn : -sgn;
+            double const phase
+                    = kx * (Rcoord_host(ir, itheta) - R0eff) + ky * Zcoord_host(ir, itheta);
             double const res_th = cyl_bessel_j(0, kperprho) * std::cos(phase);
-            double const err_J0_th
-                    = 2.0
[…]
