### 2025-07-16 540ad04
Mem object shouldn't be passed as argument
