### 2025-07-16 3a577cd
Improve `PolarSplineFEMPoissonLikeSolver::init_nnz_per_line` (#373)
The method `PolarSplineFEMPoissonLikeSolver::init_nnz_per_line` fills a
matrix containing the number of non-zero elements in the previous row
(see
[docstring](https://gyselax.github.io/gyselalibxx/gyselalibxx/classPolarSplineFEMPoissonLikeSolver.html#function-init_nnz_per_line)).
The previous implementation was problematic for the following reasons:
- The final loop filled an element which did not exist in the table
- Kokkos views were used where a DDC `Field` could be used (using a
`Field` instead of a `View` leads to a crash
[…]
