+++ assets/.doxy/gyselalibxx/gyselalibxx/polarpoissonlikesolver_8hpp_source.md
@@ -84,11 +84,18 @@ private:
 
     using IdxRangeBSPolar = IdxRange<PolarBSplinesRTheta>;
     using IdxBSPolar = Idx<PolarBSplinesRTheta>;
+    using IdxStepBSPolar = IdxStep<PolarBSplinesRTheta>;
 
     using IdxRangeBSR = IdxRange<BSplinesR>;
     using IdxRangeBSTheta = IdxRange<BSplinesTheta>;
     using IdxRangeBSRTheta = IdxRange<BSplinesR, BSplinesTheta>;
+
+    using IdxBSR = Idx<BSplinesR>;
+    using IdxBSTheta = Idx<BSplinesTheta>;
     using IdxBSRTheta = Idx<BSplinesR, BSplinesTheta>;
+
+    using IdxStepBSR = IdxStep<BSplinesR>;
+    using IdxStepBSTheta = IdxStep<BSplinesTheta>;
 
     using IdxRangeBatchedBSRTheta
             = ddc::detail::convert_type_seq_to_discrete_domain_t<ddc::type_seq_replace_t<
@@ -199,11 +206,11 @@ public:
         , m_matrix_size(ddc::discrete_space<PolarBSplinesRTheta>().nbasis() - m_nbasis_theta)
         , m_idxrange_fem_non_singular(
                   ddc::discrete_space<PolarBSplinesRTheta>().tensor_bspline_idx_range().remove_last(
-                          IdxStep<PolarBSplinesRTheta> {m_nbasis_theta}))
+                          IdxStepBSPolar {m_nbasis_theta}))
         , m_idxrange_bsplines_r(ddc::discrete_space<BSplinesR>().full_domain().remove_first(
-                  IdxStep<BSplinesR> {m_n_overlap_cells}))
+                  IdxStepBSR {m_n_overlap_cells}))
         , m_idxrange_bsplines_theta(ddc::discrete_space<BSplinesTheta>().full_domain().take_first(
-                  IdxStep<BSplinesTheta> {m_nbasis_theta}))
+                  IdxStepBSTheta {m_nbasis_theta}))
         , m_idxrange_quadrature_r(
                   Idx<QDimRMesh>(0),
                   IdxStep<QDimRMesh>(
@@ -1185,66 +1192,142 @@ public:
         return idx_theta;
     }
 
-    void init_nnz_per_line(Kokkos::View<int*, Kokkos::LayoutRight> nnz) const
+    void init_nnz_per_line(Kokkos::View<int*, Kokkos::LayoutRight> nnz_per_row) const
     {
         Kokkos::Profiling::pushRegion("PolarPoissonInitNnz");
-        size_t const mat_size = nnz.extent(0) - 1;
+        size_t const mat_size = nnz_per_row.extent(0) - 1;
+        IdxStepBSPolar radial_boundary_splines(m_nbasis_theta);
+        IdxRangeBSPolar polar_bspl_idx_range
+                = ddc::discrete_space<PolarBSplinesRTheta>().full_domain().remove_last(
+                        radial_boundary_splines);
+
+        Kokkos::deep_copy(nnz_per_row, 0.);
+
+        assert(mat_size == polar_bspl_idx_range.size());
+        // nnz per line
+        Field<int, IdxRangeBSPolar>
+                nnz(Kokkos::subview(nnz_per_row, std::pair<int, int>(2, mat_size + 1)),
+                    polar_bspl_idx_range.remove_last(IdxStepBSPolar(1)));
+
         size_t constexpr n_singular_basis = PolarBSplinesRTheta::n_singular_basis();
         size_t constexpr degree = BSplinesR::degree();
-        size_t constexpr radial_overlap = 2 * degree + 1;
+        size_t constexpr stencil_overlap = 2 * degree + 1;
[…]
