### 2025-07-16 e2822cc
Add missing type declaration
