### 2025-07-16 f4a34db
Fixed an odd comment about an even number
