### 2025-07-16 ffe4519
Use hand-written cyl_bessel_j function (#369)
Use a hand-written cyl_bessel_j function to work around #367 and unblock
#336 and #361. Fixes #367

---------

Co-authored-by: Kevin Obrejan <kevin.obrejan@cea.fr>
