+++ src/geometryRTheta/geometry/geometry.hpp
@@ -326,8 +326,8 @@ using ConstSpline2D = DConstField<IdxRangeBSRTheta>;
  *
  * Store the polar B-splines coefficients of the function.
  */
-using PolarSplineMemRTheta = PolarSplineMem<PolarBSplinesRTheta>;
-using PolarSplineRTheta = PolarSpline<PolarBSplinesRTheta>;
+using PolarSplineMemRTheta = DFieldMem<IdxRange<PolarBSplinesRTheta>>;
+using PolarSplineRTheta = DField<IdxRange<PolarBSplinesRTheta>>;
 
 /**
  * @brief Type of the index of an element of polar B-splines.
