+++ src/geometryRTheta/time_solver/bsl_predcorr.hpp
@@ -131,8 +131,7 @@ public:
         IdxRangeBSTheta polar_idx_range(ddc::discrete_space<BSplinesTheta>().full_domain());
 
         host_t<PolarSplineMemRTheta> electrostatic_potential_coef(
-                PolarBSplinesRTheta::singular_idx_range<PolarBSplinesRTheta>(),
-                IdxRangeBSRTheta(radial_bsplines, polar_idx_range));
+                ddc::discrete_space<PolarBSplinesRTheta>().full_domain());
         ddc::NullExtrapolationRule extrapolation_rule;
         PolarSplineEvaluator<PolarBSplinesRTheta, ddc::NullExtrapolationRule>
                 polar_spline_evaluator(extrapolation_rule);
@@ -159,10 +158,10 @@ public:
                     // --- compute electrostatic potential:
                     host_t<Spline2DMem> density_coef(get_spline_idx_range(m_builder));
                     m_builder(get_field(density_coef), get_const_field(density_host));
-                    m_poisson_solver(charge_density_coord, electrostatic_potential_coef);
+                    m_poisson_solver(charge_density_coord, get_field(electrostatic_potential_coef));
 
                     // --- compute advection field:
-                    advection_field_computer(electrostatic_potential_coef, advection_field_host);
+                    advection_field_computer(get_field(electrostatic_potential_coef), advection_field_host);
                 };
 
         std::function<void(host_t<DFieldRTheta>, host_t<DConstVectorFieldRTheta<X, Y>>, double)>
