### 2025-07-16 bf34246
Use Field for PolarSplineEvaluator
