+++ src/interpolation/polar_splines/polar_spline_evaluator.hpp
@@ -142,7 +142,7 @@ public:
      */
     double operator()(
             Coord<DimR, DimTheta> coord_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         return eval(coord_eval, spline_coef);
     }
@@ -161,19 +161,7 @@ public:
     void operator()(
             DField<Domain, Kokkos::HostSpace> const spline_eval,
             ConstField<Coord<DimR, DimTheta>, Domain, Kokkos::HostSpace> const coords_eval,
-            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef_field) const
-    {
-        host_t<ConstPolarSpline<PolarBSplinesType>> spline_coef(spline_coef_field);
-        using IdxEval = typename Domain::discrete_element_type;
-        ddc::for_each(get_idx_range(coords_eval), [=](IdxEval i) {
-            spline_eval(i) = eval(coords_eval(i), spline_coef);
-        });
-    }
-    template <class Domain>
-    void operator()(
-            DField<Domain, Kokkos::HostSpace> const spline_eval,
-            ConstField<Coord<DimR, DimTheta>, Domain, Kokkos::HostSpace> const coords_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         using IdxEval = typename Domain::discrete_element_type;
         ddc::for_each(get_idx_range(coords_eval), [=](IdxEval i) {
@@ -196,15 +184,6 @@ public:
     double deriv_dim_1(
             Coord<DimR, DimTheta> coord_eval,
             host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
-    {
-        return eval_no_bc(
-                coord_eval,
-                host_t<ConstPolarSpline<PolarBSplinesType>>(spline_coef),
-                eval_deriv_r_type());
-    }
-    double deriv_dim_1(
-            Coord<DimR, DimTheta> coord_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
     {
         return eval_no_bc(coord_eval, spline_coef, eval_deriv_r_type());
     }
@@ -225,18 +204,8 @@ public:
             Coord<DimR, DimTheta> coord_eval,
             host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
-        return eval_no_bc(
-                coord_eval,
-                host_t<ConstPolarSpline<PolarBSplinesType>>(spline_coef),
-                eval_deriv_theta_type());
-    }
-    double deriv_dim_2(
-            Coord<DimR, DimTheta> coord_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
-    {
         return eval_no_bc(coord_eval, spline_coef, eval_deriv_theta_type());
     }
-
 
     /**
      * @brief Get the value of the cross derivative of the spline function.
@@ -252,15 +221,6 @@ public:
     double deriv_1_and_2(
             Coord<DimR, DimTheta> coord_eval,
             host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
-    {
-        return eval_no_bc(
-                coord_eval,
-                host_t<ConstPolarSpline<PolarBSplinesType>>(spline_coef),
-                eval_deriv_r_theta_type());
-    }
-    double deriv_1_and_2(
[…]
