### 2025-07-16 a02e7a1
Clang formatting
