+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
+- Remove deprecated method `PolarBSplines::integrals`.
+
 ## [v0.2.0] - 2025-07-03
 
 ### Added
