### 2025-07-16 7d5b66f
Enable SYSTEM in cmake add_subdirectory (#336)
On compatible compilers, this silences warnings of files included from
these directories.
Requiers cmake 3.25:
https://cmake.org/cmake/help/v3.25/command/add_subdirectory.html
