+++ src/interpolation/polar_splines/polar_bsplines.hpp
@@ -14,6 +14,14 @@
 #include "polar_spline.hpp"
 #include "view.hpp"
 
+namespace PolarSplines {
+
+template <class ExecSpace, class DDim, class MemorySpace>
+DField<IdxRange<DDim>, MemorySpace> integrals(
+        ExecSpace const& execution_space,
+        DField<IdxRange<DDim>, MemorySpace> int_vals);
+}
+
 /**
  * A class containing all information describing polar B-splines.
  *
@@ -212,9 +220,9 @@ public:
         friend class Impl;
 
         template <class ExecSpace, class PBSpl, class OMemorySpace>
-        friend PolarSpline<PBSpl, OMemorySpace> integrals(
+        friend DField<IdxRange<PBSpl>, OMemorySpace> PolarSplines::integrals(
                 ExecSpace const& execution_space,
-                PolarSpline<PBSpl, OMemorySpace> int_vals);
+                DField<IdxRange<PBSpl>, OMemorySpace> int_vals);
 
     private:
         /**
@@ -536,15 +544,6 @@ public:
         eval_deriv_r_and_theta(DSpan1D singular_derivs, DSpan2D derivs, Coord<R, Theta> p) const;
 
         /**
-         * Calculate the integrals of each of the basis splines.
-         *
-         * @param[out] int_vals The integrals of the basis splines.
-         */
-        template <class MemorySpace2>
-        [[deprecated("Use `integrals` instead")]] void integrals(
-                PolarSpline<DDim, MemorySpace2> int_vals) const;
-
-        /**
          * Get the total number of basis functions.
          *
          * @returns The number of basis functions.
@@ -708,33 +707,25 @@ KOKKOS_FUNCTION Idx<BSplinesR, BSplinesTheta> PolarBSplines<BSplinesR, BSplinesT
     return Idx<BSplinesR, BSplinesTheta>(jmin_r, jmin_theta);
 }
 
+namespace PolarSplines {
+
 template <class ExecSpace, class DDim, class MemorySpace>
-PolarSpline<DDim, MemorySpace> integrals(
+DField<IdxRange<DDim>, MemorySpace> integrals(
         ExecSpace const& execution_space,
-        PolarSpline<DDim, MemorySpace> int_vals)
+        DField<IdxRange<DDim>, MemorySpace> int_vals)
 {
     static_assert(
             Kokkos::SpaceAccessibility<ExecSpace, MemorySpace>::accessible,
             "MemorySpace has to be accessible for ExecutionSpace.");
     using BSplinesR = typename DDim::BSplinesR_tag;
     using BSplinesTheta = typename DDim::BSplinesTheta_tag;
-    using tensor_product_idx_range_type = IdxRange<BSplinesR, BSplinesTheta>;
-    using tensor_product_idx_type = Idx<BSplinesR, BSplinesTheta>;
     using IdxR = Idx<BSplinesR>;
     using IdxTheta = Idx<BSplinesTheta>;
 
     auto r_bspl_space = ddc::discrete_space<BSplinesR>();
     auto theta_bspl_space = ddc::discrete_space<BSplinesTheta>();
 
-    assert(get_idx_range(int_vals.singular_spline_coef).extents() == DDim::n_singular_basis());
-    assert(get_idx_range(int_vals.spline_coef).front().template uid<BSplinesR>()
-           == DDim::continuity + 1);
-    assert(get_idx_range(int_vals.spline_coef).back().template uid<BSplinesR>()
-           == r_bspl_space.nbasis() - 1);
-    assert(get_idx_range(int_vals.spline_coef).template extent<BSplinesTheta>()
-                   == theta_bspl_space.nbasis()
-           || get_idx_range(int_vals.spline_coef).template extent<BSplinesTheta>()
[…]
