+++ src/interpolation/polar_splines/polar_spline.hpp
@@ -1,443 +1,0 @@
-// SPDX-License-Identifier: MIT
-#pragma once
-
-#include <ddc/ddc.hpp>
-
-#include "ddc_alias_inline_functions.hpp"
-#include "ddc_helper.hpp"
-
-template <class PolarBSplinesType, class MemSpace = Kokkos::DefaultExecutionSpace::memory_space>
-struct PolarSpline;
-
-template <class PolarBSplinesType, class MemSpace = Kokkos::DefaultExecutionSpace::memory_space>
-struct ConstPolarSpline;
-
-
-/**
- * @brief A structure containing the two FieldMems necessary to define a spline on a set of
- * polar basis splines.
- *
- * @tparam PolarBSplinesType The type of the polar B-splines on which this spline is defined.
- */
-template <class PolarBSplinesType, class MemSpace = Kokkos::DefaultExecutionSpace::memory_space>
-struct PolarSplineMem
-{
-public:
-    /// The radial bspline from which the polar B-splines are constructed.
-    using BSplinesR = typename PolarBSplinesType::BSplinesR_tag;
-    /// The poloidal bspline from which the polar B-splines are constructed.
-    using BSplinesTheta = typename PolarBSplinesType::BSplinesTheta_tag;
-
-public:
-    /**
-     * A FieldMem containing the coefficients in front of the b-spline elements which can be
-     * expressed as a tensor product of 1D B-splines.
-     */
-    DFieldMem<IdxRange<BSplinesR, BSplinesTheta>, MemSpace> spline_coef;
-
-    /**
-     * A FieldMem containing the coefficients in front of the b-spline elements near the
-     * singular point which cannot be expressed as a tensor product of 1D B-splines.
-     */
-    DFieldMem<IdxRange<PolarBSplinesType>, MemSpace> singular_spline_coef;
-
-public:
-    /**
-     * @brief A constructor for the PolarSplineMem.
-     *
-     * A constructor for the PolarSplineMem which takes a 2D domain of B-splines. This domain is then
-     * split into the domains which are relevant for a polar bspline constructed from these 2D B-splines.
-     * These new domains are then used to create the chunks which define the spline.
-     *
-     * @param domain A 2D domain of B-splines.
-     */
-    explicit PolarSplineMem<PolarBSplinesType, MemSpace>(IdxRange<BSplinesR, BSplinesTheta> domain)
-        : spline_coef(IdxRange<BSplinesR, BSplinesTheta>(
-                ddc::select<BSplinesR>(domain).remove_first(
-                        IdxStep<BSplinesR>(PolarBSplinesType::continuity + 1)),
-                ddc::select<BSplinesTheta>(domain)))
-        , singular_spline_coef(IdxRange<PolarBSplinesType>(
-                  Idx<PolarBSplinesType>(0),
-                  IdxStep<PolarBSplinesType>(PolarBSplinesType::n_singular_basis())))
-    {
-    }
-
-    /**
-     * @brief Initialise the FieldMems containing the coefficients of the spline representation.
-     *
-     * @param singular_domain The domain for the coefficients in front of the b-spline elements
-     *                        near the singular point. These are the elements which cannot be
-     *                        expressed as a tensor product of 1D B-splines.
-     * @param domain          The domain for the coefficients in front of the b-spline elements
-     *                        which can be expressed as a tensor product of 1D B-splines.
-     */
-    PolarSplineMem<PolarBSplinesType, MemSpace>(
-            IdxRange<PolarBSplinesType> singular_domain,
[…]
