### 2025-07-16 d387455
Port PolarSplineEvaluator to GPU
