### 2025-07-16 8cf2f57
Add evaluation method with coordinates calculated on the fly
