### 2025-07-16 dc6e578
Add GPU test
