+++ tests/interpolation/polar_splines.cpp
@@ -199,6 +199,125 @@ TEST(PolarSplineTest, ConstantEval)
     }
 }
 
+void test_polar_spline_eval_gpu()
+{
+    using PolarCoord = Coord<R, Theta>;
+    using CoordR = Coord<R>;
+    using CoordTheta = Coord<Theta>;
+    using SplineMem = host_t<DFieldMem<IdxRange<BSplines>>>;
+    using Evaluator = PolarSplineEvaluator<BSplines, ddc::NullExtrapolationRule>;
+    using BuilderRTheta = ddc::SplineBuilder2D<
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
+            BSplinesR,
+            BSplinesTheta,
+            GridR,
+            GridTheta,
+            ddc::BoundCond::GREVILLE,
+            ddc::BoundCond::GREVILLE,
+            ddc::BoundCond::PERIODIC,
+            ddc::BoundCond::PERIODIC,
+            ddc::SplineSolver::LAPACK>;
+
+    using EvaluatorRTheta = ddc::SplineEvaluator2D<
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
+            BSplinesR,
+            BSplinesTheta,
+            GridR,
+            GridTheta,
+            ddc::NullExtrapolationRule,
+            ddc::NullExtrapolationRule,
+            ddc::PeriodicExtrapolationRule<Theta>,
+            ddc::PeriodicExtrapolationRule<Theta>>;
+
+    CoordR constexpr r0(0.);
+    CoordR constexpr rN(1.);
+    CoordTheta constexpr theta0(0.);
+    CoordTheta constexpr thetaN(2. * M_PI);
+    std::size_t constexpr ncells = 20;
+
+    // 1. Create BSplines
+    {
+        IdxStep<GridR> constexpr npoints_r(ncells + 1);
+        std::vector<CoordR> breaks_r(npoints_r);
+        const double dr = (rN - r0) / ncells;
+        for (int i(0); i < npoints_r; ++i) {
+            breaks_r[i] = CoordR(r0 + i * dr);
+        }
+        ddc::init_discrete_space<BSplinesR>(breaks_r);
+#if defined(BSPLINES_TYPE_UNIFORM)
+        ddc::init_discrete_space<BSplinesTheta>(theta0, thetaN, ncells);
+#elif defined(BSPLINES_TYPE_NON_UNIFORM)
+        IdxStep<GridTheta> constexpr npoints_theta(ncells + 1);
+        std::vector<CoordTheta> breaks_theta(npoints_theta);
+        const double dp = (thetaN - theta0) / ncells;
+        for (int i(0); i < npoints_r; ++i) {
+            breaks_theta[i] = CoordTheta(theta0 + i * dp);
+        }
+        ddc::init_discrete_space<BSplinesTheta>(breaks_theta);
+#endif
+    }
+
+    ddc::init_discrete_space<GridR>(GrevillePointsR::get_sampling<GridR>());
+    ddc::init_discrete_space<GridTheta>(GrevillePointsTheta::get_sampling<GridTheta>());
+    IdxRange<GridR> interpolation_idx_range_r(GrevillePointsR::get_domain<GridR>());
+    IdxRange<GridTheta> interpolation_idx_range_theta(GrevillePointsTheta::get_domain<GridTheta>());
+    IdxRange<GridR, GridTheta>
+            interpolation_idx_range(interpolation_idx_range_r, interpolation_idx_range_theta);
+
+    BuilderRTheta builder_rtheta(interpolation_idx_range);
+
[…]
