+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -254,10 +254,15 @@ private:
 
     FieldMem<double, IdxRangeQuadratureRTheta> m_int_volume;
 
-    PolarSplineEvaluator<PolarBSplinesRTheta, ddc::NullExtrapolationRule> m_polar_spline_evaluator;
+    PolarSplineEvaluator<
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
+            PolarBSplinesRTheta,
+            ddc::NullExtrapolationRule>
+            m_polar_spline_evaluator;
     std::unique_ptr<MatrixBatchCsr<Kokkos::DefaultExecutionSpace, MatrixBatchCsrSolver::CG>>
             m_gko_matrix;
-    mutable host_t<PolarSplineMemRTheta> m_phi_spline_coef;
+    mutable PolarSplineMemRTheta m_phi_spline_coef;
     Kokkos::View<double**, Kokkos::LayoutRight> m_x_init;
 
     const int m_batch_idx {0}; // TODO: Remove when batching is supported
@@ -1010,13 +1015,7 @@ public:
                 Kokkos::DefaultExecutionSpace(),
                 get_idx_range(phi),
                 KOKKOS_LAMBDA(IdxRTheta idx) { coords_eval(idx) = ddc::coordinate(idx); });
-        auto coords_eval_host = ddc::create_mirror_and_copy(coords_eval);
-        auto phi_host = ddc::create_mirror_and_copy(phi);
-        m_polar_spline_evaluator(
-                get_field(phi_host),
-                get_const_field(coords_eval_host),
-                get_const_field(m_phi_spline_coef));
-        ddc::parallel_deepcopy(phi, phi_host);
+        m_polar_spline_evaluator(phi, coords_eval, get_const_field(m_phi_spline_coef));
     }
 
     /**
