### 2025-07-16 fc08296
Remove unused broken function
