+++ tests/interpolation/polar_splines.cpp
@@ -11,7 +11,6 @@
 #include "discrete_mapping_builder.hpp"
 #include "discrete_to_cartesian.hpp"
 #include "polar_bsplines.hpp"
-#include "polar_spline.hpp"
 #include "polar_spline_evaluator.hpp"
 #include "view.hpp"
 
