### 2025-07-16 9b090b2
Fix handling of missing theta duplicates
