+++ CHANGELOG.md
@@ -23,7 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
-- Remove unused broken metho `PolarSplineEvaluator::integrate`.
+- Remove unhelpful `PolarSpline` classes in favour of `DField<IdxRange<PolarBSplines>>` types.
+- Remove unused broken method `PolarSplineEvaluator::integrate`.
 
 ## [v0.2.0] - 2025-07-03
 
