### 2025-07-17 7dac77c
Return std::string::npos if no index is found to avoid uninitialised errors
