+++ src/multipatch/connectivity/edge_transformation.hpp
@@ -48,9 +48,6 @@ class EdgeTransformation
     using IdxRangeEdge1 = IdxRange<EdgeGrid1>;
     using IdxRangeEdge2 = IdxRange<EdgeGrid2>;
 
-    using IdxEdge1 = Idx<EdgeGrid1>;
-    using IdxEdge2 = Idx<EdgeGrid2>;
-
     using Patch1 = typename Interface::Edge1::associated_patch;
     using Patch2 = typename Interface::Edge2::associated_patch;
 
@@ -179,11 +176,10 @@ public:
     template <class CurrentIdx>
     auto operator()(CurrentIdx const& current_idx) const
     {
-        using IdxTarget
-                = std::conditional_t<std::is_same_v<CurrentIdx, IdxEdge1>, IdxEdge2, IdxEdge1>;
-        IdxTarget target_idx;
-        [[maybe_unused]] bool is_equivalent_idx_found = search_for_match(target_idx, current_idx);
-        assert(is_equivalent_idx_found);
+        using TargetGrid = std::
+                conditional_t<std::is_same_v<CurrentIdx, Idx<EdgeGrid1>>, EdgeGrid2, EdgeGrid1>;
+        Idx<TargetGrid> target_idx = search_for_match<TargetGrid>(current_idx);
+        assert(target_idx != Idx<TargetGrid>(std::string::npos));
         return target_idx;
     }
 
@@ -210,10 +206,10 @@ public:
     template <class CurrentIdx>
     bool is_match_available(CurrentIdx const& current_idx) const
     {
-        using IdxTarget
-                = std::conditional_t<std::is_same_v<CurrentIdx, IdxEdge1>, IdxEdge2, IdxEdge1>;
-        IdxTarget potential_target_idx;
-        return search_for_match(potential_target_idx, current_idx);
+        using TargetGrid = std::
+                conditional_t<std::is_same_v<CurrentIdx, Idx<EdgeGrid1>>, EdgeGrid2, EdgeGrid1>;
+        Idx<TargetGrid> target_idx = search_for_match<TargetGrid>(current_idx);
+        return search_for_match<TargetGrid>(current_idx) != Idx<TargetGrid>(std::string::npos);
     }
 
 
@@ -245,8 +241,8 @@ public:
      * @return Boolean stating if there is an equivalent index. 
      * 
      */
-    template <class CurrentGrid, class TargetGrid>
-    bool search_for_match(Idx<TargetGrid>& target_idx, Idx<CurrentGrid> current_idx) const
+    template <class TargetGrid, class CurrentGrid>
+    Idx<TargetGrid> search_for_match(Idx<CurrentGrid> current_idx) const
     {
         static_assert(
                 std::is_same_v<CurrentGrid, EdgeGrid1> || std::is_same_v<CurrentGrid, EdgeGrid2>,
@@ -329,7 +325,7 @@ public:
             if (is_equivalent_idx) {
                 double const rescaling_factor = double(n_cells_current) / double(n_cells_target);
                 IdxStepTarget target_idx_step_rescaled(int(current_idx_value / rescaling_factor));
-                target_idx = target_idx_range.front() + target_idx_step_rescaled;
+                Idx<TargetGrid> target_idx = target_idx_range.front() + target_idx_step_rescaled;
 
                 if constexpr (!Interface::orientations_agree) {
                     IdxStepTarget target_idx_step = target_idx_range.back() - target_idx;
@@ -345,17 +341,20 @@ public:
                         target_idx = target_idx_range.front();
                     }
                 }
+                return target_idx;
             }
         } else { // Non uniform case
             // Dichotomy method comparing the coordinates of indexes of the target edge.
-            target_idx = get_target_idx(target_idx_range, current_idx);
[…]
