### 2025-07-17 25be7a1
Unused variables
