### 2025-07-17 0badf42
Clean up with out_of_bounds_idx static constexpr
