### 2025-07-17 35dd5bc
Periodic handling removed
