### 2025-07-17 f7b9ad0
Make is_equivalent local
