### 2025-07-17 0fddefc
Unused variable
