+++ src/geometryRTheta/time_solver/bsl_predcorr_second_order_explicit.hpp
@@ -165,11 +165,6 @@ public:
         // Grid. ------------------------------------------------------------------------------------------
         IdxRangeRTheta const grid(get_idx_range<GridR, GridTheta>(density_host));
 
-        host_t<FieldMemRTheta<CoordRTheta>> coords(grid);
-        ddc::for_each(grid, [&](IdxRTheta const irtheta) {
-            coords(irtheta) = ddc::coordinate(irtheta);
-        });
-
         // --- Electrostatic potential (phi). -------------------------------------------------------------
         DFieldMemRTheta electrical_potential(grid);
         host_t<DFieldMemRTheta> electrical_potential_host(grid);
@@ -213,7 +208,6 @@ public:
 
             polar_spline_evaluator(
                     get_field(electrical_potential_host),
-                    get_const_field(coords),
                     get_const_field(electrostatic_potential_coef));
 
             ddc::PdiEvent("iteration")
