+++ src/geometryRTheta/advection_field/advection_field_rtheta.hpp
@@ -227,21 +227,14 @@ private:
         IdxRangeRTheta const grid = get_idx_range(advection_field_xy);
         host_t<DVectorFieldMemRTheta<X, Y>> electric_field(grid);
 
-        host_t<FieldMemRTheta<CoordRTheta>> coords(grid);
-        ddc::for_each(grid, [&](IdxRTheta const irtheta) {
-            coords(irtheta) = ddc::coordinate(irtheta);
-        });
-
         // > computation of the phi derivatives
         host_t<DVectorFieldMemRTheta<R_cov, Theta_cov>> deriv_phi(grid);
 
         evaluator.deriv_dim_1(
                 ddcHelper::get<R_cov>(deriv_phi),
-                get_const_field(coords),
                 get_const_field(electrostatic_potential_coef));
         evaluator.deriv_dim_2(
                 ddcHelper::get<Theta_cov>(deriv_phi),
-                get_const_field(coords),
                 get_const_field(electrostatic_potential_coef));
 
         InverseJacobianMatrix inv_jacobian_matrix(m_mapping);
