### 2025-07-17 7884f57
Remove deprecated method `PolarBSplines::integrals` (#379)
Remove deprecated method `PolarBSplines::integrals`.
