+++ assets/.doxy/gyselalibxx/gyselalibxx/polar__bsplines_8hpp_source.md
@@ -319,10 +319,6 @@ public:
         KOKKOS_FUNCTION tensor_product_index_type
         eval_deriv_r_and_theta(DSpan1D singular_derivs, DSpan2D derivs, Coord<R, Theta> p) const;
 
-        template <class MemorySpace2>
-        [[deprecated("Use `integrals` instead")]] void integrals(
-                PolarSpline<DDim, MemorySpace2> int_vals) const;
-
         std::size_t nbasis() const noexcept
         {
             std::size_t nr = ddc::discrete_space<BSplinesR>().nbasis() - C - 1;
@@ -559,15 +555,6 @@ PolarSpline<DDim, MemorySpace> integrals(
         ddc::parallel_fill(execution_space, int_vals.spline_coef[repeat_idx_range], 0.0);
     }
     return int_vals;
-}
-
-template <class BSplinesR, class BSplinesTheta, int C>
-template <class DDim, class MemorySpace>
-template <class MemorySpace2>
-void PolarBSplines<BSplinesR, BSplinesTheta, C>::Impl<DDim, MemorySpace>::integrals(
-        PolarSpline<DDim, MemorySpace2> int_vals) const
-{
-    integrals(Kokkos::DefaultHostExecutionSpace(), int_vals);
 }
 ```
 
