### 2025-07-17 35dc9ff
Correct get_idx_range
