### 2025-07-17 25cda0e
Use Kokkos math function
