+++ toolchains/v100.persee/environment.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 if [ "${BASH_SOURCE[0]}" -ef "$0" ]
 then
     echo "This script must be sourced not executed."
