+++ toolchains/v100.ruche/toolchain.cmake
@@ -18,6 +18,13 @@ set(Kokkos_ENABLE_CUDA_CONSTEXPR ON CACHE BOOL "" FORCE )
 set(Kokkos_ENABLE_CUDA_RELOCATABLE_DEVICE_CODE ON CACHE BOOL "" FORCE)
 set(Kokkos_ARCH_VOLTA70 ON CACHE BOOL "Indicate that the GPU architecture is V100" FORCE)
 set(Kokkos_ENABLE_DEPRECATED_CODE_4 OFF CACHE BOOL "" FORCE)
+
+# Kokkos Kernels options
+set(KokkosKernels_ENABLE_ALL_COMPONENTS OFF CACHE BOOL "")
+set(KokkosKernels_ENABLE_COMPONENT_BATCHED ON CACHE BOOL "")
+set(KokkosKernels_ENABLE_COMPONENT_BLAS ON CACHE BOOL "")
+set(KokkosKernels_ADD_DEFAULT_ETI OFF CACHE BOOL "")
+
 # Activate/deactivate parts of the code
 if (DEFINED ENV{DDC_BUILD_TESTING})
     set(DDC_BUILD_TESTS $ENV{DDC_BUILD_TESTING})
