+++ toolchains/cpu.spack.gyselalibxx_env/toolchain.cmake
@@ -7,6 +7,12 @@ set(Python3_EXECUTABLE $ENV{PYTHON_EXECUTABLE})
 # CMake options
 set(CMAKE_BUILD_TYPE Release)
 
+# Kokkos Kernels options
+set(KokkosKernels_ENABLE_ALL_COMPONENTS OFF CACHE BOOL "")
+set(KokkosKernels_ENABLE_COMPONENT_BATCHED ON CACHE BOOL "")
+set(KokkosKernels_ENABLE_COMPONENT_BLAS ON CACHE BOOL "")
+set(KokkosKernels_ADD_DEFAULT_ETI OFF CACHE BOOL "")
+
 # Compiler options
 set(CMAKE_CXX_FLAGS "-Wall -Wno-sign-compare")
 
