+++ toolchains/a100.leonardo.spack/toolchain.cmake
@@ -17,6 +17,12 @@ set(Kokkos_ENABLE_SERIAL ON CACHE BOOL "Allow serial code to run" FORCE)
 set(Kokkos_ENABLE_CUDA ON CACHE BOOL "Activate GPU usage via cuda" FORCE)
 set(Kokkos_ARCH_AMPERE80 ON CACHE BOOL "Indicate that the GPU architecture is A100" FORCE)
 
+# Kokkos Kernels options
+set(KokkosKernels_ENABLE_ALL_COMPONENTS OFF CACHE BOOL "")
+set(KokkosKernels_ENABLE_COMPONENT_BATCHED ON CACHE BOOL "")
+set(KokkosKernels_ENABLE_COMPONENT_BLAS ON CACHE BOOL "")
+set(KokkosKernels_ADD_DEFAULT_ETI OFF CACHE BOOL "")
+
 # Activate/deactivate parts of the code
 if (DEFINED ENV{DDC_BUILD_TESTING})
     set(DDC_BUILD_TESTS $ENV{DDC_BUILD_TESTING})
