### 2025-07-17 123cea8
Clang format
