### 2025-07-17 b658d0e
Improve docstring
Co-authored-by: alex-m-h <81037585+alex-m-h@users.noreply.github.com>
