### 2025-07-17 650f90a
Prefer DFieldMem over Kokkos::View
