+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -54,6 +54,8 @@ class PolarSplineFEMPoissonLikeSolver
             std::is_same_v<IdxRangeFull, IdxRange<GridR, GridTheta>>,
             "PolarSplineFEMPoissonLikeSolver is not yet batched");
 
+    class InternalBatchDim;
+
 public:
     /// The radial dimension
     using R = typename GridR::continuous_dimension_type;
@@ -262,8 +264,8 @@ private:
             m_polar_spline_evaluator;
     std::unique_ptr<MatrixBatchCsr<Kokkos::DefaultExecutionSpace, MatrixBatchCsrSolver::CG>>
             m_gko_matrix;
-    mutable host_t<PolarSplineMemRTheta> m_phi_spline_coef;
-    Kokkos::View<double**, Kokkos::LayoutRight> m_x_init;
+    mutable PolarSplineMemRTheta m_phi_spline_coef;
+    mutable DFieldMem<IdxRange<InternalBatchDim, PolarBSplinesRTheta>> m_x_init;
 
     const int m_batch_idx {0}; // TODO: Remove when batching is supported
 public:
@@ -334,13 +336,16 @@ public:
         , m_phi_spline_coef(ddc::discrete_space<PolarBSplinesRTheta>().full_domain())
         , m_x_init(
                   "x_init",
-                  1,
-                  ddc::discrete_space<PolarBSplinesRTheta>().nbasis()
-                          - ddc::discrete_space<BSplinesTheta>().nbasis())
+                  IdxRange<InternalBatchDim, PolarBSplinesRTheta>(
+                          Idx<InternalBatchDim, PolarBSplinesRTheta>(0, 0),
+                          IdxStep<InternalBatchDim, PolarBSplinesRTheta>(
+                                  1,
+                                  ddc::discrete_space<PolarBSplinesRTheta>().nbasis()
+                                          - ddc::discrete_space<BSplinesTheta>().nbasis())))
     {
         static_assert(has_jacobian_v<Mapping>);
         //initialise x_init
-        Kokkos::deep_copy(m_x_init, 0);
+        ddc::parallel_fill(get_field(m_x_init), 0);
         // Get break points
         IdxRange<KnotsR> idxrange_r_edges = ddc::discrete_space<BSplinesR>().break_point_domain();
         IdxRange<KnotsTheta> idxrange_theta_edges
@@ -865,25 +870,31 @@ public:
                 std::is_invocable_r_v<double, RHSFunction, CoordRTheta>,
                 "RHSFunction must have an operator() which takes a coordinate and returns a "
                 "double");
-        const int b_size = ddc::discrete_space<PolarBSplinesRTheta>().nbasis()
-                           - ddc::discrete_space<BSplinesTheta>().nbasis();
-        const int batch_size = 1;
+
+        assert(get_idx_range(spline) == ddc::discrete_space<PolarBSplinesRTheta>().full_domain());
+
+        IdxRange<InternalBatchDim> batch_idx_range(get_idx_range(m_x_init));
+
+        assert(batch_idx_range.size() == 1);
+
+        Idx<InternalBatchDim> batch_idx = batch_idx_range.front();
+
         // Create b for rhs
-        Kokkos::View<double**, Kokkos::LayoutRight, Kokkos::HostSpace>
-                b_host("b_host", batch_size, b_size);
+        host_t<DFieldMem<IdxRange<InternalBatchDim, PolarBSplinesRTheta>>> b_host_alloc(
+                get_idx_range(m_x_init));
         //Create an initial guess
-        Kokkos::View<double**, Kokkos::LayoutRight, Kokkos::HostSpace>
[…]
