+++ src/geometryRTheta/advection_field/advection_field_rtheta.hpp
@@ -8,7 +8,6 @@
 #include "iqnsolver.hpp"
 #include "metric_tensor_evaluator.hpp"
 #include "poisson_like_rhs_function.hpp"
-#include "polar_spline.hpp"
 #include "polar_spline_evaluator.hpp"
 #include "polarpoissonlikesolver.hpp"
 #include "vector_field.hpp"
@@ -183,7 +182,7 @@ public:
      *      The advection field on the physical axis. 
      */
     void operator()(
-            host_t<PolarSplineMemRTheta>& electrostatic_potential_coef,
+            host_t<PolarSplineRTheta> electrostatic_potential_coef,
             host_t<DVectorFieldRTheta<X, Y>> advection_field_xy) const
     {
         compute_advection_field_XY(
@@ -218,7 +217,7 @@ private:
                             Evaluator,
                             PolarSplineEvaluator<
                                     PolarBSplinesRTheta,
-                                    ddc::NullExtrapolationRule>> && std::is_same_v<SplineType, host_t<PolarSplineMemRTheta>>));
+                                    ddc::NullExtrapolationRule>> && std::is_same_v<SplineType, host_t<PolarSplineRTheta>>));
 
         IdxRangeRTheta const grid = get_idx_range(advection_field_xy);
         host_t<DVectorFieldMemRTheta<X, Y>> electric_field(grid);
@@ -407,7 +406,7 @@ public:
      *      The advection field at the centre point on the Cartesian basis.
      */
     void operator()(
-            host_t<PolarSplineMemRTheta>& electrostatic_potential_coef,
+            host_t<PolarSplineRTheta>& electrostatic_potential_coef,
             host_t<DVectorFieldRTheta<R, Theta>> advection_field_rtheta,
             DVector<X, Y>& advection_field_xy_centre) const
     {
@@ -449,7 +448,7 @@ private:
                             Evaluator,
                             PolarSplineEvaluator<
                                     PolarBSplinesRTheta,
-                                    ddc::NullExtrapolationRule>> && std::is_same_v<SplineType, host_t<PolarSplineMemRTheta>>));
+                                    ddc::NullExtrapolationRule>> && std::is_same_v<SplineType, host_t<PolarSplineRTheta>>));
 
         IdxRangeRTheta const grid_without_Opoint = get_idx_range(advection_field_rtheta);
 
