+++ tests/interpolation/polar_splines.cpp
@@ -11,7 +11,6 @@
 #include "discrete_mapping_builder.hpp"
 #include "discrete_to_cartesian.hpp"
 #include "polar_bsplines.hpp"
-#include "polar_spline.hpp"
 #include "polar_spline_evaluator.hpp"
 #include "view.hpp"
 
@@ -91,7 +90,7 @@ TEST(PolarSplineTest, ConstantEval)
     using PolarCoord = Coord<R, Theta>;
     using CoordR = Coord<R>;
     using CoordTheta = Coord<Theta>;
-    using SplineMem = PolarSplineMem<BSplines, Kokkos::HostSpace>;
+    using SplineMem = host_t<DFieldMem<IdxRange<BSplines>>>;
     using Evaluator = PolarSplineEvaluator<BSplines, ddc::NullExtrapolationRule>;
     using BuilderRTheta = ddc::SplineBuilder2D<
             Kokkos::DefaultHostExecutionSpace,
@@ -176,14 +175,8 @@ TEST(PolarSplineTest, ConstantEval)
     DiscreteToCartesian mapping = mapping_builder();
     ddc::init_discrete_space<BSplines>(mapping);
 
-    SplineMem coef(get_spline_idx_range(builder_rtheta));
-
-    ddc::for_each(get_idx_range(coef.singular_spline_coef), [&](Idx<BSplines> const i) {
-        coef.singular_spline_coef(i) = 1.0;
-    });
-    ddc::for_each(get_idx_range(coef.spline_coef), [&](Idx<BSplinesR, BSplinesTheta> const i) {
-        coef.spline_coef(i) = 1.0;
-    });
+    SplineMem coef(ddc::discrete_space<BSplines>().full_domain());
+    ddc::parallel_fill(get_field(coef), 1.0);
 
     ddc::NullExtrapolationRule extrapolation_rule;
     const Evaluator spline_evaluator(extrapolation_rule);
@@ -210,11 +203,11 @@ void test_polar_integrals()
 {
     using CoordR = Coord<R>;
     using CoordTheta = Coord<Theta>;
-    using SplineMem = PolarSplineMem<BSplines>;
-    using Spline = PolarSpline<BSplines>;
+    using SplineMem = DFieldMem<IdxRange<BSplines>>;
+    using Spline = DField<IdxRange<BSplines>>;
     using BuilderRTheta = ddc::SplineBuilder2D<
-            Kokkos::DefaultHostExecutionSpace,
-            Kokkos::DefaultHostExecutionSpace::memory_space,
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
             BSplinesR,
             BSplinesTheta,
             GridR,
@@ -226,8 +219,8 @@ void test_polar_integrals()
             ddc::SplineSolver::LAPACK>;
 
     using EvaluatorRTheta = ddc::SplineEvaluator2D<
-            Kokkos::DefaultHostExecutionSpace,
-            Kokkos::DefaultHostExecutionSpace::memory_space,
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
             BSplinesR,
             BSplinesTheta,
             GridR,
@@ -288,32 +281,22 @@ void test_polar_integrals()
     CircToCart const coord_changer(0.3, 1.4);
 #endif
     DiscreteToCartesianBuilder<X, Y, BuilderRTheta, EvaluatorRTheta> mapping_builder(
-            Kokkos::DefaultHostExecutionSpace(),
+            Kokkos::DefaultExecutionSpace(),
             coord_changer,
             builder_rtheta,
             evaluator_rtheta);
[…]
