+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -10,7 +10,6 @@
 #include "math_tools.hpp"
 #include "matrix_batch_csr.hpp"
 #include "metric_tensor_evaluator.hpp"
-#include "polar_spline.hpp"
 #include "polar_spline_evaluator.hpp"
 #include "quadrature_coeffs_nd.hpp"
 #include "view.hpp"
@@ -170,7 +169,8 @@ private:
     using KnotsTheta = ddc::NonUniformBsplinesKnots<BSplinesTheta>;
 
     using ConstSpline2D = DConstField<IdxRangeBatchedBSRTheta>;
-    using PolarSplineMemRTheta = PolarSplineMem<PolarBSplinesRTheta>;
+    using PolarSplineMemRTheta = DFieldMem<IdxRange<PolarBSplinesRTheta>>;
+    using PolarSplineRTheta = DField<IdxRange<PolarBSplinesRTheta>>;
 
     using CoordFieldMemRTheta = FieldMem<CoordRTheta, IdxRangeRTheta>;
     using CoordFieldRTheta = Field<CoordRTheta, IdxRangeRTheta>;
@@ -326,11 +326,7 @@ public:
                           ThetaBasisSubset,
                           QDimThetaMesh>(m_non_zero_bases_theta, m_idxrange_quadrature_theta))
         , m_polar_spline_evaluator(ddc::NullExtrapolationRule())
-        , m_phi_spline_coef(
-                  PolarBSplinesRTheta::template singular_idx_range<PolarBSplinesRTheta>(),
-                  IdxRangeBSRTheta(
-                          m_idxrange_bsplines_r,
-                          ddc::discrete_space<BSplinesTheta>().full_domain()))
+        , m_phi_spline_coef(ddc::discrete_space<PolarBSplinesRTheta>().full_domain())
         , m_x_init(
                   "x_init",
                   1,
@@ -856,7 +852,7 @@ public:
      *      The spline representation of the solution @f$\phi@f$, also used as initial data for the iterative solver.
      */
     template <class RHSFunction>
-    void operator()(RHSFunction const& rhs, host_t<PolarSplineMemRTheta>& spline) const
+    void operator()(RHSFunction const& rhs, host_t<PolarSplineRTheta> spline) const
     {
         Kokkos::Profiling::pushRegion("PolarPoissonRHS");
 
@@ -965,7 +961,6 @@ public:
         IdxRangeBSRTheta dirichlet_boundary_idx_range(
                 m_idxrange_bsplines_r.take_last(IdxStep<BSplinesR> {1}),
                 m_idxrange_bsplines_theta);
-        IdxRangeBSTheta idxrange_polar(ddc::discrete_space<BSplinesTheta>().full_domain());
 
         // Fill the spline
         ddc::for_each(
@@ -975,27 +970,13 @@ public:
                                          - PolarBSplinesRTheta::template singular_idx_range<
                                                    PolarBSplinesRTheta>()
                                                    .front();
-                    spline.singular_spline_coef(idx) = x_init_host(0, bspl_idx);
+                    spline(idx) = x_init_host(0, bspl_idx);
                 });
         ddc::for_each(m_idxrange_fem_non_singular, [&](IdxBSPolar const idx) {
-            const IdxBSRTheta idx_2d(PolarBSplinesRTheta::get_2d_index(idx));
-            spline.spline_coef(idx_2d) = x_init_host(0, idx.uid());
+            spline(idx) = x_init_host(0, idx.uid());
         });
         ddc::for_each(dirichlet_boundary_idx_range, [&](IdxBSRTheta const idx) {
-            spline.spline_coef(idx) = 0.0;
-        });
-
-        // Copy the periodic elements
[…]

[…]
