### 2025-07-17 2a793f7
Remove `PolarSpline` type (#355)
The `PolarSpline` type causes more problems than it solves so this PR
removes it. Fixes #289 .
Instead a simple `Field` defined over the basis splines is used. Helper
methods are provided in `PolarBSplines` to extract the subset of fields
that were previously stored in `PolarSpline` so this can be used when it
is useful.

The untested method `PolarSplineEvaluator::integrate` containing code
which does not compile (in particular `mapping.determinant(i)`) is
removed.

Following this change there are no duplicate elements in the field of
coefficients. This means t
[…]
