+++ src/interpolation/polar_splines/polar_spline_evaluator.hpp
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
-#include "polar_spline.hpp"
 #include "view.hpp"
 
 /**
@@ -142,7 +141,7 @@ public:
      */
     double operator()(
             Coord<DimR, DimTheta> coord_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         return eval(coord_eval, spline_coef);
     }
@@ -161,7 +160,7 @@ public:
     void operator()(
             DField<Domain, Kokkos::HostSpace> const spline_eval,
             ConstField<Coord<DimR, DimTheta>, Domain, Kokkos::HostSpace> const coords_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         using IdxEval = typename Domain::discrete_element_type;
         ddc::for_each(get_idx_range(coords_eval), [=](IdxEval i) {
@@ -183,7 +182,7 @@ public:
      */
     double deriv_dim_1(
             Coord<DimR, DimTheta> coord_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         return eval_no_bc(coord_eval, spline_coef, eval_deriv_r_type());
     }
@@ -202,11 +201,10 @@ public:
      */
     double deriv_dim_2(
             Coord<DimR, DimTheta> coord_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         return eval_no_bc(coord_eval, spline_coef, eval_deriv_theta_type());
     }
-
 
     /**
      * @brief Get the value of the cross derivative of the spline function.
@@ -221,7 +219,7 @@ public:
      */
     double deriv_1_and_2(
             Coord<DimR, DimTheta> coord_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         return eval_no_bc(coord_eval, spline_coef, eval_deriv_r_theta_type());
     }
@@ -241,7 +239,7 @@ public:
     void deriv_dim_1(
             DField<Domain, Kokkos::HostSpace> const spline_eval,
             ConstField<Coord<DimR, DimTheta>, Domain, Kokkos::HostSpace> const coords_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         using IdxEval = typename Domain::discrete_element_type;
         ddc::for_each(get_idx_range(coords_eval), [=](IdxEval i) {
@@ -264,7 +262,7 @@ public:
     void deriv_dim_2(
             DField<Domain, Kokkos::HostSpace> const spline_eval,
             ConstField<Coord<DimR, DimTheta>, Domain, Kokkos::HostSpace> const coords_eval,
-            host_t<ConstPolarSpline<PolarBSplinesType>> const spline_coef) const
+            host_t<DConstField<IdxRange<PolarBSplinesType>>> const spline_coef) const
     {
         using IdxEval = typename Domain::discrete_element_type;
         ddc::for_each(get_idx_range(coords_eval), [=](IdxEval i) {
[…]
