+++ assets/.doxy/gyselalibxx/gyselalibxx/polar__spline__evaluator_8hpp.md
@@ -10,7 +10,6 @@
 
 
 
-* `#include "polar_spline.hpp"`
 * `#include "view.hpp"`
 
 
