+++ assets/.doxy/gyselalibxx/gyselalibxx/structPolarBSplinesRTheta.md
@@ -103,6 +103,8 @@ See [PolarBSplines](classPolarBSplines.md)
 | ---: | :--- |
 |  KOKKOS\_FUNCTION [**tensor\_product\_index\_type**](classPolarBSplines.md#typedef-tensor_product_index_type) | [**get\_2d\_index**](classPolarBSplines.md#function-get_2d_index) (Idx&lt; DDim &gt; const & idx) <br> |
 |  KOKKOS\_FUNCTION Idx&lt; DDim &gt; | [**get\_polar\_index**](classPolarBSplines.md#function-get_polar_index) ([**tensor\_product\_index\_type**](classPolarBSplines.md#typedef-tensor_product_index_type) const & idx) <br> |
+|  Field&lt; ElementType, IdxRange&lt; DDim &gt;, MemorySpace &gt; | [**get\_singular\_subset**](classPolarBSplines.md#function-get_singular_subset) (Field&lt; ElementType, IdxRange&lt; DDim &gt;, MemorySpace &gt; coeffs) <br>_Get the subset of the coefficients associated with singular basis splines._  |
+|  Field&lt; ElementType, [**tensor\_product\_idx\_range\_type**](classPolarBSplines.md#typedef-tensor_product_idx_range_type), MemorySpace &gt; | [**get\_tensor\_product\_subset**](classPolarBSplines.md#function-get_tensor_product_subset) (Field&lt; ElementType, IdxRange&lt; DDim &gt;, MemorySpace &gt; coeffs) <br>_Get the subset of the coefficients of tensor product basis splines._  |
 |  constexpr std::size\_t | [**n\_singular\_basis**](classPolarBSplines.md#function-n_singular_basis) () <br> |
 |  constexpr IdxRange&lt; DDim &gt; | [**singular\_idx\_range**](classPolarBSplines.md#function-singular_idx_range) () <br>_Get the IdxRange containing the indices of the b-splines which traverse the singular point._  |
 
