### 2025-07-17 dc156a8
Change interface of `EdgeTransformation::search_for_match` to fix uninitialised warnings (#380)
Change interface of `EdgeTransformation::search_for_match` to return an
`out_of_bounds_idx` instead of a boolean. This fixes the warnings about
uninitialised values.
