+++ src/multipatch/connectivity/edge_transformation.hpp
@@ -11,6 +11,8 @@
 #include "edge.hpp"
 #include "interface.hpp"
 
+template <class TargetGrid>
+static constexpr Idx<TargetGrid> out_of_bounds_idx = Idx<TargetGrid>(std::string::npos);
 
 /**
  * @brief Transform a coordinate or an index from one edge to the one on the other edge.
@@ -47,9 +49,6 @@ class EdgeTransformation
 
     using IdxRangeEdge1 = IdxRange<EdgeGrid1>;
     using IdxRangeEdge2 = IdxRange<EdgeGrid2>;
-
-    using IdxEdge1 = Idx<EdgeGrid1>;
-    using IdxEdge2 = Idx<EdgeGrid2>;
 
     using Patch1 = typename Interface::Edge1::associated_patch;
     using Patch2 = typename Interface::Edge2::associated_patch;
@@ -179,11 +178,10 @@ public:
     template <class CurrentIdx>
     auto operator()(CurrentIdx const& current_idx) const
     {
-        using IdxTarget
-                = std::conditional_t<std::is_same_v<CurrentIdx, IdxEdge1>, IdxEdge2, IdxEdge1>;
-        IdxTarget target_idx;
-        [[maybe_unused]] bool is_equivalent_idx_found = search_for_match(target_idx, current_idx);
-        assert(is_equivalent_idx_found);
+        using TargetGrid = std::
+                conditional_t<std::is_same_v<CurrentIdx, Idx<EdgeGrid1>>, EdgeGrid2, EdgeGrid1>;
+        Idx<TargetGrid> target_idx = search_for_match<TargetGrid>(current_idx);
+        assert(target_idx != out_of_bounds_idx<TargetGrid>);
         return target_idx;
     }
 
@@ -210,10 +208,9 @@ public:
     template <class CurrentIdx>
     bool is_match_available(CurrentIdx const& current_idx) const
     {
-        using IdxTarget
-                = std::conditional_t<std::is_same_v<CurrentIdx, IdxEdge1>, IdxEdge2, IdxEdge1>;
-        IdxTarget potential_target_idx;
-        return search_for_match(potential_target_idx, current_idx);
+        using TargetGrid = std::
+                conditional_t<std::is_same_v<CurrentIdx, Idx<EdgeGrid1>>, EdgeGrid2, EdgeGrid1>;
+        return search_for_match<TargetGrid>(current_idx) != out_of_bounds_idx<TargetGrid>;
     }
 
 
@@ -229,24 +226,23 @@ public:
      * can use a dichotomy method. 
      * 
      * @warning target_idx is always replaced by the suspected index. 
-     * If there is not equivalent index, the returned index is wrong but the 
-     * closest that the algorithm found. 
+     * If there is not equivalent index, the returned index is out_of_bounds_idx.
      * 
-     * @tparam CurrentGrid The grid where the input index is defined. 
      * @tparam TargetGrid The grid where the output index is defined.
-
-     * @param[out] target_idx
-     *      A index on the edge of the target patch.
+     * @tparam CurrentGrid The grid where the input index is defined.
+     *                     This template parameter can be deduced from the function call.
+     *
      * @param[in] current_idx
      *      A index on the edge of the current patch.
-     * @tparam CurrentIdx 
-     *      The current index type of the given coordinate index. 
+     *
+     * @returns target_idx
+     *      A index on the edge of the target patch.
      * 
      * @return Boolean stating if there is an equivalent index. 
      * 
      */
-    template <class CurrentGrid, class TargetGrid>
-    bool search_for_match(Idx<TargetGrid>& target_idx, Idx<CurrentGrid> current_idx) const
[…]
