+++ assets/.doxy/gyselalibxx/gyselalibxx/class_member_typedefs.md
@@ -310,8 +310,6 @@
 * **IdxRangeSplines** ([**DiscreteToCartesianBuilder**](classDiscreteToCartesianBuilder.md), [**RefinedDiscreteToCartesianBuilder**](classRefinedDiscreteToCartesianBuilder.md))
 * **InterpolationField** ([**DiscreteToCartesianBuilder**](classDiscreteToCartesianBuilder.md), [**RefinedDiscreteToCartesianBuilder**](classRefinedDiscreteToCartesianBuilder.md))
 * **InterpolationFieldMem** ([**DiscreteToCartesianBuilder**](classDiscreteToCartesianBuilder.md), [**RefinedDiscreteToCartesianBuilder**](classRefinedDiscreteToCartesianBuilder.md))
-* **IdxEdge1** ([**EdgeTransformation**](classEdgeTransformation.md))
-* **IdxEdge2** ([**EdgeTransformation**](classEdgeTransformation.md))
 * **IdxRangeEdge1** ([**EdgeTransformation**](classEdgeTransformation.md))
 * **IdxRangeEdge2** ([**EdgeTransformation**](classEdgeTransformation.md))
 * **IdxBSplines** ([**FEM1DPoissonSolver**](classFEM1DPoissonSolver.md))
