+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Port `PolarSplineEvaluator` methods to GPU.
+- Add methods to `PolarSplineEvaluator` to avoid unnecessary creation of fields of coordinates.
+
 ### Fixed
 
 - Modify `ruche.v100/environment.sh` file to fix tokamaxi simulation segfault issues.
@@ -22,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change interface of `EdgeTransformation::search_for_match` to return an `out_of_bounds_idx` instead of a boolean.
+- Change template parameters of `PolarSplineEvaluator` to add execution and memory space information.
 
 ### Deprecated
 
