### 2025-07-17 53d6111
Remove some unnecessary calculations
