### 2025-07-18 994e229
CHANGELOG: mention cpu spack install changes from #361
