### 2025-07-18 4f2c6f9
Fix Adastra environment
- Remove cache usage for `CMAKE_CXX_FLAGS`
- Prefer serial over OpenMP libsci for consistency with Cray wrapper
