### 2025-07-18 1333ab7
Fix Adastra toolchain (#385)
- Remove cache usage for `CMAKE_CXX_FLAGS`, it did not work on the first
cmake configuration
- Prefer serial over OpenMP libsci for consistency with Cray wrapper
