### 2025-07-18 87f8d17
Fix Adastra toolchain (#385)
- Remove cache usage for `CMAKE_CXX_FLAGS`, it did not work on the first
cmake configuration
- Prefer serial over OpenMP libsci for consistency with Cray wrapper
