### 2025-07-18 67bf3c9
Add missing KOKKOS_FUNCTION tags
