### 2025-07-18 36d5798
Correct KOKKOS_LAMBDA->KOKKOS_CLASS_LAMBDA. Correct memspace
