### 2025-07-18 7033d0d
Clang formatting
