### 2025-07-18 5111b5e
Use _INIT variables in toolchains
