+++ toolchains/common_toolchains/openmp_release_toolchain.cmake
@@ -5,7 +5,7 @@ include(${CMAKE_CURRENT_LIST_DIR}/importable_defaults.cmake)
 set(CMAKE_BUILD_TYPE Release)
 
 # Compiler options
-set(CMAKE_CXX_FLAGS "-Wall -Wno-sign-compare")
+set(CMAKE_CXX_FLAGS_INIT "-Wall -Wno-sign-compare")
 
 # Kokkos options
 set(Kokkos_ENABLE_OPENMP ON CACHE BOOL "Activate OpenMP parallelisation" FORCE)
