+++ toolchains/xeon.persee/toolchain.cmake
@@ -5,7 +5,7 @@ include(${CMAKE_CURRENT_LIST_DIR}/../common_toolchains/importable_defaults.cmake
 set(CMAKE_BUILD_TYPE Release)
 
 # Compiler options
-set(CMAKE_CXX_FLAGS "-Wall -Wno-sign-compare -Wno-unused-but-set-variable")
+set(CMAKE_CXX_FLAGS_INIT "-Wall -Wno-sign-compare -Wno-unused-but-set-variable")
 
 # Kokkos options
 set(Kokkos_ENABLE_OPENMP ON CACHE BOOL "Activate OpenMP parallelisation" FORCE)
