### 2025-07-18 0aa9b05
Add mandatory options to Leonardo toolchain
