### 2025-07-18 00440d1
extend cpu toolchain prepare script (#361)
cpu toolchain prepare.sh script: set paths to use spack standalone
configuration (also in environment.sh)

This is useful for people who want to set up many different spack
instances, or for CI in an un-containerized environment.

---------

Co-authored-by: Emily Bourne <emily.bourne@epfl.ch>
