### 2025-07-18 2151b9e
Update CHANGELOG.md
Co-authored-by: Emily Bourne <emily.bourne@epfl.ch>
