+++ src/multipatch/data_types/multipatch_field.hpp
@@ -159,7 +159,7 @@ public:
      *
      * @returns The set of index ranges on which the set of fields stored in this class are defined.
      */
-    auto idx_range() const
+    KOKKOS_FUNCTION auto idx_range() const
     {
         return MultipatchType<InternalIdxRangeOnPatch, Patches...>(
                 get_idx_range(std::get<T<Patches>>(base_type::m_tuple))...);
