+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Port `PolarSplineEvaluator` methods to GPU.
+- Add methods to `PolarSplineEvaluator` to avoid unnecessary creation of fields of coordinates.
+
 ### Fixed
 
 - Modify `ruche.v100/environment.sh` file to fix tokamaxi simulation segfault issues.
@@ -24,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change interface of `EdgeTransformation::search_for_match` to return an `out_of_bounds_idx` instead of a boolean.
 - Change spack setup in CPU installation script (`prepare.sh`) to create and use independent spack installation.
+- Change template parameters of `PolarSplineEvaluator` to add execution and memory space information.
+- Allow `get_idx_range` to be called from a GPU execution space.
 
 ### Deprecated
 
