### 2025-07-18 7bd91b2
Port polar spline evaluator to GPU (#381)
Port polar spline evaluator functions to GPU and add tests. Fixes #42 
New methods are also added so the coordinates can be calculated on the
fly instead of being stored in a field. These changes require a change
in the template parameters of `PolarSplineEvaluator`.
Function calls are updated to avoid unnecessary allocations.
`get_idx_range` can now be called from GPU
