+++ tests/interpolation/polar_splines.cpp
@@ -91,7 +91,11 @@ TEST(PolarSplineTest, ConstantEval)
     using CoordR = Coord<R>;
     using CoordTheta = Coord<Theta>;
     using SplineMem = host_t<DFieldMem<IdxRange<BSplines>>>;
-    using Evaluator = PolarSplineEvaluator<BSplines, ddc::NullExtrapolationRule>;
+    using Evaluator = PolarSplineEvaluator<
+            Kokkos::DefaultHostExecutionSpace,
+            Kokkos::HostSpace,
+            BSplines,
+            ddc::NullExtrapolationRule>;
     using BuilderRTheta = ddc::SplineBuilder2D<
             Kokkos::DefaultHostExecutionSpace,
             Kokkos::HostSpace,
@@ -197,6 +201,129 @@ TEST(PolarSplineTest, ConstantEval)
             EXPECT_LE(fabs(deriv_2), 1.0e-13);
         }
     }
+}
+
+void test_polar_spline_eval_gpu()
+{
+    using CoordR = Coord<R>;
+    using CoordTheta = Coord<Theta>;
+    using SplineMem = DFieldMem<IdxRange<BSplines>>;
+    using Evaluator = PolarSplineEvaluator<
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
+            BSplines,
+            ddc::NullExtrapolationRule>;
+    using BuilderRTheta = ddc::SplineBuilder2D<
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
+            BSplinesR,
+            BSplinesTheta,
+            GridR,
+            GridTheta,
+            ddc::BoundCond::GREVILLE,
+            ddc::BoundCond::GREVILLE,
+            ddc::BoundCond::PERIODIC,
+            ddc::BoundCond::PERIODIC,
+            ddc::SplineSolver::LAPACK>;
+
+    using EvaluatorRTheta = ddc::SplineEvaluator2D<
+            Kokkos::DefaultExecutionSpace,
+            Kokkos::DefaultExecutionSpace::memory_space,
+            BSplinesR,
+            BSplinesTheta,
+            GridR,
+            GridTheta,
+            ddc::NullExtrapolationRule,
+            ddc::NullExtrapolationRule,
+            ddc::PeriodicExtrapolationRule<Theta>,
+            ddc::PeriodicExtrapolationRule<Theta>>;
+
+    CoordR constexpr r0(0.);
+    CoordR constexpr rN(1.);
+    CoordTheta constexpr theta0(0.);
+    CoordTheta constexpr thetaN(2. * M_PI);
+    std::size_t constexpr ncells = 20;
+
+    // 1. Create BSplines
+    {
+        IdxStep<GridR> constexpr npoints_r(ncells + 1);
+        std::vector<CoordR> breaks_r(npoints_r);
+        const double dr = (rN - r0) / ncells;
+        for (int i(0); i < npoints_r; ++i) {
+            breaks_r[i] = CoordR(r0 + i * dr);
+        }
+        ddc::init_discrete_space<BSplinesR>(breaks_r);
+#if defined(BSPLINES_TYPE_UNIFORM)
+        ddc::init_discrete_space<BSplinesTheta>(theta0, thetaN, ncells);
+#elif defined(BSPLINES_TYPE_NON_UNIFORM)
+        IdxStep<GridTheta> constexpr npoints_theta(ncells + 1);
+        std::vector<CoordTheta> breaks_theta(npoints_theta);
+        const double dp = (thetaN - theta0) / ncells;
[…]
