+++ assets/.doxy/gyselalibxx/gyselalibxx/classPolarSplineEvaluator.md
@@ -2,7 +2,7 @@
 
 # Class PolarSplineEvaluator
 
-**template &lt;class PolarBSplinesType, class OuterExtrapolationRule&gt;**
+**template &lt;class ExecSpace, class MemorySpace, class PolarBSplinesType, class OuterExtrapolationRule&gt;**
 
 
 
@@ -39,6 +39,8 @@ _Define an evaluator on polar B-splines._ [More...](#detailed-description)
 | typedef typename BSplinesR::continuous\_dimension\_type | [**DimR**](#typedef-dimr)  <br>_Tag the first dimension of the space._  |
 | typedef typename BSplinesTheta::continuous\_dimension\_type | [**DimTheta**](#typedef-dimtheta)  <br>_Tag the second dimension of the space._  |
 | typedef PolarBSplinesType | [**bsplines\_type**](#typedef-bsplines_type)  <br>_Tag the type of B-splines._  |
+| typedef ExecSpace | [**exec\_space**](#typedef-exec_space)  <br>_The type of the Kokkos execution space used by this class._  |
+| typedef MemorySpace | [**memory\_space**](#typedef-memory_space)  <br>_The type of the Kokkos memory space used by this class._  |
 
 
 
@@ -72,16 +74,20 @@ _Define an evaluator on polar B-splines._ [More...](#detailed-description)
 |   | [**PolarSplineEvaluator**](#function-polarsplineevaluator-24) (OuterExtrapolationRule const & outer\_bc) <br>_Instantiate a_ [_**PolarSplineEvaluator**_](classPolarSplineEvaluator.md) _with boundary evaluation conditions._ |
 |   | [**PolarSplineEvaluator**](#function-polarsplineevaluator-34) ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md) const & x) = default<br>_Instantiate a_ [_**PolarSplineEvaluator**_](classPolarSplineEvaluator.md) _from another._ |
 |   | [**PolarSplineEvaluator**](#function-polarsplineevaluator-44) ([**PolarSplineEvaluator**](classPolarSplineEvaluator.md) && x) = default<br>_Instantiate a_ [_**PolarSplineEvaluator**_](classPolarSplineEvaluator.md) _from another temporary._ |
-|  double | [**deriv\_1\_and\_2**](#function-deriv_1_and_2) (Coord&lt; [**DimR**](classPolarSplineEvaluator.md#typedef-dimr), [**DimTheta**](classPolarSplineEvaluator.md#typedef-dimtheta) &gt; coord\_eval, host\_t&lt; DConstField&lt; IdxRange&lt; PolarBSplinesType &gt; &gt; &gt; const spline\_coef) const<br>_Get the value of the cross derivative of the spline function._  |
-|  double | [**deriv\_dim\_1**](#function-deriv_dim_1-12) (Coord&lt; [**DimR**](classPolarSplineEvaluator.md#typedef-dimr), [**DimTheta**](classPolarSplineEvaluator.md#typedef-dimtheta) &gt; coord\_eval, host\_t&lt; DConstField&lt; IdxRange&lt; PolarBSplinesType &gt; &gt; &gt; const spline\_coef) const<br>_Get the value of the derivative of the spline function on the first dimension._  |
[…]
