### 2025-07-21 090204a
Uniformise toolchains (#382)
- Introduce missing shebangs
- Introduce missing guards
- Uniformise Kokkos Kernels options (in particular disables ETI that
should speed up some toolchains)
- Uniformise Kokkos options (in particular for Leonardo)
- Use `CMAKE_CXX_FLAGS_INIT` instead of `CMAKE_CXX_FLAGS` in toolchains

Note that the Leonardo toolchain is likely not working anymore. For
example the `mdspan` package is not part anymore of the pdi Spack
repository.
