+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change spack setup in CPU installation script (`prepare.sh`) to create and use independent spack installation.
 - Change template parameters of `PolarSplineEvaluator` to add execution and memory space information.
 - Allow `get_idx_range` to be called from a GPU execution space.
+- Uniformise toolchains.
 
 ### Deprecated
 
