+++ toolchains/common_toolchains/ci_coverage_toolchain.cmake
@@ -5,7 +5,7 @@ include(${CMAKE_CURRENT_LIST_DIR}/importable_defaults.cmake)
 set(CMAKE_BUILD_TYPE Debug)
 
 # Compiler options
-set(CMAKE_CXX_FLAGS "-Og --coverage -fprofile-update=atomic -g")
+set(CMAKE_CXX_FLAGS_INIT "-Og --coverage -fprofile-update=atomic -g")
 
 # Kokkos options
 set(Kokkos_ENABLE_SERIAL ON CACHE BOOL "Allow serial code to run" FORCE)
