### 2025-07-21 a8a0f39
Move packages and compilers inside environment definition
