+++ toolchains/mi250.hipcc.adastra.spack/packages.yaml
@@ -1,358 +1,0 @@
-# -------------------------------------------------------------------------
-# This file controls default concretization preferences for Spack, and
-# will serve as the basis for the configuration of the Adastra packages
-#
-# Settings here are versioned with Spack and are intended to provide
-# sensible defaults out of the box. Spack maintainers should edit this
-# file to keep it current.
-#
-# Users can override these settings by editing the following files.
-#
-# Per-spack-instance settings (overrides defaults):
-#   $SPACK_ROOT/etc/spack/packages.yaml
-#
-# Per-user settings (overrides default and site settings):
-#   ~/.spack/packages.yaml
-# -------------------------------------------------------------------------
-packages:
-  all:
-    compiler: [gcc@13.2.1.generic, gcc@13.2.1.genoa, gcc@13.2.1.mi250, gcc@13.2.0.generic, gcc@13.2.0.genoa, cce@17.0.0.genoa, aocc@3.2.0.genoa,
-      rocmcc@6.3.3.genoa, rocmcc@5.7.1.genoa, cce@17.0.0.mi250, rocmcc@6.3.3.mi250,
-      rocmcc@5.7.1.mi250, gcc@13.2.0.mi250, aocc@3.2.0.mi250, cce@16.0.0.genoa, cce@15.0.0.genoa,
-      cce@16.0.0.mi250, gcc@12.1.generic, gcc@11.2.generic, gcc@12.1.genoa, gcc@11.2.genoa,
-      gcc@12.1.mi250, gcc@11.2.mi250, cce@18.0.0.mi300, gcc@13.2.mi300]
-    target: [zen3, zen4]
-    permissions:
-      write: group
-    providers:
-      mpi:
-      - cray-mpich
-      # version 1.0.0 seems to struggle to find cray-libsci/24.07 library, here the error message, at install time :
-      # Error: NoLibrariesError: Unable to recursively locate cray-libsci libraries in /opt/cray/pe/libsci/24.07.0/CRAYCLANG/18.0/x86_64
-      blas: [cray-libsci, openblas, acc_cray-libsci] 
-      lapack: [cray-libsci, openblas, acc_cray-libsci]
-      blacs: [cray-libsci, acc_cray-libsci]
-      scalapack: [cray-libsci]
-      netlib-scalapack: [cray-libsci] 
-      mkl: [cray-libsci]
-      fftw:
-      - cray-fftw
-      opencl:
-      - rocm-opencl@6.3.3
-      opencl-runtime:
-      - rocm-opencl-runtime@6.3.3
-    require:
-    # GPU parameter propagation:
-    # - Avoids handling cray-mpich special case in each MPI package
-    # - Uses package constraint to automatically propagate GPU variants
-    # For MI250X:
-    - one_of: [amdgpu_target=gfx90a, amdgpu_target=gfx942]
-      when: +rocm
-    - spec: +rocm amdgpu_target=gfx90a ^cray-mpich+rocm amdgpu_target=gfx90a
-      when: +rocm amdgpu_target=gfx90a ^cray-mpich
-    # For MI300A:
-    - spec: +rocm amdgpu_target=gfx942 ^cray-mpich+rocm amdgpu_target=gfx942
-      when: +rocm amdgpu_target=gfx942 ^cray-mpich
-    # TODO(Etienne M): Fix all packages requiring MPICH + rocm to effectively
-    # propagate the variant.
-  acc_cray-libsci:
-    buildable: false
-    version: [24.07.0]
-    externals:
[…]

[…]
