### 2025-07-22 b953443
Simplify packages
