+++ toolchains/mi250.hipcc.adastra.spack/gyselalibxx-spack-environment.yaml
@@ -21,7 +21,6 @@ spack:
           - libfabric/1.23.1
           - craype-network-ofi
           - cray-mpich/8.1.30
-          - cray-libsci/24.07.0
           - PrgEnv-gnu/8.5.0
           - amd-mixed/6.3.3
         environment: {}
@@ -78,43 +77,17 @@ spack:
           - cray-mpich
         # version 1.0.0 seems to struggle to find cray-libsci/24.07 library, here the error message, at install time :
         # Error: NoLibrariesError: Unable to recursively locate cray-libsci libraries in /opt/cray/pe/libsci/24.07.0/CRAYCLANG/18.0/x86_64
-        blas: [cray-libsci, openblas, acc_cray-libsci]
-        lapack: [cray-libsci, openblas, acc_cray-libsci]
-        blacs: [cray-libsci, acc_cray-libsci]
-        scalapack: [cray-libsci]
-        netlib-scalapack: [cray-libsci]
-        mkl: [cray-libsci]
-        fftw:
-          - cray-fftw
+        blas: [openblas]
+        lapack: [openblas]
+        fftw-api:
+          - fftw
         opencl:
           - rocm-opencl@6.3.3
         opencl-runtime:
           - rocm-opencl-runtime@6.3.3
       require:
-        # GPU parameter propagation:
-        # - Avoids handling cray-mpich special case in each MPI package
-        # - Uses package constraint to automatically propagate GPU variants
-        # For MI250X:
         - one_of: [amdgpu_target=gfx90a, amdgpu_target=gfx942]
           when: +rocm
-        - spec: +rocm amdgpu_target=gfx90a ^cray-mpich+rocm amdgpu_target=gfx90a
-          when: +rocm amdgpu_target=gfx90a ^cray-mpich
-        # For MI300A:
-        - spec: +rocm amdgpu_target=gfx942 ^cray-mpich+rocm amdgpu_target=gfx942
-          when: +rocm amdgpu_target=gfx942 ^cray-mpich
-          # TODO(Etienne M): Fix all packages requiring MPICH + rocm to effectively
-          # propagate the variant.
-    acc_cray-libsci:
-      buildable: false
-      version: [24.07.0]
-      externals:
-        - spec: acc_cray-libsci@24.07.0
-          modules:
-            - cray-libsci_acc/24.07.0
-            # The GPU version depends on the CPU version.
-            - cray-libsci/24.07.0
-            - amd-mixed/6.3.3
-            - craype-accel-amd-gfx90a
     binutils:
       version: [2.38]
       externals:
@@ -129,30 +102,6 @@ spack:
         - spec: cray-mpich@8.1.30
           modules:
             - cray-mpich/8.1.30
-    cray-libsci:
-      buildable: false
-      version: [23.12.5, 24.07.0]
-      externals:
-        - spec: cray-libsci@23.12.5
-          modules:
-            - cray-libsci/23.12.5
-        - spec: cray-libsci@24.07.0  # When i tried with this one I have this warning : Warning: Extracted path from     module does not exist [module=cray-libsci/24.07.0, path=module]
-          modules:
-            - cray-libsci/24.07.0
-    cray-fftw:
-      buildable: false
-      version: [3.3.10.8]
-      externals:
-        - spec: cray-fftw@3.3.10.8
-          modules:
-            - cray-fftw/3.3.10.8
-    cuda:
-      buildable: false
-      version:
-        - '11.7'
-      externals:
-        - spec: cuda@11.7 arch=linux-rhel8-zen3
[…]
