+++ toolchains/mi250.hipcc.adastra.spack/gyselalibxx-spack-environment.yaml
@@ -12,24 +12,20 @@ spack:
         operating_system: rhel8
         target: x86_64
         modules:
-          - cpe/24.07
-          - craype-accel-amd-gfx90a
-          - craype-x86-trento
           - gcc-native/13.2
+          - amd-mixed/6.3.3
           - craype/2.7.32
           - cray-dsmml/0.3.0
           - libfabric/1.23.1
           - craype-network-ofi
           - cray-mpich/8.1.30
-          - PrgEnv-gnu/8.5.0
-          - amd-mixed/6.3.3
         environment: {}
         extra_rpaths: []
   definitions:
     - packages:
         - cmake
         - fftw
-        - ginkgo~openmp+rocm amdgpu_target=gfx90a
+        - ginkgo~openmp+rocm
         - googletest
         - ninja
         - paraconf
