### 2025-07-24 79763e3
Remove overwrite of previous initialisation value to reduce number of required iterations
