### 2025-07-24 f2d7307
Use logical default value
