### 2025-07-24 8a5cde3
Allow batched solver convergence parameters to be passed as input
