### 2025-07-24 4a8d8c8
Improve matrix handling for polar poisson solver (#386)
Add the batch CSR convergence parameters to the constructor of the polar
poisson solver so they can be modified in different ways for different
simulations. Remove the line `Kokkos::deep_copy(m_x_init, x_init_host);`
which prevents the previous solution being used as an initial guess for
the convergence.
Fix typo `preconditionner`->`preconditioner`.
