+++ assets/.doxy/gyselalibxx/gyselalibxx/classMatrixBatchCsr.md
@@ -82,8 +82,8 @@ See [MatrixBatch](classMatrixBatch.md)
 
 | Type | Name |
 | ---: | :--- |
-|   | [**MatrixBatchCsr**](#function-matrixbatchcsr-12) (const int batch\_size, const int mat\_size, const int nnz\_per\_system, std::optional&lt; int &gt; max\_iter=std::nullopt, std::optional&lt; double &gt; res\_tol=std::nullopt, std::optional&lt; bool &gt; logger=std::nullopt, std::optional&lt; int &gt; preconditionner\_max\_block\_size=1u) <br>_The constructor for_ [_**MatrixBatchCsr**_](classMatrixBatchCsr.md) _class._ |
-|   | [**MatrixBatchCsr**](#function-matrixbatchcsr-22) (Kokkos::View&lt; double \*\*, Kokkos::LayoutRight, ExecSpace &gt; batch\_values, Kokkos::View&lt; int \*, Kokkos::LayoutRight, ExecSpace &gt; cols\_idx, Kokkos::View&lt; int \*, Kokkos::LayoutRight, ExecSpace &gt; nnz\_per\_row, std::optional&lt; int &gt; max\_iter=std::nullopt, std::optional&lt; double &gt; res\_tol=std::nullopt, std::optional&lt; bool &gt; logger=std::nullopt, std::optional&lt; int &gt; preconditionner\_max\_block\_size=1u) <br>_Constructor for_ [_**MatrixBatchCsr**_](classMatrixBatchCsr.md) _class._ |
+|   | [**MatrixBatchCsr**](#function-matrixbatchcsr-12) (const int batch\_size, const int mat\_size, const int nnz\_per\_system, std::optional&lt; int &gt; max\_iter=std::nullopt, std::optional&lt; double &gt; res\_tol=std::nullopt, std::optional&lt; bool &gt; logger=std::nullopt, std::optional&lt; int &gt; preconditioner\_max\_block\_size=std::nullopt) <br>_The constructor for_ [_**MatrixBatchCsr**_](classMatrixBatchCsr.md) _class._ |
+|   | [**MatrixBatchCsr**](#function-matrixbatchcsr-22) (Kokkos::View&lt; double \*\*, Kokkos::LayoutRight, ExecSpace &gt; batch\_values, Kokkos::View&lt; int \*, Kokkos::LayoutRight, ExecSpace &gt; cols\_idx, Kokkos::View&lt; int \*, Kokkos::LayoutRight, ExecSpace &gt; nnz\_per\_row, std::optional&lt; int &gt; max\_iter=std::nullopt, std::optional&lt; double &gt; res\_tol=std::nullopt, std::optional&lt; bool &gt; logger=std::nullopt, std::optional&lt; int &gt; preconditioner\_max\_block\_size=std::nullopt) <br>_Constructor for_ [_**MatrixBatchCsr**_](classMatrixBatchCsr.md) _class._ |
 |  std::tuple&lt; Kokkos::View&lt; double \*\*, Kokkos::LayoutRight, ExecSpace &gt;, Kokkos::View&lt; int \*, Kokkos::LayoutRight, ExecSpace &gt;, Kokkos::View&lt; int \*, Kokkos::LayoutRight, ExecSpace &gt; &gt; | [**get\_batch\_csr**](#function-get_batch_csr) () <br>_A function to update information about values,indices and the number of non-zero per row for the whole batch. Data is managed by Kokkos Views stored on the host._  |
 |  double | [**norm**](#function-norm) (int batch\_idx) const<br>_A function returning the norm of a matrix located at batch\_idx._  |
[…]
