### 2025-07-24 08b6572
Add GAMs to dictionary (#387)
