+++ .typos.toml
@@ -31,6 +31,7 @@ ax = "ax"
 color = "color"
 colors = "colors"
 artifacts = "artifacts"
+GAM = "GAM"
 
 [files]
 extend-exclude = ["*.tex",
