### 2025-07-27 7f5549f
External packages and compilers files
