+++ toolchains/mi250.hipcc.adastra.spack/packages.yaml
@@ -1,0 +1,239 @@
+---
+packages:
+  all:
+    compiler: [gcc@13.2.1.mi250.raw]
+    target: [zen3, zen4]
+    providers:
+      mpi: [cray-mpich]
+      blas: [openblas]
+      lapack: [openblas]
+      fftw-api: [fftw]
+      opencl: [rocm-opencl@6.3.3]
+      opencl-runtime: [rocm-opencl-runtime@6.3.3]
+    require:
+      - one_of: [amdgpu_target=gfx90a, amdgpu_target=gfx942]
+        when: +rocm
+  cray-mpich:
+    buildable: false
+    version: [8.1.30]
+    externals:
+      - spec: cray-mpich@8.1.30
+        prefix: /opt/cray/pe/mpich/8.1.30/ofi/gnu/11.2
+        modules:
+          - gcc-native/13.2
+          - craype-network-ofi
+          - craype/2.7.32
+          - cray-mpich/8.1.30
+  hip:
+    buildable: false
+    externals:
+      - extra_attributes:
+          compilers:
+            c: /opt/rocm-6.3.3/llvm/bin/clang++
+            c++: /opt/rocm-6.3.3/llvm/bin/clang++
+            hip: /opt/rocm-6.3.3/bin/hipcc
+        prefix: /opt/rocm-6.3.3
+        spec: hip@6.3.3
+    version: [6.3.3]
+  hipblas:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hipblas@6.3.3
+    version: [6.3.3]
+  hipcub:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hipcub@6.3.3
+    version: [6.3.3]
+  hipfft:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hipfft@6.3.3
+    version: [6.3.3]
+  hipfort:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hipfort@6.3.3
+    version: [6.3.3]
+  hipify-clang:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hipify-clang@6.3.3
+    version: [6.3.3]
+  hiprand:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hiprand@6.3.3
+    version: [6.3.3]
+  hip-rocclr:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3/rocclr
+        spec: hip-rocclr@6.3.3
+    version: [6.3.3]
+  hipsolver:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hipsolver@6.3.3
+    version: [6.3.3]
+  hipsparse:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hipsparse@6.3.3
+    version: [6.3.3]
+  hsa-rocr-dev:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hsa-rocr-dev@6.3.3
+        extra_attributes:
+          compilers:
+            c: /opt/rocm-6.3.3/llvm/bin/clang++
+            cxx: /opt/rocm-6.3.3/llvm/bin/clang++
+    version: [6.3.3]
+  hsakmt-roct:
+    buildable: false
+    externals:
+      - prefix: /opt/rocm-6.3.3
+        spec: hsakmt-roct@6.3.3
+    version: [6.3.3]
+  libfabric:
+    buildable: false
[…]
