+++ toolchains/mi250.hipcc.adastra.spack/gyselalibxx-spack-environment.yaml
@@ -1,54 +1,39 @@
----
 spack:
-  compilers:
-    - compiler:
-        spec: gcc@13.2.1.mi250.raw
-        paths:
-          cc: /opt/rh/gcc-toolset-13/root/usr/bin/gcc
-          cxx: /opt/rh/gcc-toolset-13/root/usr/bin/g++
-          f77: /opt/rh/gcc-toolset-13/root/usr/bin/gfortran
-          fc: /opt/rh/gcc-toolset-13/root/usr/bin/gfortran
-        flags: {}
-        operating_system: rhel8
-        target: x86_64
-        modules: []
-        environment: {}
-        extra_rpaths: []
   definitions:
-    - packages:
-        - cmake
-        - fftw
-        - ginkgo~openmp+rocm
-        - googletest
-        - ninja
-        - paraconf
-        - pdi
-        - pdiplugin-decl-hdf5
-        - pdiplugin-decl-netcdf
-        - pdiplugin-mpi
-        - pdiplugin-set-value
-        - pdiplugin-trace
-        - python
+  - packages:
+    - cmake
+    - fftw
+    - ginkgo~openmp+rocm amdgpu_target=gfx90a
+    - googletest
+    - ninja
+    - paraconf
+    - pdi
+    - pdiplugin-decl-hdf5
+    - pdiplugin-decl-netcdf
+    - pdiplugin-mpi
+    - pdiplugin-set-value
+    - pdiplugin-trace
+    - python
 
-        - py-dask
-        - py-h5py
-        - py-imageio
-        - py-matplotlib
-        - py-netcdf4
-        - py-numpy
-        - py-scipy
-        - py-sympy
-        - py-xarray
-        - py-pyyaml
-    - compilers:
-        - gcc@13.2.1.mi250.raw
-    - architectures:
-        - arch=linux-rhel8-zen3
+    - py-dask
+    - py-h5py
+    - py-imageio
+    - py-matplotlib
+    - py-netcdf4
+    - py-numpy
+    - py-scipy
+    - py-sympy
+    - py-xarray
+    - py-pyyaml
+  - compilers:
+    - gcc@13.2.1.mi250
+  - architectures:
+    - arch=linux-rhel8-zen3
   specs:
-    - matrix:
-        - [$packages]
-        - [$%compilers]
-        - [$architectures]
+  - matrix:
+    - [$packages]
+    - [$%compilers]
+    - [$architectures]
   concretizer:
     unify: true
     reuse: false
@@ -59,248 +44,6 @@ spack:
       permissions:
         read: world
         write: group
-      compiler: [gcc@13.2.1.mi250.raw]
-      target: [zen3, zen4]
-      providers:
-        mpi:
-          - cray-mpich
-        # version 1.0.0 seems to struggle to find cray-libsci/24.07 library, here the error message, at install time :
-        # Error: NoLibrariesError: Unable to recursively locate cray-libsci libraries in /opt/cray/pe/libsci/24.07.0/CRAYCLANG/18.0/x86_64
-        blas: [openblas]
-        lapack: [openblas]
-        fftw-api:
-          - fftw
-        opencl:
-          - rocm-opencl@6.3.3
-        opencl-runtime:
-          - rocm-opencl-runtime@6.3.3
-      require:
-        - one_of: [amdgpu_target=gfx90a, amdgpu_target=gfx942]
-          when: +rocm
-    cray-mpich:
-      buildable: false
-      version: [8.1.30]
-      externals:
-        - spec: cray-mpich@8.1.30
-          prefix: /opt/cray/pe/mpich/8.1.30/ofi/gnu/11.2
-          modules:
-            - gcc-native/13.2
-            - craype-network-ofi
[…]
