+++ src/multipatch/interface_derivatives/interface_derivative_matrix.hpp
@@ -58,8 +58,9 @@
 template <
         class Connectivity,
         class Grid1D,
-        template <typename P>
-        typename ValuesOnPatch, // Fix type ?
+        // template <typename P>
+        // typename ValuesOnPatch, // Fix type ?
+        class SingleInterfaceDerivativesCalculatorCollectionType,
         ddc::BoundCond LowerBound = ddc::BoundCond::HERMITE,
         ddc::BoundCond UpperBound = ddc::BoundCond::HERMITE,
         class ExecSpace = Kokkos::DefaultHostExecutionSpace,
@@ -187,9 +188,9 @@ class InterfaceDerivativeMatrix
     using DerivFieldOnPatch = DerivField<
             double,
             IdxRange<
-                    ddc::Deriv<typename Patch::Grid1>,
-                    ddc::Deriv<typename Patch::Grid2>,
+                    ddc::Deriv<typename Patch::Dim1>,
                     typename Patch::Grid1,
+                    ddc::Deriv<typename Patch::Dim2>,
                     typename Patch::Grid2>>;
 
     template <class Patch>
@@ -309,10 +310,12 @@ private:
 
     MultipatchType<IdxRangeOnPatch, Patches...> const& m_idx_ranges;
 
-    // TODO: WARNING, BE ABLE TO DEAL WITH DIFFERENT ORIENTATION INTERFACES.
+    // TODO: WARNING, BE ABLE TO DEAL WITH DIFFERENT ORIENTATION/ORDER INTERFACES.
     using SingleInterfaceDerivativesCalculatorTuple
             = get_tuple_deriv_calculator_t<inner_interface_collection>;
-    SingleInterfaceDerivativesCalculatorTuple const& m_derivatives_calculators;
+    // SingleInterfaceDerivativesCalculatorTuple const& m_derivatives_calculators;
+
+    SingleInterfaceDerivativesCalculatorCollectionType const& m_derivatives_calculators;
 
 public:
     ~InterfaceDerivativeMatrix() {}
@@ -331,9 +334,41 @@ public:
      * @warning The interface derivative calculators have to be ordered in the tuple. 
      *      The order has to be the same as the interfaces in the Grid1D direction. 
      */
+    // InterfaceDerivativeMatrix(
+    //         MultipatchType<IdxRangeOnPatch, Patches...> const& idx_ranges,
+    //         SingleInterfaceDerivativesCalculatorTuple const derivatives_calculators,
+    //         double const& reduction_factor = 1e-6)
+    //     : m_idx_ranges(idx_ranges)
+    //     , m_derivatives_calculators(derivatives_calculators)
+    // {
+    //     const auto exec = gko::ReferenceExecutor::create();
+    //     m_matrix = gko::share(Matrix::create(exec, gko::dim<2>(n_inner_interfaces)));
+
+    //     set_matrix(std::make_integer_sequence<std::size_t, n_inner_interfaces> {});
+
+    //     // Define a solver.
+    //     auto solver_factory
+    //             = SolverCG::build()
+    //                       .with_criteria(
+    //                               gko::stop::Iteration::build()
+    //                                       .with_max_iters(n_inner_interfaces)
+    //                                       .on(exec),
+    //                               gko::stop::ResidualNorm<>::build()
+    //                                       .with_reduction_factor(reduction_factor)
+    //                                       .on(exec))
+    //                       .with_preconditioner(
+    //                               PreconditionerJ::build().with_max_block_size(8u).on(exec))
+    //                       .on(exec);
+    //     // Create solver
+    //     m_solver = std::move(solver_factory->generate(m_matrix));
+
[…]
