+++ tests/multipatch/interface_derivatives/interface_derivative_matrix_test.cpp
@@ -26,7 +26,7 @@
         |  1  |  2  |  3  |  1 ...
         -------------------
         |  4  |  5  |  6  |  4 ...
-        -------------------
+        -------------------b
         |  7  |  8  |  9  |  7 ... 
 
         with the global X dimension periodic and the global Y spline
@@ -137,11 +137,11 @@ template <class Patch>
 using DerivFieldMemOnPatch = DerivFieldMem<
         double,
         IdxRange<
-                ddc::Deriv<typename Patch::Grid1>,
-                ddc::Deriv<typename Patch::Grid2>,
+                ddc::Deriv<typename Patch::Dim1>,
                 typename Patch::Grid1,
+                ddc::Deriv<typename Patch::Dim2>,
                 typename Patch::Grid2>,
-        2>;
+        1>;
 
 template <class Patch>
 using DerivFieldMemOnPatch_host = host_t<DerivFieldMemOnPatch<Patch>>;
@@ -150,9 +150,9 @@ template <class Patch>
 using DerivFieldOnPatch = DerivField<
         double,
         IdxRange<
-                ddc::Deriv<typename Patch::Grid1>,
-                ddc::Deriv<typename Patch::Grid2>,
+                ddc::Deriv<typename Patch::Dim1>,
                 typename Patch::Grid1,
+                ddc::Deriv<typename Patch::Dim2>,
                 typename Patch::Grid2>>;
 
 template <class Patch>
@@ -175,8 +175,8 @@ Coord<Xg, Yg> get_global_coord(Coord<X, Y> const& local_coord)
 }
 
 /// @brief Initialise the function with f(x,y) = cos(2/3*pi*x)sin(y).
-template <class Grid1, class Grid2>
-void initialise_2D_function(host_t<DField<IdxRange<Grid1, Grid2>>> function)
+template <class Grid1, class Grid2, class Layout>
+void initialise_2D_function(DField<IdxRange<Grid1, Grid2>, Kokkos::HostSpace, Layout> function)
 {
     ddc::for_each(get_idx_range(function), [&](Idx<Grid1, Grid2> idx) {
         // Get the coordinate on the equivalent global mesh.
@@ -187,153 +187,162 @@ void initialise_2D_function(host_t<DField<IdxRange<Grid1, Grid2>>> function)
 }
 
 /// @brief Initialise all the functions defined on the patches.
+// template <class... Patches>
+// void initialise_all_functions(MultipatchField<DFieldOnPatch_host, Patches...> const& functions)
+// {
+//     (initialise_2D_function<typename Patches::Grid1, typename Patches::Grid2>(
+//              functions.template get<Patches>()),
+//      ...);
+// }
+
 template <class... Patches>
-void initialise_all_functions(MultipatchField<DFieldOnPatch_host, Patches...> const& functions)
+void initialise_all_functions(
+        MultipatchField<DerivFieldOnPatch_host, Patches...> const& functions_and_derivs)
 {
     (initialise_2D_function<typename Patches::Grid1, typename Patches::Grid2>(
-             functions.template get<Patches>()),
+             (functions_and_derivs.template get<Patches>()).get_values_field()),
      ...);
 }
 
-/// @brief Initialise the y-derivatives from the global spline.
-template <class Grid1, class Grid2>
-void initialise_2D_y_derivative(
-        host_t<DField<IdxRange<Grid1, ddc::Deriv<typename Grid2::continuous_dimension_type>>>>
-                deriv_y,
-        Idx<Grid2> const& idx_y,
-        SplineRThetagEvaluator const& evaluator_g,
-        host_t<DConstField<IdxRange<BSplinesXg, BSplinesYg>>> const& function_g_coef)
-{
[…]
