+++ src/data_types/derivative_field_common.hpp
@@ -274,6 +274,12 @@ protected:
                                ddc::Deriv<typename QueryDDim::continuous_dimension_type>,
                                deriv_tags>)) {
                     // If the derivative is being requested
+                    std::cout << IdxRangeSlice<QueryDDim>(m_cross_derivative_idx_range)
+                                         .contains(idx_range_requested.front())
+                              << std::endl;
+                    std::cout << (idx_range_requested.extents() == 1) << std::endl;
+                    std::cout << (IdxRangeSlice<QueryDDim>(m_cross_derivative_idx_range).strides()
+                            == 1) << std::endl;
                     assert(IdxRangeSlice<QueryDDim>(m_cross_derivative_idx_range)
                                    .contains(idx_range_requested.front())
                            && (idx_range_requested.extents() == 1
