### 2025-07-29 ef7bfd8
Manage periodicity
