### 2025-07-29 94864a5
Add get_vals_and_derivs method to calculate bspline values and derivatives on the fly
