+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -123,6 +123,7 @@ private:
 
     using IdxStepBSR = IdxStep<BSplinesR>;
     using IdxStepBSTheta = IdxStep<BSplinesTheta>;
+    using IdxStepBSRTheta = IdxStep<BSplinesR, BSplinesTheta>;
 
     using IdxRangeBatchedBSRTheta
             = ddc::detail::convert_type_seq_to_discrete_domain_t<ddc::type_seq_replace_t<
@@ -551,10 +552,10 @@ public:
                             Idx<QDimRMesh> const idx_r(idx_quad);
                             Idx<QDimThetaMesh> const idx_theta(idx_quad);
                             return weak_integral_element(
+                                    idx_test,
+                                    idx_trial,
                                     idx_r,
                                     idx_theta,
-                                    singular_basis_vals_and_derivs(idx_test, idx_r, idx_theta),
-                                    singular_basis_vals_and_derivs(idx_trial, idx_r, idx_theta),
                                     coeff_alpha,
                                     coeff_beta,
                                     spline_evaluator,
@@ -680,14 +681,10 @@ public:
                                     IdxQuadratureR const idx_r(idx_quad);
                                     IdxQuadratureTheta const idx_theta(idx_quad);
                                     return weak_integral_element<Mapping>(
+                                            idx_test,
+                                            idx_trial_polar,
                                             idx_r,
                                             idx_theta,
-                                            singular_basis_vals_and_derivs(
-                                                    idx_test,
-                                                    idx_r,
-                                                    idx_theta),
-                                            r_basis_vals_and_derivs(ib_trial_r, idx_r),
-                                            theta_basis_vals_and_derivs(ib_trial_theta, idx_theta),
                                             coeff_alpha,
                                             coeff_beta,
                                             spline_evaluator,
@@ -1029,8 +1026,8 @@ public:
      * @return 
      *      The quadrature range corresponding to the  @f$(r,\theta)@f$ indices.
      */
-    static KOKKOS_FUNCTION IdxRangeQuadratureRTheta
-    get_quadrature_points_in_cell(int cell_idx_r, int cell_idx_theta)
+    KOKKOS_FUNCTION IdxRangeQuadratureRTheta
+    get_quadrature_points_in_cell(int cell_idx_r, int cell_idx_theta) const
     {
         const IdxQuadratureR first_quad_point_r(cell_idx_r * s_n_gauss_legendre_r);
         const IdxQuadratureTheta first_quad_point_theta(cell_idx_theta * s_n_gauss_legendre_theta);
@@ -1041,142 +1038,6 @@ public:
         return IdxRangeQuadratureRTheta(quad_points_r, quad_points_theta);
     }
 
-    /**
-     * @brief compute the weak integral value.
-     *
-     * @param[in] idx_r
-     *      The index for radial direction.
-     * @param[in] idx_theta
-     *      The index for poloidal direction
-     * @param[in] test_bspline_val_and_deriv
-     *      The data structure containing the derivatives over radial and poloidal directions for test space.
-     * @param[in] trial_bspline_val_and_deriv
-     *      The data structure containing the derivatives over radial and poloidal directions for trial space.
-     * @param[in] coeff_alpha
-     *      The spline representation of the @f$ \alpha @f$ function in the
-     *      definition of the Poisson-like equation.
[…]
