### 2025-07-29 4e1578a
Remove calculated values
