+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -246,13 +246,6 @@ private:
     FieldMem<double, IdxRangeQuadratureR> m_weights_r;
     FieldMem<double, IdxRangeQuadratureTheta> m_weights_theta;
 
-    // Basis Spline values and derivatives at Gauss-Legendre points
-    host_t<FieldMem<EvalDeriv2DType, IdxRange<PolarBSplinesRTheta, QDimRMesh, QDimThetaMesh>>>
-            m_singular_basis_vals_and_derivs;
-    host_t<FieldMem<EvalDeriv1DType, IdxRange<RBasisSubset, QDimRMesh>>> m_r_basis_vals_and_derivs;
-    host_t<FieldMem<EvalDeriv1DType, IdxRange<ThetaBasisSubset, QDimThetaMesh>>>
-            m_theta_basis_vals_and_derivs;
-
     FieldMem<double, IdxRangeQuadratureRTheta> m_int_volume;
 
     PolarSplineEvaluator<PolarBSplinesRTheta, ddc::NullExtrapolationRule> m_polar_spline_evaluator;
@@ -316,16 +309,6 @@ public:
                   m_idxrange_quadrature_theta)
         , m_weights_r(m_idxrange_quadrature_r)
         , m_weights_theta(m_idxrange_quadrature_theta)
-        , m_singular_basis_vals_and_derivs(IdxRange<PolarBSplinesRTheta, QDimRMesh, QDimThetaMesh>(
-                  PolarBSplinesRTheta::template singular_idx_range<PolarBSplinesRTheta>(),
-                  ddc::select<QDimRMesh>(m_idxrange_quadrature_singular),
-                  ddc::select<QDimThetaMesh>(m_idxrange_quadrature_singular)))
-        , m_r_basis_vals_and_derivs(
-                  IdxRange<RBasisSubset, QDimRMesh>(m_non_zero_bases_r, m_idxrange_quadrature_r))
-        , m_theta_basis_vals_and_derivs(
-                  IdxRange<
-                          ThetaBasisSubset,
-                          QDimThetaMesh>(m_non_zero_bases_theta, m_idxrange_quadrature_theta))
         , m_polar_spline_evaluator(ddc::NullExtrapolationRule())
         , m_phi_spline_coef(ddc::discrete_space<PolarBSplinesRTheta>().full_domain())
         , m_x_init(
@@ -356,74 +339,6 @@ public:
                 mapping,
                 gauss_legendre_quadrature_coefficients<
                         Kokkos::DefaultExecutionSpace>(gl_coeffs_r, gl_coeffs_theta));
-
-        // Find value and derivative of 1D B-splines in radial direction
-        ddc::for_each(m_idxrange_quadrature_r, [&](IdxQuadratureR const idx_r) {
-            std::array<double, 2 * m_n_non_zero_bases_r> data;
-            DSpan2D vals(data.data(), m_n_non_zero_bases_r, 2);
-            ddc::discrete_space<BSplinesR>()
-                    .eval_basis_and_n_derivs(vals, ddc::coordinate(idx_r), 1);
-            for (auto ib : m_non_zero_bases_r) {
-                const int ib_idx = ib - m_non_zero_bases_r.front();
-                m_r_basis_vals_and_derivs(ib, idx_r).value = vals(ib_idx, 0);
-                m_r_basis_vals_and_derivs(ib, idx_r).derivative = vals(ib_idx, 1);
-            }
-        });
-
-        // Find value and derivative of 1D B-splines in poloidal direction
-        ddc::for_each(m_idxrange_quadrature_theta, [&](IdxQuadratureTheta const idx_theta) {
-            std::array<double, 2 * m_n_non_zero_bases_theta> data;
[…]
