### 2025-07-29 a287e86
Remove unused variables. Reduce duplication of long expression
