+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -222,11 +222,6 @@ private:
     static constexpr IdxStep<ThetaBasisSubset> m_n_non_zero_bases_theta
             = IdxStep<ThetaBasisSubset>(BSplinesTheta::degree() + 1);
 
-    static constexpr IdxRange<RBasisSubset> m_non_zero_bases_r
-            = IdxRange<RBasisSubset>(Idx<RBasisSubset> {0}, m_n_non_zero_bases_r);
-    static constexpr IdxRange<ThetaBasisSubset> m_non_zero_bases_theta
-            = IdxRange<ThetaBasisSubset>(Idx<ThetaBasisSubset> {0}, m_n_non_zero_bases_theta);
-
     const int m_nbasis_r;
     const int m_nbasis_theta;
 
@@ -529,9 +524,7 @@ public:
         // Calculate the matrix elements where bspline products overlap the B-splines which cover the singular point
         ddc::for_each(idxrange_singular, [&](IdxBSPolar const idx_test) {
             ddc::for_each(idxrange_non_singular_near_centre, [&](IdxBSRTheta const idx_trial) {
-                const IdxBSPolar idx_trial_polar(
-                        PolarBSplinesRTheta::template get_polar_index<PolarBSplinesRTheta>(
-                                idx_trial));
+                const IdxBSPolar idx_trial_polar(to_polar(idx_trial));
                 const Idx<BSplinesR> idx_trial_r(ddc::select<BSplinesR>(idx_trial));
                 const Idx<BSplinesTheta> idx_trial_theta(ddc::select<BSplinesTheta>(idx_trial));
 
@@ -562,10 +555,6 @@ public:
 
                         const IdxRangeQuadratureRTheta cell_quad_points(
                                 get_quadrature_points_in_cell(cell_idx_r, cell_idx_theta));
-                        // Find the column where the non-zero data is stored
-                        Idx<RBasisSubset> ib_trial_r(idx_trial_r.uid() - cell_idx_r);
-                        Idx<ThetaBasisSubset> ib_trial_theta(
-                                theta_mod(idx_trial_theta.uid() - cell_idx_theta));
                         // Calculate the weak integral
                         element += ddc::parallel_transform_reduce(
                                 Kokkos::DefaultExecutionSpace(),
@@ -651,8 +640,7 @@ public:
             ddc::for_each(remaining_theta, [&](Idx<BSplinesTheta> const idx_trial_theta) {
                 IdxBSRTheta idx_trial(Idx<BSplinesR>(idx_test_r), idx_trial_theta);
                 IdxBSPolar idx_trial_polar(
-                        PolarBSplinesRTheta::template get_polar_index<PolarBSplinesRTheta>(
-                                IdxBSRTheta(idx_test_r, theta_mod(idx_trial_theta.uid()))));
+                        to_polar(IdxBSRTheta(idx_test_r, theta_mod(idx_trial_theta.uid()))));
                 double element = get_matrix_stencil_element(
                         idx_test,
                         idx_trial,
@@ -697,8 +685,7 @@ public:
                 const int idx_trial_r(ddc::select<BSplinesR>(idx_trial).uid());
                 const int idx_trial_theta(ddc::select<BSplinesTheta>(idx_trial).uid());
                 IdxBSPolar idx_trial_polar(
-                        PolarBSplinesRTheta::template get_polar_index<PolarBSplinesRTheta>(
-                                IdxBSRTheta(idx_trial_r, theta_mod(idx_trial_theta))));
+                        to_polar(IdxBSRTheta(idx_trial_r, theta_mod(idx_trial_theta))));
                 double element = get_matrix_stencil_element(
                         idx_test,
                         idx_trial,
@@ -847,7 +834,8 @@ public:
                             IdxQuadratureR const idx_r(idx_quad);
[…]
