### 2025-07-29 cfc39ae
Simplify
