### 2025-07-29 dc4dc7a
Remove RBasisSubset and ThetaBasisSubset classes
