### 2025-07-29 6f7a7c1
Reduce code duplication.
