+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -446,13 +446,10 @@ public:
                         0.0,
                         ddc::reducer::sum<double>(),
                         KOKKOS_LAMBDA(Idx<QDimRMesh, QDimThetaMesh> const& idx_quad) {
-                            Idx<QDimRMesh> const idx_r(idx_quad);
-                            Idx<QDimThetaMesh> const idx_theta(idx_quad);
                             return weak_integral_element(
                                     idx_test,
                                     idx_trial,
-                                    idx_r,
-                                    idx_theta,
+                                    idx_quad,
                                     coeff_alpha,
                                     coeff_beta,
                                     spline_evaluator,
@@ -554,13 +551,10 @@ public:
                                 0.0,
                                 ddc::reducer::sum<double>(),
                                 KOKKOS_LAMBDA(IdxQuadratureRTheta const idx_quad) {
-                                    IdxQuadratureR const idx_r(idx_quad);
-                                    IdxQuadratureTheta const idx_theta(idx_quad);
                                     return weak_integral_element<Mapping>(
                                             idx_test,
                                             idx_trial_polar,
-                                            idx_r,
-                                            idx_theta,
+                                            idx_quad,
                                             coeff_alpha,
                                             coeff_beta,
                                             spline_evaluator,
@@ -822,7 +816,7 @@ public:
                             IdxQuadratureR const idx_r(idx_quad);
                             IdxQuadratureTheta const idx_theta(idx_quad);
                             CoordRTheta coord(ddc::coordinate(idx_quad));
-                            return rhs(coord) * get_vals(idx, idx_quad)
+                            return rhs(coord) * get_polar_bspline_vals(idx, idx_quad)
                                    * int_volume_host(idx_r, idx_theta);
                         });
             });
@@ -928,18 +922,12 @@ public:
      * @brief Computes a quadrature summand corresponding to the 
      *        inner product.
      *
-     * @param[in] idx_r
-     *      The index for radial direction.
-     * @param[in] idx_theta
-     *      The index for poloidal direction
-     * @param[in] test_bspline_val_and_deriv
-     *      The data structure containing the derivatives over radial and poloidal directions for test space.
-     * @param[in] trial_bspline_val_and_deriv
-     *      The data structure containing the derivatives over radial and poloidal directions for trial space.
-     * @param[in] test_bspline_val_and_deriv_theta
-     *       The data structure containing the value and derivative along poloidal direction for test space.
-     * @param[in] trial_bspline_val_and_deriv_theta
-     *       The data structure containing the value and derivative along poloidal direction for trial space.
+     * @param[in] idx_test
+     *      The index of the test basis spline.
+     * @param[in] idx_trial
+     *      The index of the trial basis spline.
+     * @param[in] idx_quad
+     *      The index for the point in the quadrature scheme.
      * @param[in] coeff_alpha
      *      The spline representation of the @f$ \alpha @f$ function in the
      *      definition of the Poisson-like equation.
@@ -962,8 +950,7 @@ public:
     KOKKOS_FUNCTION double weak_integral_element(
             IdxBSPolar idx_test,
[…]
