### 2025-07-29 bcfc19a
Clang format
