### 2025-07-29 4fcb168
Start removing cells
