+++ src/pde_solvers/polarpoissonlikesolver.hpp
@@ -99,6 +99,9 @@ private:
     /// The 1D B-splines in the poloidal direction
     using BSplinesTheta = typename PolarBSplinesRTheta::BSplinesTheta_tag;
 
+    using KnotsR = ddc::knot_discrete_dimension_t<BSplinesR>;
+    using KnotsTheta = ddc::knot_discrete_dimension_t<BSplinesTheta>;
+
     using IdxRangeRTheta = IdxRange<GridR, GridTheta>;
     using IdxRTheta = Idx<GridR, GridTheta>;
 
@@ -159,9 +162,6 @@ private:
      * @brief Tag a vector on the second dimension of the quadrature mesh.
      */
     using IdxStepQuadratureTheta = IdxStep<QDimThetaMesh>;
-
-    using KnotsR = ddc::NonUniformBsplinesKnots<BSplinesR>;
-    using KnotsTheta = ddc::NonUniformBsplinesKnots<BSplinesTheta>;
 
     using ConstSpline2D = DConstField<IdxRangeBatchedBSRTheta>;
     using PolarSplineMemRTheta = DFieldMem<IdxRange<PolarBSplinesRTheta>>;
@@ -539,54 +539,62 @@ public:
         ddc::for_each(idxrange_singular, [&](IdxBSPolar const idx_test) {
             ddc::for_each(idxrange_non_singular_near_centre, [&](IdxBSRTheta const idx_trial) {
                 const IdxBSPolar idx_trial_polar(to_polar(idx_trial));
-                const Idx<BSplinesR> idx_trial_r(idx_trial);
-                const Idx<BSplinesTheta> idx_trial_theta(idx_trial);
+                const IdxBSR idx_trial_r(idx_trial);
+                const IdxBSTheta idx_trial_theta(idx_trial);
+
+                auto& bspl_r = ddc::discrete_space<BSplinesR>();
+                auto& bspl_theta = ddc::discrete_space<BSplinesTheta>();
 
                 // Find the index range covering the cells where both the test and trial functions are non-zero
-                const Idx<RCellDim> first_overlap_element_r(
-                        idx_trial_r.uid() < BSplinesR::degree()
-                                ? 0
-                                : idx_trial_r.uid() - BSplinesR::degree());
-                const Idx<ThetaCellDim> first_overlap_element_theta(
-                        theta_mod(idx_trial_theta.uid() - BSplinesTheta::degree()));
+                const Idx<KnotsR> start_non_zero_r(
+                        ::max(bspl_r.break_point_domain().front(),
+                              bspl_r.get_first_support_knot(idx_trial_r)));
+                const Idx<KnotsR> end_non_zero_r(
+                        ::min(bspl_r.get_last_support_knot(IdxBSR(PolarBSplinesRTheta::continuity)),
+                              bspl_r.get_last_support_knot(idx_trial_r)));
 
-                const IdxStep<RCellDim> n_overlap_r(
-                        m_n_overlap_cells - first_overlap_element_r.uid());
-                const IdxStep<ThetaCellDim> n_overlap_theta(BSplinesTheta::degree() + 1);
+                const Idx<KnotsTheta> start_non_zero_theta(
+                        bspl_theta.get_first_support_knot(idx_trial_theta));
+                const Idx<KnotsTheta> end_non_zero_theta(
+                        bspl_theta.get_last_support_knot(idx_trial_theta));
 
-                const IdxRange<RCellDim> r_cells(first_overlap_element_r, n_overlap_r);
-                const IdxRange<ThetaCellDim>
-                        theta_cells(first_overlap_element_theta, n_overlap_theta);
-                const IdxRange<RCellDim, ThetaCellDim> non_zero_cells(r_cells, theta_cells);
[…]
