### 2025-07-30 2914642
Fix GPU compilation
