### 2025-07-30 d3a0839
Reduce duplication
