### 2025-07-30 6ee1aca
Avoid recalculation of coord
