### 2025-07-30 8a42ef5
Update jmin_r to avoid needing corrections later
