### 2025-07-30 9becada
Remove unnecessary if
