### 2025-07-30 6e93723
Clang format
