### 2025-07-30 b28223d
Unnecessary variables
