### 2025-07-30 b8b9ec0
Add patched recipes
