### 2025-07-30 a388b9d
Simplify polar bspline usage (#389)
Evaluating polar bsplines involves 2 arrays to handle the singular
bsplines and the tensor product bsplines. The method `eval_basis`
returns an `Idx<BSplinesR, BSplinesTheta>` which identifies the smallest
non-zero bsplines. However on devel this element can be a tensor bspline
which doesn't exist in the polar bspline basis. As a result when calling
`eval_basis` a small adjustment was required to ensure that fileds were
not evaluated outside their domain. This PR fixes this logic issue. This
simplifies the work on the polar Poisson solver.
