### 2025-08-06 gyselax/gyselalibxx#403 @pratikvn
[JOSS paper]: Main issue for JOSS review,
### Issue Type

JOSS

### Location

https://github.com/gyselax/gyselalibxx/blob/joss/docs/joss_paper/paper.md

### Description

Hello everyone!

I am a reviewer for the JOSS paper that you have submitted: https://github.com/openjournals/joss-reviews/issues/8582.

This main issue is meant to track any subsequent issues in the code, installation or tests, or in the JOSS paper. To enable the editor to follow along the progress, I will link any concerns back to this issue.

I will also try to restrict any JOSS related issues only to this issue to reduce clutter.
