### 2025-08-11 gyselax/gyselalibxx#403 @pratikvn
Mandatory checklist:

- [ ] Documentation: Accessible: https://gyselax.github.io/gyselalibxx/
- [ ] Examples/Tutorials: Accessible: https://gyselax.github.io/gyselalibxx/docs/first_steps/landau_damping_tutorial.html
- [ ] Building and tests:
     - [ ] Docker: Builds and all tests pass.
     - [ ] On a HPC cluster: Builds and all tests pass.
- [ ] CI: Exists and run regularly, with each PR merge.

Recommended checklist:
- [ ] Performance showcase: Missing
- [ ] Functionality showcase: Partial (https://gyselax.github.io/gyselalibxx/simulations/index.html), but not with a list of available features and capabilities.
