### 2025-08-11 #8582 @pratikvn
@EmilyBourne and @piyueh, I have reviewed the software and the paper, and it looks mostly good from my side. Only one thing that is missing in the paper is the section on State of the Field, which compares Gyselalib++ to other existing software with similar functionality. While there is a brief mention of GYSELA and SeLalib, it is mixed in the same section, and should probably be separated out into another section.
