### 2025-08-11 gyselax/gyselalibxx#403 @pratikvn
@EmilyBourne,

> I agree that we are missing documentation on how to measure performance. As we are using Kokkos it can be done easily with [Kokkos-tools](https://github.com/kokkos/kokkos-tools). In particular with the [SimpleKernelTimer](https://github.com/kokkos/kokkos-tools/wiki/SimpleKernelTimer) or on Nvidia GPUs with [NVTXConnector](https://github.com/kokkos/kokkos-tools/wiki/NVTXConnector) which allows performance to be examined with NVIDIA Nsight Systems.
> I will open a PR to explain this in more detail.

Thanks!

> Otherwise the more reader-friendly version is in the READMEs of each folder which are included in the docs here:
https://gyselax.github.io/gyselalibxx/src/index.html

Yes, I think this is already good. Maybe you can highlight this, with maybe something like "Feature list/Building blocks of Gyselalib++" or something like that?
