### 2025-08-11 #8582 @EmilyBourne
https://github.com/gyselax/gyselalibxx/pull/407
