### 2025-08-20 gyselax/gyselalibxx#403 @EmilyBourne
https://github.com/gyselax/gyselalibxx/pull/407 has now been merged
