### 2025-08-28 4138278
Add curl operator (#414)
Add a curl operator working on arbitrary geometry.
