### 2025-08-28 #8582 @paulwilhelmvlasov
@EmilyBourne  and @piyueh, sorry for making you wait so long for my review/reply. I now was able to read through it and also install and look at the code. Overall it looks good to me. I was very pleasantly surprised by how smooth the installation went. The code is well documented and it's easy to find oneself around it.
Likewise, the paper was also written clearly, gave a good overview over the code and its objectives as well as how it fits and what it adds into the community.

I have just 3 (minor) comments or suggestions:
In the paper:
1) In the statement of need, the very first sentence: "[...] the necessary changes [...]". Maybe reformulate/clarify what you mean by that. From how it's written, it wasn't directly clear (to me) what this refers to, i.e., whether you mean necessary changes due to advancement of hardware/software requirements like exascale systems or availability of GPUs or whether this refers to the choice of grid or whether you mean all of that.
2)  It would be nice to write out the long name of DDC once for clarity. Right now you only use the abbreviation and refer to a library.
For the code:
3) It's a very nice idea to provide installation scripts for a number of clusters, especially as there are seemingly a lot of dependencies. As far as I have seen, there are scripts for local computers, Leonardo and a number of French clusters (?). Especially as this project is also co-funded by EuroHPC and a lot of potential users may want to use it on other EuroHPC clusters, it would be nice to have such scripts for all 8 (or however many of those clusters there are). Alternatively, it would be good to at least have a recipe or list of dependencies such that users on other clusters, than the ones listed, can more easily install it there. I know that this is likely a lot of work and I wouldn't require it for this publication, but in general it would be a nice "quality of life" feature to include in the future.

@piyueh: The paper looks good to me as well and hence I would also recommend it for publication.
