### 2025-08-28 #8582 @EmilyBourne
@pratikvn @paulwilhelmvlasov Thank you very much for taking the time to review the paper and for your comments.

@paulwilhelmvlasov I have updated the paper to address your 1st and 2nd comment.
For the third I agree that it would be nice to have such scripts for all clusters however I don't think this is feasible currently. We don't have access to all of the clusters and we don't want to include scripts if we don't have collaborators using them regularly as we cannot assure their quality in this case.
I have previously attempted to write general documentation describing how to create such scripts for new clusters but every cluster seems to have subtle differences that have so far made this an exercice in futility unfortunately. We would recommend that people open an issue in this case so we can help them out.
We are working on a spack recipe to make this simpler in the future
