### 2025-09-02 #8582 @EmilyBourne
> Thank you [@pratikvn](https://github.com/pratikvn) and [@paulwilhelmvlasov](https://github.com/paulwilhelmvlasov) for reviewing this paper.
>
> [@EmilyBourne](https://github.com/EmilyBourne) I am starting the post-review stage. In the meantime, if there were any changes made to the code at a reviewer’s request, please create a new release. If there were no code change, then continue with the version noted in this submission (v0.2.0 in your case). Archive this specific version on an online service such as Zenodo or Figshare. Then, post the final release version number and the DOI from Zenodo/Figshare/etc. here. This JOSS paper will only cover this specific version of the code. See [here](https://github.com/openjournals/joss-reviews/issues/8582#issuecomment-3245869197) for more information.

The only change was a documentation change
