### 2025-09-02 #8582 @piyueh
@pratikvn @paulwilhelmvlasov Could you let us know which snapshot of code did you review? (e.g., the `devel` branch, the `v0.2.0` release, or a specific commit) Thank you!

@EmilyBourne I'll wait for reviewers' replies and see what's the best option. The bottom line is option 1: `I take the only commit explicitly asked for by reviewers and release a v0.2.1 for JOSS`. But if reviewers actually did the review on any commits that are newer than v0.2.0, then option 2 or 3 seems also okay to me.
