### 2025-09-03 #8582 @EmilyBourne
Thank you for your quick replies. I have created v0.3.0 using all commits up to 41382788a (all commits which have been reviewed by someone): https://github.com/gyselax/gyselalibxx/releases/tag/v0.3.0
The DOI is: https://doi.org/10.5281/zenodo.17044083

I have completed the additional author tasks after review is complete.
@piyueh over to you
