### 2025-09-03 #8582 @piyueh
Hi @EmilyBourne, here is a patch for the paper after proofread by @@danielskatz (thank you!): https://github.com/gyselax/gyselalibxx/compare/joss...danielskatz:gyselalibxx:patch-1 If you agree with these revisions, you can create a PR to merge the changes.

In addition, here are my notes after proofread (line numbers are based on [this PDF draft](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08582/joss.08582/10.21105.joss.08582.pdf)):
* line 29: acronym ITER doesn't seem to be introduced
* Special terms in References:
  * line 103: semi-lagrangian -> semi-Lagragian
  * line 131: vlasov -> Vlasov
  * line 146: reynolds -> Reynolds
* And the reference at line 144-145 is a JOSS submission that has not yet been accepted/published. So I think it should be changed to the URL or repository of that software. Alternatively, you can create an archive of a specific version of that software on Zenodo and cite the Zonodo record. (And then after that JOSS submission is accepted, you can update the version of that software on Zenodo.)
