### 2025-09-03 #8582 @danielskatz
I've made my part into a PR - https://github.com/gyselax/gyselalibxx/pull/422
