### 2025-09-04 #8582 @piyueh
@EmilyBourne Thanks for the clarification. Would it be possible to add a reference to ITER in the paper? Like citing their website if no official citation format is provided? I think this would be helpful for readers who are new to the field.

Feel free on the editing suggestions. It’s likely my mistake. I compiled some initial edits and then asked other editors to proofread again. Since I wasn’t familiar with British English standards, I suggested many of the British → American English changes, and those suggestions may have influenced other editors.
