### 2025-09-09 #8582 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
